### 2024-11-07 554a25d
remove deprecated neural_net access from utils
