### 2024-11-08 d0a24e9
remove tests for deprecated neural_nets_factory
