### 2024-11-12 653d0b4
Fix type checks for EnsemblePosterior weights (#1299)
* minor fix for EnsemblePosterior weights.setter

* Update sbi/inference/posteriors/ensemble_posterior.py

Co-authored-by: Jan <janfb@users.noreply.github.com>

---------

Co-authored-by: Jan <janfb@users.noreply.github.com>
