### 2024-11-13 3ff1f51
fix #1301: remove deprecated neural_net access from `utils` to avoid circular imports (#1302)
* remove deprecated neural_net access from utils

* remove tests for deprecated neural_nets_factory
