### 2024-11-14 #7428 @janfb
Hi @crvernon , thanks a lot for starting this process. Yes of course. There are lots of new features and general changes to the package. We implemented several new SBI algorithms, new validation methods and more plotting tools, and provided access to the widely-used sampling libraries for posterior sampling and visualization (pyro, pymc, arviz). We also added many more tutorials, revised the entire documentation website, and improved the package maintenance pipelines.

In the paper, we added a figure to give an overview of the new features:

In lines 71-78 in the draft PDF, we also give a summary of what changed compared to the last published version.

Please let me know, if you need further clarification.
Thanks,
Jan
