### 2024-12-02 3bd8aa9
test: ensemble posterior weights validation (#1307)
Signed-off-by: samadpls <abdulsamadsid1@gmail.com>
