### 2024-12-02 59eb1e7
rename ci to tests for readme badge
