+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: Continuous Integration
+name: Tests
 
 on: [pull_request, workflow_dispatch]
 
