### 2024-12-03 4b3d409
Add discord link to README.md (#1317)
* Update README.md

* rename ci to tests for readme badge
