### 2024-12-03 9db9e9e
expose batched sampling option; error handling
