+++ sbi/utils/diagnostics_utils.py
@@ -1,4 +1,5 @@
 import warnings
+from typing import Optional
 
 import torch
 from joblib import Parallel, delayed
@@ -18,6 +19,7 @@ def get_posterior_samples_on_batch(
     num_workers: int = 1,
     show_progress_bar: bool = False,
     use_batched_sampling: bool = True,
+    batch_size: Optional[int] = None,
 ) -> Tensor:
     """Get posterior samples for a batch of xs.
 
@@ -28,22 +30,37 @@ def get_posterior_samples_on_batch(
         num_workers: number of workers to use for parallelization.
         show_progress_bars: whether to show progress bars.
         use_batched_sampling: whether to use batched sampling if possible.
-
+        batch_size: batch size for batched sampling. Useful for batched sampling with
+            large batches of xs for avoiding memory overflow.
     Returns:
         posterior_samples: of shape (num_samples, batch_size, dim_parameters).
     """
-    batch_size = len(xs)
+    num_xs = len(xs)
+    if batch_size is None:
+        batch_size = num_xs
 
-    # Try using batched sampling when implemented.
-    try:
-        # has shape (num_samples, batch_size, dim_parameters)
-        if use_batched_sampling:
-            posterior_samples = posterior.sample_batched(
-                sample_shape, x=xs, show_progress_bars=show_progress_bar
+    if use_batched_sampling:
+        try:
+            # distribute the batch of xs into smaller batches
+            batched_xs = xs.split(batch_size)
+            posterior_samples = torch.cat(
+                [  # has shape (num_samples, num_xs, dim_parameters)
+                    posterior.sample_batched(
+                        sample_shape, x=xs_batch, show_progress_bars=show_progress_bar
+                    )
+                    for xs_batch in batched_xs
+                ],
+                dim=1,
             )
-        else:
-            raise NotImplementedError
-    except (NotImplementedError, AssertionError):
+        except (NotImplementedError, AssertionError):
+            warnings.warn(
+                "Batched sampling not implemented for this posterior. "
+                "Falling back to non-batched sampling.",
+                stacklevel=2,
+            )
+            use_batched_sampling = False
+
+    if not use_batched_sampling:
         # We need a function with extra training step for new x for VIPosterior.
         def sample_fun(
             posterior: NeuralPosterior, sample_shape: Shape, x: Tensor, seed: int = 0
@@ -57,13 +74,13 @@ def get_posterior_samples_on_batch(
         if isinstance(posterior, (VIPosterior, MCMCPosterior)):
             warnings.warn(
                 "Using non-batched sampling. Depending on the number of different xs "
-                f"( {batch_size}) and the number of parallel workers {num_workers}, "
-                "this might be slow.",
+                f"( {num_xs}) and the number of parallel workers {num_workers}, "
+                "this might take a lot of time.",
                 stacklevel=2,
             )
 
         # Run in parallel with progress bar.
-        seeds = torch.randint(0, 2**32, (batch_size,))
+        seeds = torch.randint(0, 2**32, (num_xs,))
         outputs = list(
             tqdm(
                 Parallel(return_as="generator", n_jobs=num_workers)(
@@ -72,7 +89,7 @@ def get_posterior_samples_on_batch(
                 ),
                 disable=not show_progress_bar,
                 total=len(xs),
-                desc=f"Sampling {batch_size} times {sample_shape} posterior samples.",
+                desc=f"Sampling {num_xs} times {sample_shape} posterior samples.",
             )
         )  # (batch_size, num_samples, dim_parameters)
         # Transpose to shape convention: (sample_shape, batch_size, dim_parameters)
[…]
