### 2024-12-10 4c69fa4
undo batch_size option
