### 2024-12-12 b2fe636
add method for iid-batched conditioning.
- deprecate MNLE-based potential (can be nle-based)
- adapt tests for conditioned mnle.
