+++ tests/mnle_test.py
@@ -1,29 +1,31 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
+from typing import Union
+
 import pytest
 import torch
 from pyro.distributions import InverseGamma
-from torch.distributions import Beta, Binomial, Categorical, Gamma
+from torch import Tensor
+from torch.distributions import Beta, Binomial, Distribution, Gamma
 
 from sbi.inference import MNLE, MCMCPosterior
 from sbi.inference.posteriors.rejection_posterior import RejectionPosterior
 from sbi.inference.posteriors.vi_posterior import VIPosterior
 from sbi.inference.potentials.base_potential import BasePotential
 from sbi.inference.potentials.likelihood_based_potential import (
-    MixedLikelihoodBasedPotential,
+    likelihood_estimator_based_potential,
 )
 from sbi.neural_nets import likelihood_nn
 from sbi.neural_nets.embedding_nets import FCEmbedding
 from sbi.utils import BoxUniform, mcmc_transform
-from sbi.utils.conditional_density_utils import ConditionedPotential
 from sbi.utils.torchutils import atleast_2d, process_device
 from sbi.utils.user_input_checks_utils import MultipleIndependent
 from tests.test_utils import check_c2st
 
 
 # toy simulator for mixed data
-def mixed_simulator(theta, stimulus_condition=2.0):
+def mixed_simulator(theta: Tensor, stimulus_condition: Union[Tensor, float] = 2.0):
     """Simulator for mixed data."""
     # Extract parameters
     beta, ps = theta[:, :1], theta[:, 1:]
@@ -190,11 +192,28 @@ def test_mnle_accuracy_with_different_samplers_and_trials(
 
 
 class BinomialGammaPotential(BasePotential):
-    def __init__(self, prior, x_o, concentration_scaling=1.0, device="cpu"):
+    """Binomial-Gamma potential for mixed data."""
+
+    def __init__(
+        self,
+        prior: Distribution,
+        x_o: Tensor,
+        concentration_scaling: Union[Tensor, float] = 1.0,
+        device="cpu",
+    ):
         super().__init__(prior, x_o, device)
+
+        # concentration_scaling needs to be a float or match the batch size
+        if isinstance(concentration_scaling, Tensor):
+            num_trials = x_o.shape[0]
+            assert concentration_scaling.shape[0] == num_trials
+
+            # Reshape to match convention (batch_size, num_trials, *event_shape)
+            concentration_scaling = concentration_scaling.reshape(1, num_trials, -1)
+
         self.concentration_scaling = concentration_scaling
 
-    def __call__(self, theta, track_gradients: bool = True):
+    def __call__(self, theta: Tensor, track_gradients: bool = True) -> Tensor:
         theta = atleast_2d(theta)
 
         with torch.set_grad_enabled(track_gradients):
@@ -202,11 +221,12 @@ class BinomialGammaPotential(BasePotential):
 
         return iid_ll + self.prior.log_prob(theta)
 
-    def iid_likelihood(self, theta):
+    def iid_likelihood(self, theta: Tensor) -> Tensor:
         batch_size = theta.shape[0]
         num_trials = self.x_o.shape[0]
         theta = theta.reshape(batch_size, 1, -1)
         beta, rho = theta[:, :, :1], theta[:, :, 1:]
+
         # vectorized
         logprob_choices = Binomial(probs=rho).log_prob(
             self.x_o[:, 1:].reshape(1, num_trials, -1)
[…]
