### 2024-12-12 074efa0
update notebook, bugfixes
