### 2024-12-16 65bd398
Start of mini_sbibm
