### 2024-12-17 8a95c8f
add batch dim for x, add test.
