+++ sbi/inference/potentials/likelihood_based_potential.py
@@ -116,19 +116,27 @@ class LikelihoodBasedPotential(BasePotential):
                 )
             return log_likelihood_batches + self.prior.log_prob(theta)  # type: ignore
 
-    def condition_on(self, condition: Tensor, dims_to_sample: List[int]) -> Callable:
-        """Returns a potential conditioned on a subset of theta dimensions.
+    def condition_on_theta(
+        self, theta_condition: Tensor, dims_to_sample: List[int]
+    ) -> Callable:
+        """Returns a potential function conditioned on a subset of theta dimensions.
 
         The condition is a part of theta, but is assumed to correspond to a batch of iid
-        x_o.
+        x_o. For example, it can be a batch of experimental conditions that corresponds
+        to a batch of i.i.d. trials in x_o.
 
         Args:
-            condition: The condition to fix.
-            dims_to_sample: The indices of the parameters to sample.
+            theta_condition: The condition values to be conditioned.
+            dims_to_sample: The indices of the columns in theta that will be sampled,
+                i.e., that *not* conditioned. For example, if original theta has shape
+                `(batch_dim, 3)`, and `dims_to_sample=[0, 1]`, then the potential will
+                set `theta[:, 3] = theta_condition` at inference time.
 
         Returns:
-            A potential function conditioned on the condition.
+            A potential function conditioned on the theta_condition.
         """
+
+        assert self.x_is_iid, "Conditioning is only supported for iid data."
 
         def conditioned_potential(
             theta: Tensor, x_o: Optional[Tensor] = None, track_gradients: bool = True
@@ -138,10 +146,10 @@ class LikelihoodBasedPotential(BasePotential):
             ), "dims_to_sample must match the number of parameters to sample."
             theta_without_condition = theta[:, dims_to_sample]
 
-            return _log_likelihood_with_iid_condition(
+            return _log_likelihood_over_iid_conditions(
                 x=x_o if x_o is not None else self.x_o,
                 theta_without_condition=theta_without_condition,
-                condition=condition,
+                condition=theta_condition,
                 estimator=self.likelihood_estimator,
                 track_gradients=track_gradients,
             )
@@ -205,63 +213,75 @@ def _log_likelihoods_over_trials(
     return log_likelihood_trial_sum
 
 
-def _log_likelihood_with_iid_condition(
+def _log_likelihood_over_iid_conditions(
     x: Tensor,
     theta_without_condition: Tensor,
     condition: Tensor,
     estimator: ConditionalDensityEstimator,
     track_gradients: bool = False,
 ) -> Tensor:
-    """Return log likelihoods summed over iid trials of `x` with a matching batch of
-    conditions.
+    """Returns $\\log(p(x_o|\theta, condition)$, where x_o is a batch of iid data, and
+    condition is a matching batch of conditions.
 
     This function is different from `_log_likelihoods_over_trials` in that it moves the
-    iid batch dimension of `x` onto the batch dimension of `theta`. This is useful when
+    iid batch dimension of `x` onto the batch dimension of `theta`. This is needed when
     the likelihood estimator is conditioned on a batch of conditions that are iid with
-    the batch of `x`. It avoid the evaluation of the likelihood for every combination of
-    `x` and `condition`. Instead, it manually constructs a batch covering all
-    combination of iid trial and theta batch and reshapes to sum over the iid
+    the batch of `x`. It avoids the evaluation of the likelihood for every combination
+    of `x` and `condition`. Instead, it manually constructs a batch covering all
[…]
