### 2024-12-17 ced7e15
fix shape handling, adapt tutorial.
