### 2024-12-18 621a875
Working prototype
