+++ tests/conftest.py
@@ -1,37 +1,15 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+import shutil
 
 import pytest
 import torch
 
 from sbi.utils.sbiutils import seed_all_backends
-from pytest_harvest import get_session_results_dct
 
 # Seed for `set_seed` fixture. Change to random state of all seeded tests.
 seed = 1
-
-
-# Pytest harvest
-@pytest.fixture(scope="session", autouse=True)
-def collect_results(request):
-    """
-    Automatically runs after the test session.
-    We'll collect the test results and store them for later use (e.g. in a dataframe).
-    """
-    print(request)
-    yield  # Run all tests first
-    # At this point, tests are all done. Let's harvest the results:
-    session_results = get_session_results_dct(request.session)
-
-    # Now `session_results` is a dict with entries for each test function.
-    # You can transform it into a DataFrame if you want.
-    try:
-        import pandas as pd
-
-        df = pd.DataFrame.from_dict(session_results, orient='index')
-        print("\nHarvested Test Results:\n", df)
-    except ImportError:
-        print("\nHarvested Test Results (raw dictionary):\n", session_results)
+harvested_fixture_data = None
 
 
 # Use seed automatically for every test function.
@@ -56,6 +34,106 @@ def pytest_collection_modifyitems(config, items):
         for item in items:
             if "gpu" in item.keywords:
                 item.add_marker(skip_gpu)
+
+    # Filter tests to only those with the 'bm' marker
+    if config.getoption("--bm"):
+        # Filter tests to only those with the 'bm' marker
+        filtered_items = []
+        for item in items:
+            # Using newer pytest versions (>=4.6):
+            if item.get_closest_marker("benchmark"):
+                filtered_items.append(item)
+
+        items[:] = filtered_items  # Inplace!
+
+
+# Run mini-benchmark tests with `pytest --print-harvest`
+def pytest_addoption(parser):
+    parser.addoption(
+        "--bm",
+        action="store_true",
+        default=False,
+        help="Print the harvest results at the end of the test session",
+    )
+
+
+@pytest.fixture(scope="session", autouse=True)
+def finalize_fixture_store(request, fixture_store):
+    # The code before `yield` runs at the start of the session (before tests).
+    yield
+    # The code after `yield` runs after all tests have completed.
+    # At this point, fixture_store should have all the harvested data.
+    global harvested_fixture_data
+    harvested_fixture_data = dict(fixture_store)
+
+
+def pytest_terminal_summary(terminalreporter, exitstatus, config):
+    if config.getoption("--bm"):
+        terminal_width = shutil.get_terminal_size().columns
+        summary_text = " mini SBIBM results "
+        centered_line = summary_text.center(terminal_width, '=')
+        colored_line = f"\033[96m{centered_line}\033[0m"
+        terminalreporter.write_line(colored_line)
+
+        if harvested_fixture_data is not None:
+            terminalreporter.write_line("Harvested Fixture Data:")
+
+            results = harvested_fixture_data["results_bag"]
+
+            # Extract relevant data (method, task, metric)
+            methods = set()
+            tasks = set()
+            data = {}  # (method, task) -> metric
+
+            for _, info in results.items():
+                method = info.get('method')
+                task = info.get('task_name')
+                metric = info.get('metric')
+                # You can also choose another metric or value to display
+
+                if method is not None and task is not None:
+                    methods.add(method)
+                    tasks.add(task)
+                    data[(method, task)] = metric
+
+            # Sort methods and tasks for consistent display
+            methods = sorted(methods)
+            tasks = sorted(tasks)
+
+            if not methods or not tasks:
[…]
