+++ tests/mini_sbibm/__init__.py
@@ -1,4 +1,3 @@
-
 from .two_moons import TwoMoons
 
 
