+++ tests/mini_sbibm/two_moons.py
@@ -1,4 +1,3 @@
-
 import math
 import os
 from typing import Callable
