+++ tests/bm_test.py
@@ -1,31 +1,27 @@
 import pytest
 import torch
 
-from sbi.inference import NPE, NRE
+from sbi.inference import FMPE, NLE, NPE, NPSE, NRE
 from sbi.utils.metrics import c2st
 
 from .mini_sbibm import get_task
 
+# The probably should be some user control on this
+SEED = 0
+TASKS = ["two_moons", "linear_mvg_2d", "gaussian_linear", "slcp"]
+NUM_SIMULATIONS = 2000
+EVALUATION_POINTS = 4  # Currently only 3 observation tested for speed
 
-@pytest.mark.benchmark
-@pytest.mark.parametrize('task_name', ['two_moons'], ids=str)
-@pytest.mark.parametrize('density_estimator', ["maf", "nsf"], ids=str)
-def test_benchmark_npe_methods(
-    task_name, density_estimator, results_bag, method=None, num_simulations=1000, seed=0
-):
-    torch.manual_seed(seed)
-    task = get_task(task_name)
-    thetas, xs = task.get_data(num_simulations)
-    assert thetas.shape[0] == num_simulations
-    assert xs.shape[0] == num_simulations
+TRAIN_KWARGS = {
+    # "training_batch_size": 200,  # To speed up training
+}
 
-    inference = NPE(density_estimator=density_estimator)
-    _ = inference.append_simulations(thetas, xs).train()
+# Amortized benchmarking
 
-    posterior = inference.build_posterior()
 
+def standard_eval_c2st_loop(posterior, task):
     metrics = []
-    for i in range(1, 2):  # Currently only one observation tested for speed
+    for i in range(1, EVALUATION_POINTS):
         x_o = task.get_observation(i)
         posterior_samples = task.get_reference_posterior_samples(i)
         approx_posterior_samples = posterior.sample((1000,), x=x_o)
@@ -37,43 +33,117 @@ def test_benchmark_npe_methods(
     mean_c2st = sum(metrics) / len(metrics)
     # Convert to float rounded to 3 decimal places
     mean_c2st = float(f"{mean_c2st:.3f}")
+    return mean_c2st
 
+
+DENSITY_estimators = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
+DENSITY_estimators = ["maf", "nsf"]  # Fast
+
+
+@pytest.mark.benchmark
+@pytest.mark.parametrize('task_name', TASKS, ids=str)
+@pytest.mark.parametrize('density_estimator', DENSITY_estimators, ids=str)
+def test_benchmark_npe_methods(task_name, density_estimator, results_bag):
+    torch.manual_seed(SEED)
+    task = get_task(task_name)
+    thetas, xs = task.get_data(NUM_SIMULATIONS)
+    prior = task.get_prior()
+
+    print(thetas.shape, xs.shape)
+
+    inference = NPE(prior, density_estimator=density_estimator)
+    _ = inference.append_simulations(thetas, xs).train(**TRAIN_KWARGS)
+
+    posterior = inference.build_posterior()
+
+    mean_c2st = standard_eval_c2st_loop(posterior, task)
+
+    # Cache results
     results_bag.metric = mean_c2st
-    results_bag.num_simulations = num_simulations
+    results_bag.num_simulations = NUM_SIMULATIONS
     results_bag.task_name = task_name
     results_bag.method = "NPE_" + density_estimator
 
 
 @pytest.mark.benchmark
-@pytest.mark.parametrize('task_name', ['two_moons'], ids=str)
-def test_benchmark_nre_methods(task_name, results_bag, num_simulations=1000, seed=0):
-    torch.manual_seed(seed)
[…]
