### 2024-12-19 3d3bd00
add release template.
