### 2024-12-19 0b66fca
add changelog, fix dependencies, bump version.
