### 2024-12-19 82185e7
fix docstrings about condition sample_dim.
