+++ sbi/neural_nets/estimators/base.py
@@ -119,15 +119,15 @@ class ConditionalEstimator(nn.Module, ABC):
 class ConditionalDensityEstimator(ConditionalEstimator):
     r"""Base class for density estimators.
 
-    The density estimator class is a wrapper around neural networks that
-    allows to evaluate the `log_prob`, `sample`, and provide the `loss` of $\theta,x$
-    pairs. Here $\theta$ would be the `input` and $x$ would be the `condition`.
+    The density estimator class is a wrapper around neural networks that allows to
+    evaluate the `log_prob`, `sample`, and provide the `loss` of $\theta,x$ pairs. Here
+    $\theta$ would be the `input` and $x$ would be the `condition`.
 
     Note:
         We assume that the input to the density estimator is a tensor of shape
-        (batch_size, input_size), where input_size is the dimensionality of the input.
-        The condition is a tensor of shape (batch_size, *condition_shape), where
-        condition_shape is the shape of the condition tensor.
+        (sample_dim, batch_dim, *input_shape), where input_shape is the dimensionality
+        of the input. The condition is a tensor of shape (batch_size, *condition_shape),
+        where condition_shape is the shape of the condition tensor.
 
     """
 
@@ -226,15 +226,15 @@ class ConditionalDensityEstimator(ConditionalEstimator):
 class ConditionalVectorFieldEstimator(ConditionalEstimator):
     r"""Base class for vector field (e.g., score and ODE flow) estimators.
 
-    The density estimator class is a wrapper around neural networks that
-    allows to evaluate the `vector_field`, and provide the `loss` of $\theta,x$
-    pairs. Here $\theta$ would be the `input` and $x$ would be the `condition`.
+    The density estimator class is a wrapper around neural networks that allows to
+    evaluate the `vector_field`, and provide the `loss` of $\theta,x$ pairs. Here
+    $\theta$ would be the `input` and $x$ would be the `condition`.
 
     Note:
         We assume that the input to the density estimator is a tensor of shape
-        (batch_size, input_size), where input_size is the dimensionality of the input.
-        The condition is a tensor of shape (batch_size, *condition_shape), where
-        condition_shape is the shape of the condition tensor.
+        (sample_dim, batch_dim, *input_shape), where input_shape is the dimensionality
+        of the input. The condition is a tensor of shape (batch_dim, *condition_shape),
+        where condition_shape is the shape of the condition tensor.
 
     """
 
