### 2024-12-20 5f5ba6b
docs: Add conda-forge install instructions
* Add install instructions from conda-forge for pixi and conda to README.
   - c.f. https://github.com/conda-forge/sbi-feedstock
* Add conda-forge badge.
* Add install instructions from conda-forge for pixi and conda to docs.
   - Correct minimum required Python to Python 3.9.
   - Update example python version in conda env creation to Python 3.12.
