### 2024-12-21 737764d
feedback; fix texts
