+++ sbi/inference/potentials/likelihood_based_potential.py
@@ -117,23 +117,25 @@ class LikelihoodBasedPotential(BasePotential):
             return log_likelihood_batches + self.prior.log_prob(theta)  # type: ignore
 
     def condition_on_theta(
-        self, theta_condition: Tensor, dims_to_sample: List[int]
+        self, local_theta: Tensor, dims_global_theta: List[int]
     ) -> Callable:
-        """Returns a potential function conditioned on a subset of theta dimensions.
+        r"""Returns a potential function conditioned on a subset of theta dimensions.
 
-        The condition is a part of theta, but is assumed to correspond to a batch of iid
-        x_o. For example, it can be a batch of experimental conditions that corresponds
-        to a batch of i.i.d. trials in x_o.
+        The goal of this function is to divide the original `theta` into a
+        `global_theta` we do inference over, and a `local_theta` we condition on (in
+        addition to conditioning on `x_o`). Thus, the returned potential function will
+        calculate $\prod_{i=1}^{N}p(x_i | local_theta_i, \global_theta)$, where `x_i`
+        and `local_theta_i` are fixed and `global_theta` varies at inference time.
 
         Args:
-            theta_condition: The condition values to be conditioned.
-            dims_to_sample: The indices of the columns in theta that will be sampled,
-                i.e., that *not* conditioned. For example, if original theta has shape
-                `(batch_dim, 3)`, and `dims_to_sample=[0, 1]`, then the potential will
-                set `theta[:, 3] = theta_condition` at inference time.
+            local_theta: The condition values to be conditioned.
+            dims_global_theta: The indices of the columns in `theta` that will be
+                sampled, i.e., that *not* conditioned. For example, if original theta
+                has shape `(batch_dim, 3)`, and `dims_global_theta=[0, 1]`, then the
+                potential will set `theta[:, 3] = local_theta` at inference time.
 
         Returns:
-            A potential function conditioned on the theta_condition.
+            A potential function conditioned on the `local_theta`.
         """
 
         assert self.x_is_iid, "Conditioning is only supported for iid data."
@@ -142,9 +144,9 @@ class LikelihoodBasedPotential(BasePotential):
             theta: Tensor, x_o: Optional[Tensor] = None, track_gradients: bool = True
         ) -> Tensor:
             assert (
-                len(dims_to_sample) == theta.shape[1]
-            ), "dims_to_sample must match the number of parameters to sample."
-            theta_without_condition = theta[:, dims_to_sample]
+                len(dims_global_theta) == theta.shape[1]
+            ), "dims_global_theta must match the number of parameters to sample."
+            global_theta = theta[:, dims_global_theta]
             x_o = x_o if x_o is not None else self.x_o
             # x needs shape (sample_dim (iid), batch_dim (xs), *event_shape)
             if x_o.dim() < 3:
@@ -152,10 +154,10 @@ class LikelihoodBasedPotential(BasePotential):
                     x_o, event_shape=x_o.shape[1:], leading_is_sample=self.x_is_iid
                 )
 
-            return _log_likelihood_over_iid_conditions(
+            return _log_likelihood_over_iid_trials_and_local_theta(
                 x=x_o,
-                theta_without_condition=theta_without_condition,
-                condition=theta_condition,
+                global_theta=global_theta,
+                local_theta=local_theta,
                 estimator=self.likelihood_estimator,
                 track_gradients=track_gradients,
             )
@@ -219,51 +221,50 @@ def _log_likelihoods_over_trials(
     return log_likelihood_trial_sum
 
 
[…]
