### 2024-12-21 390a518
docs: Fix docstrings for condition sample_dim. (#1338)
