### 2024-12-21 e7940dc
feat: NLE with multiple iid conditions (#1331)
* add method for iid-batched conditioning.

- deprecate MNLE-based potential (can be nle-based)
- adapt tests for conditioned mnle.

* update notebook, bugfixes

* add batch dim for x, add test.

* fix shape handling, adapt tutorial.
