+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * expose batched sampling option; error handling by @janfb in https://github.com/sbi-dev/sbi/pull/1321
 * Fix #1316: remove sample_dim docstring for condition. by @janfb in https://github.com/sbi-dev/sbi/pull/1338
 * docs: fix tutorial typos by @janfb in https://github.com/sbi-dev/sbi/pull/1341
+* docs: run and seed SBC tutorial by @manuel-morales-a in https://github.com/sbi-dev/sbi/pull/1336
 
 ## New Contributors 🎉
 
@@ -24,6 +25,7 @@
 * @CompiledAtBirth made their first contribution in https://github.com/sbi-dev/sbi/pull/1299
 * @tvwenger made their first contribution in https://github.com/sbi-dev/sbi/pull/1302
 * @matthewfeickert made their first contribution in https://github.com/sbi-dev/sbi/pull/1340
+* @manuel-morales-a made their first contribution in https://github.com/sbi-dev/sbi/pull/1336
 
 **Full Changelog**: https://github.com/sbi-dev/sbi/compare/v0.23.2...v0.23.3
 
