### 2024-12-30 9152e93
docs: prepare release, bump version (#1337)
* add release template.

* add changelog, fix dependencies, bump version.

* update changelog.

* add latest contributions.
