+++ sbi/__version__.py
@@ -1,6 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
-VERSION = (0, 23, 2)
+VERSION = (0, 23, 3)
 
 __version__ = ".".join(map(str, VERSION))
