### 2025-01-07 624682b
Adding some base, might changing class to a more functional form integrated into the potential
