+++ sbi/samplers/score/score_fn.py
@@ -1,0 +1,372 @@
+import re
+import torch
+from torch import Tensor
+from torch.distributions import Distribution
+from typing import Optional, Callable
+from abc import abstractmethod
+
+from sbi.neural_nets.estimators import ConditionalScoreEstimator
+from sbi.utils.torchutils import ensure_theta_batched
+from sbi.samplers.score.denoise_and_marginalize import denoise, marginalize
+
+
+def mv_diag_or_dense(A_diag_or_dense: Tensor, b: Tensor) -> Tensor:
+    """Dot product of a diagonal matrix and a dense matrix.
+
+    Args:
+        A_diag_or_dense (Tensor): Diagonal matrix.
+        b (Tensor): Dense matrix.
+
+    Returns:
+        Tensor: Dot product.
+    """
+    if A_diag_or_dense.ndim == 1:
+        return A_diag_or_dense * b
+    else:
+        return torch.matmul(A_diag_or_dense, b)
+
+
+def solve_diag_or_dense(A_diag_or_dense: Tensor, b: Tensor) -> Tensor:
+    """Solve a linear system with a diagonal matrix or a dense matrix.
+
+    Args:
+        A_diag_or_dense (Tensor): Diagonal matrix or dense matrix.
+        b (Tensor): Dense matrix.
+
+    Returns:
+        Tensor: Solution to the linear system.
+    """
+    if A_diag_or_dense.ndim == 1:
+        return b / A_diag_or_dense
+    else:
+        return torch.linalg.solve(A_diag_or_dense, b)
+
+
+def add_diag_or_dense(A_diag_or_dense: Tensor, B_diag_or_dense: Tensor) -> Tensor:
+    """Add two diagonal matrices or two dense matrices.
+
+    Args:
+        A_diag_or_dense (Tensor): Diagonal matrix or dense matrix.
+        B_diag_or_dense (Tensor): Diagonal matrix or dense matrix.
+
+    Returns:
+        Tensor: Sum of the matrices.
+    """
+    if (
+        A_diag_or_dense.ndim == 1
+        and B_diag_or_dense.ndim == 1
+        or A_diag_or_dense.ndim == 2
+        and B_diag_or_dense.ndim == 2
+    ):
+        return A_diag_or_dense + B_diag_or_dense
+    elif A_diag_or_dense.ndim == 2 and B_diag_or_dense.ndim == 1:
+        return A_diag_or_dense + torch.diag(B_diag_or_dense)
+    elif A_diag_or_dense.ndim == 1 and B_diag_or_dense.ndim == 2:
+        return torch.diag(A_diag_or_dense) + B_diag_or_dense
+    else:
+        raise ValueError("Incompatible dimensions for addition")
+
+
+class ScoreFnIID:
+    def __init__(
+        self,
+        score_estimator: ConditionalScoreEstimator,
+        prior: Distribution,
+        device: str = "cpu",
+    ):
+        r"""Initializes the ScoreFnIID class.
+
+        Args:
+            score_estimator: The neural network modelling the score.
+            prior: The prior distribution.
+            device: The device on which to evaluate the potential.
+        """
+        self.score_estimator = score_estimator.to(device)
+        self.prior = prior
+        self.score_estimator.eval()
+
+    @abstractmethod
+    def __call__(
+        self,
+        theta: Tensor,
+    ) -> Tensor:
+        r"""Abstract method to be implemented by subclasses.
+
+        Args:
+            theta: The parameters at which to evaluate the potential.
+
+        Returns:
+            The score function.
+        """
+        pass
+
+    def prior_score_fn(self, theta: Tensor) -> Tensor:
+        r"""Computes the score of the prior distribution.
+
+        Args:
[…]
