### 2025-01-08 9c34d34
wip: update toy simulator
