### 2025-01-09 a567fc8
wip: build_mnpe integrated, build_mnle refactored, mnle_test fails; TODO: MNPE class + test
