+++ sbi/neural_nets/net_builders/mnle.py
@@ -2,201 +2,22 @@
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
 import warnings
-from typing import Optional
 
-import torch
-from torch import Tensor, nn
-
-from sbi.neural_nets.estimators import MixedDensityEstimator
-from sbi.neural_nets.estimators.mixed_density_estimator import (
-    _is_discrete,
-    _separate_input,
-)
-from sbi.neural_nets.net_builders.categorial import (
-    build_categoricalmassestimator,
-)
-from sbi.neural_nets.net_builders.flow import (
-    build_made,
-    build_maf,
-    build_maf_rqs,
-    build_nsf,
-    build_zuko_bpf,
-    build_zuko_gf,
-    build_zuko_maf,
-    build_zuko_naf,
-    build_zuko_ncsf,
-    build_zuko_nice,
-    build_zuko_nsf,
-    build_zuko_sospf,
-    build_zuko_unaf,
-)
-from sbi.neural_nets.net_builders.mdn import build_mdn
-from sbi.utils.sbiutils import standardizing_net, z_score_parser
-from sbi.utils.user_input_checks import check_data_device
-
-model_builders = {
-    "mdn": build_mdn,
-    "made": build_made,
-    "maf": build_maf,
-    "maf_rqs": build_maf_rqs,
-    "nsf": build_nsf,
-    "zuko_nice": build_zuko_nice,
-    "zuko_maf": build_zuko_maf,
-    "zuko_nsf": build_zuko_nsf,
-    "zuko_ncsf": build_zuko_ncsf,
-    "zuko_sospf": build_zuko_sospf,
-    "zuko_naf": build_zuko_naf,
-    "zuko_unaf": build_zuko_unaf,
-    "zuko_gf": build_zuko_gf,
-    "zuko_bpf": build_zuko_bpf,
-}
+from .mixed_nets import build_mnle as _build_mnle
 
 
-def build_mnle(
-    batch_x: Tensor,
-    batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
-    flow_model: str = "nsf",
-    num_categorical_columns: Optional[Tensor] = None,
-    embedding_net: nn.Module = nn.Identity(),
-    combined_embedding_net: Optional[nn.Module] = None,
-    num_transforms: int = 2,
-    num_components: int = 5,
-    num_bins: int = 5,
-    hidden_features: int = 50,
-    hidden_layers: int = 2,
-    tail_bound: float = 10.0,
-    log_transform_x: bool = False,
-    **kwargs,
-):
-    """Returns a density estimator for mixed data types.
-
-    Uses an autoregressive categorical density estimator to model the discrete part
-    and a conditional density estimator to model the continuous part of the data.
-
-    Note: If the condition y is > 1D, an embedding net must be provided. Then,
-    during inference, we need to combine the embedded condition with the
-    discrete part of the input data. To do this, we use a combined embedding net
-    that takes the discrete part of the input and the embedded condition as
-    input and passes it on to the flow model.
-    To this end, we build the z-scoring and the embedding net in this function
-    here, i.e., outside of the flow model, and then pass z_score_x="none" to the
-    flow builder.
-    The y-embedding is passed to MixedDensityEstimator, which then uses it to
-    embed the continuous part and concatenate it with the discrete part during
-    log_prob evaluation and sampling.
-    The combined embedding net is passed to the flow model, which then uses it
-    to combine the already embedded and combined condition.
-
-    Args:
-        batch_x: Batch of xs, used to infer dimensionality.
-        batch_y: Batch of ys, used to infer dimensionality and (optional)
-            z-scoring.
[…]
