### 2025-01-09 af9423d
wip: build_mnpe integrated, build_mnle refactored, mnle_test fails; TODO: MNPE class + test
