### 2025-01-09 d04af8c
rm: rm legacy CategoricalNet
