### 2025-01-09 a498695
wip: added MNPE class and test case for it
