+++ sbi/inference/trainers/npe/__init__.py
@@ -1,3 +1,4 @@
+from sbi.inference.trainers.npe.mnpe import MNPE  # noqa: F401
 from sbi.inference.trainers.npe.npe_a import NPE_A  # noqa: F401
 from sbi.inference.trainers.npe.npe_b import NPE_B  # noqa: F401
 from sbi.inference.trainers.npe.npe_base import PosteriorEstimator  # noqa: F401
