### 2025-01-09 b9a5b92
wip: added MNPE class and test case for it
