+++ sbi/inference/trainers/npe/mnpe.py
@@ -1,0 +1,181 @@
+# This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
+# under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+
+from copy import deepcopy
+from typing import Any, Callable, Dict, Optional, Union
+
+from torch.distributions import Distribution
+
+from sbi.inference.posteriors import MCMCPosterior, RejectionPosterior, VIPosterior
+from sbi.inference.potentials import posterior_estimator_based_potential
+from sbi.inference.trainers.npe.npe_c import NPE_C
+from sbi.neural_nets.estimators import MixedDensityEstimator
+from sbi.sbi_types import TensorboardSummaryWriter, TorchModule
+from sbi.utils.sbiutils import del_entries
+from sbi.utils.user_input_checks import check_prior
+
+
+class MNPE(NPE_C):
+    def __init__(
+        self,
+        prior: Optional[Distribution] = None,
+        density_estimator: Union[str, Callable] = "mnpe",
+        device: str = "cpu",
+        logging_level: Union[int, str] = "WARNING",
+        summary_writer: Optional[TensorboardSummaryWriter] = None,
+        show_progress_bars: bool = True,
+    ):
+        r"""Mixed Neural Posterior Estimation (MNPE).
+
+        Like NPE-C, but designed to be applied to data with mixed types, e.g.,
+        continuous parameters and discrete parameters like they occur in models with
+        switching components.
+
+        Args:
+            prior: A probability distribution that expresses prior knowledge about the
+                parameters, e.g. which ranges are meaningful for them. If `None`, the
+                prior must be passed to `.build_posterior()`.
+            density_estimator: If it is a string, it must be "mnpe" to use the
+                preconfigured neural nets for MNPE. Alternatively, a function
+                that builds a custom neural network can be provided. The function will
+                be called with the first batch of simulations (theta, x), which can
+                thus be used for shape inference and potentially for z-scoring. It
+                needs to return a PyTorch `nn.Module` implementing the density
+                estimator. The density estimator needs to provide the methods
+                `.log_prob`, `.log_prob_iid()` and `.sample()`.
+            device: Training device, e.g., "cpu", "cuda" or "cuda:{0, 1, ...}".
+            logging_level: Minimum severity of messages to log. One of the strings
+                INFO, WARNING, DEBUG, ERROR and CRITICAL.
+            summary_writer: A tensorboard `SummaryWriter` to control, among others, log
+                file location (default is `<current working directory>/logs`.)
+            show_progress_bars: Whether to show a progressbar during simulation and
+                sampling.
+        """
+
+        if isinstance(density_estimator, str):
+            assert (
+                density_estimator == "mnpe"
+            ), f"""MNPE can be used with preconfigured 'mnpe' density estimator only,
+                not with {density_estimator}."""
+        kwargs = del_entries(locals(), entries=("self", "__class__"))
+        super().__init__(**kwargs)
+
+    def train(
+        self,
+        training_batch_size: int = 200,
+        learning_rate: float = 5e-4,
+        validation_fraction: float = 0.1,
+        stop_after_epochs: int = 20,
+        max_num_epochs: int = 2**31 - 1,
+        clip_max_norm: Optional[float] = 5.0,
+        resume_training: bool = False,
+        discard_prior_samples: bool = False,
+        retrain_from_scratch: bool = False,
+        show_train_summary: bool = False,
+        dataloader_kwargs: Optional[Dict] = None,
+    ) -> MixedDensityEstimator:
+        density_estimator = super().train(
+            **del_entries(locals(), entries=("self", "__class__"))
+        )
+        assert isinstance(
+            density_estimator, MixedDensityEstimator
+        ), f"""Internal net must be of type
[…]
