+++ tests/mnpe_test.py
@@ -1,0 +1,255 @@
+# This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
+# under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+
+import pytest
+import torch
+from torch import Tensor
+from torch.distributions import Bernoulli, Beta, Normal
+
+from sbi.inference import MNPE, MCMCPosterior
+from sbi.inference.posteriors.rejection_posterior import RejectionPosterior
+from sbi.inference.posteriors.vi_posterior import VIPosterior
+from sbi.neural_nets import posterior_nn
+from sbi.neural_nets.embedding_nets import FCEmbedding
+from sbi.utils import mcmc_transform
+from sbi.utils.torchutils import process_device
+from sbi.utils.user_input_checks_utils import MultipleIndependent
+from tests.test_utils import check_c2st
+
+
+# toy simulator for continuous data with mixed parameters
+def mixed_param_simulator(theta: Tensor) -> Tensor:
+    """Simulator for continuous data with mixed parameters.
+
+    Args:
+        theta: Parameters with mixed types - continuous and discrete.
+               First component is continuous (Beta distributed)
+               Second component is discrete (0 or 1)
+
+    Returns:
+        x: Continuous observation drawn from a mixture of two normals.
+    """
+    # Extract parameters
+    mixing_weight, component = theta[:, 0], theta[:, 1]
+
+    # Generate data from mixture of normals
+    # component 0: N(0,1), component 1: N(3,1)
+    means = torch.where(component > 0.5, torch.tensor(3.0), torch.tensor(0.0))
+
+    # Use mixing weight to determine variance
+    stds = 0.5 + mixing_weight
+
+    return Normal(means, stds).sample().reshape(-1, 1)
+
+
+@pytest.mark.mcmc
+@pytest.mark.gpu
+@pytest.mark.parametrize("device", ("cpu", "gpu"))
+def test_mnpe_on_device(
+    device,
+    mcmc_params_fast: dict,
+    num_simulations: int = 100,
+    mcmc_method: str = "slice_np",
+):
+    """Test MNPE API on device."""
+
+    device = process_device(device)
+
+    # Generate data with mixed parameter types
+    theta = torch.cat(
+        (
+            torch.rand(num_simulations, 1),  # continuous
+            torch.randint(0, 2, (num_simulations, 1)),  # discrete
+        ),
+        dim=1,
+    ).to(device)
+    x = mixed_param_simulator(theta).to(device)
+
+    # Train and infer
+    prior = MultipleIndependent([
+        Beta(torch.ones(1), torch.ones(1)),  # continuous
+        Bernoulli(probs=0.5 * torch.ones(1)),  # discrete (Bernoulli)
+    ]).to(device)
+
+    trainer = MNPE(prior=prior, device=device)
+    trainer.append_simulations(theta, x).train(max_num_epochs=1)
+
+    # Test sampling on device
+    posterior = trainer.build_posterior()
+    posterior.sample(
+        (1,),
+        x=x[0],
+        show_progress_bars=False,
+        mcmc_method=mcmc_method,
+        **mcmc_params_fast,
+    )
+
+
+@pytest.mark.parametrize(
+    "sampler", (pytest.param("mcmc", marks=pytest.mark.mcmc), "rejection", "vi")
+)
+@pytest.mark.parametrize("flow_model", ("mdn", "maf", "nsf", "zuko_nsf"))
+@pytest.mark.parametrize("z_score_theta", ("independent", "none"))
+def test_mnpe_api(flow_model: str, sampler, mcmc_params_fast: dict, z_score_theta: str):
+    """Test MNPE API."""
+
+    # Generate data
[…]
