### 2025-01-09 #7428 @boisgera
Hi @janfb,

I wish you a happy new year!

I am sorry for the delay here, I have been far too busy to handle JOSS matters properly at the end of 2024.

Let me tell you where we stand: I have received a majority of "no" and a few "maybe but later" to my first batch of proposals to review your project (nothing suprising unfortunately, and not connected at all with the quality of your submission!).

I am contacting again the "maybes" and opening the proposal to a few other potential reviewers. I hope that we'll be ready to start the review soon!

Best regards,

Sébastien
