### 2025-01-09 05b4c3b
fix: correct i/o shapes, updated tutorial
