### 2025-01-10 fe274c9
wip: added MNPE class and test case for it, not working yet
