### 2025-01-10 a1ccd82
wip: added MNPE class and test case for it, not working yet
