+++ tests/mnpe_test.py
@@ -4,17 +4,16 @@
 import pytest
 import torch
 from torch import Tensor
-from torch.distributions import Bernoulli, Beta, Normal
+from torch.distributions import Bernoulli
 
-from sbi.inference import MNPE, MCMCPosterior
+from sbi.inference import MNPE
 from sbi.inference.posteriors.rejection_posterior import RejectionPosterior
 from sbi.inference.posteriors.vi_posterior import VIPosterior
 from sbi.neural_nets import posterior_nn
 from sbi.neural_nets.embedding_nets import FCEmbedding
-from sbi.utils import mcmc_transform
+from sbi.utils import BoxUniform
 from sbi.utils.torchutils import process_device
 from sbi.utils.user_input_checks_utils import MultipleIndependent
-from tests.test_utils import check_c2st
 
 
 # toy simulator for continuous data with mixed parameters
@@ -23,94 +22,110 @@ def mixed_param_simulator(theta: Tensor) -> Tensor:
 
     Args:
         theta: Parameters with mixed types - continuous and discrete.
-               First component is continuous (Beta distributed)
-               Second component is discrete (0 or 1)
-
     Returns:
-        x: Continuous observation drawn from a mixture of two normals.
+        x: Continuous observation.
     """
+    device = theta.device
+
     # Extract parameters
-    mixing_weight, component = theta[:, 0], theta[:, 1]
-
-    # Generate data from mixture of normals
-    # component 0: N(0,1), component 1: N(3,1)
-    means = torch.where(component > 0.5, torch.tensor(3.0), torch.tensor(0.0))
-
-    # Use mixing weight to determine variance
-    stds = 0.5 + mixing_weight
-
-    return Normal(means, stds).sample().reshape(-1, 1)
+    a, b = theta[:, 0], theta[:, 1]
+    noise = 0.05 * torch.randn(a.shape, device=device).reshape(-1, 1)
+    return (a + 2 * b).reshape(-1, 1) + noise
 
 
-@pytest.mark.mcmc
 @pytest.mark.gpu
 @pytest.mark.parametrize("device", ("cpu", "gpu"))
 def test_mnpe_on_device(
     device,
-    mcmc_params_fast: dict,
     num_simulations: int = 100,
-    mcmc_method: str = "slice_np",
 ):
-    """Test MNPE API on device."""
-
+    """Test MNPE API on device (CPU/GPU)."""
+    num_samples = 10
     device = process_device(device)
 
     # Generate data with mixed parameter types
-    theta = torch.cat(
-        (
-            torch.rand(num_simulations, 1),  # continuous
-            torch.randint(0, 2, (num_simulations, 1)),  # discrete
-        ),
-        dim=1,
-    ).to(device)
+    theta_true = torch.tensor([[0.5, 1.0]], device=device)
+    x_o = mixed_param_simulator(theta_true).to(device)
+
+    prior = MultipleIndependent(
+        [
+            BoxUniform(low=torch.zeros(1), high=torch.ones(1)),  # continuous
+            Bernoulli(probs=0.8 * torch.ones(1)),  # discrete (Bernoulli)
+        ],
+        device=device,
+    )
+    theta = prior.sample((num_simulations,))
     x = mixed_param_simulator(theta).to(device)
 
-    # Train and infer
-    prior = MultipleIndependent([
-        Beta(torch.ones(1), torch.ones(1)),  # continuous
-        Bernoulli(probs=0.5 * torch.ones(1)),  # discrete (Bernoulli)
-    ]).to(device)
-
     trainer = MNPE(prior=prior, device=device)
-    trainer.append_simulations(theta, x).train(max_num_epochs=1)
[…]
