+++ sbi/utils/restriction_estimator.py
@@ -692,6 +692,11 @@ class RestrictedPrior(Distribution):
                 max_sampling_batch_size=max_sampling_batch_size,
                 alternative_method="sample_with='sir'",
             )
+            # NOTE: This currently requires a float acceptance rate. As previous versions
+            # of accept_reject_sample returned a float. In favour to batched sampling
+            # it now returns a tensor.
+            acceptance_rate = acceptance_rate.min().item()
+
             if save_acceptance_rate:
                 self.acceptance_rate = torch.as_tensor(acceptance_rate)
             if print_rejected_frac:
