### 2025-01-10 7f319c3
Improving test to cover batched log_prob on bounded support priors
