### 2025-01-10 f385ee5
fix device in mog_log_prob, add test, fix tests.
