+++ sbi/utils/sbiutils.py
@@ -841,7 +841,9 @@ def mog_log_prob(
 
     # Split up evaluation into parts.
     weights = logits_pp - torch.logsumexp(logits_pp, dim=-1, keepdim=True)
-    constant = -(output_dim / 2.0) * torch.log(torch.tensor([2 * pi]))
+    constant = -(output_dim / 2.0) * torch.log(
+        torch.tensor([2 * pi], device=theta.device)
+    )
     log_det = 0.5 * torch.log(torch.det(precisions_pp))
     theta_minus_mean = theta.expand_as(means_pp) - means_pp
     exponent = -0.5 * batched_mixture_vmv(precisions_pp, theta_minus_mean)
