### 2025-01-10 06ba452
Improve tests to resolve circular imports
