+++ tests/circular_import_test.py
@@ -2,6 +2,10 @@ import importlib
 import inspect
 import pkgutil
 import random
+import contextlib
+import sys
+
+from torch import le
 
 
 def import_first_function(module_name: str):
@@ -25,11 +29,25 @@ def find_submodules(package_name):
     submodules = []
     package = __import__(package_name)
 
-    for _, name, _ in pkgutil.walk_packages(package.__path__):
-        full_name = package.__name__ + "." + name
-        submodules.append(full_name)
+    # Now crawls all submodules
+    def walk_submodules(package):
+        for _, name, is_pkg in pkgutil.walk_packages(
+            package.__path__, package.__name__ + "."
+        ):
+            submodules.append(name)
+            if is_pkg:
+                # There are some wanted import errors for deprecated modules
+                with contextlib.suppress(ImportError):
+                    walk_submodules(importlib.import_module(name))
 
+    walk_submodules(package)
     return submodules
+
+def reset_environment():
+    # This is a helper which resets the environment
+    for module_name in list(sys.modules.keys()):
+        if "sbi" in module_name:
+            del sys.modules[module_name]
 
 
 def test_for_circular_imports():
@@ -37,16 +55,25 @@ def test_for_circular_imports():
     # Permute the list of modules
     random.shuffle(modules)
 
+    errors = []
     for module_name in modules:
+
         # Try to import
         if "sbi.examples" in module_name:
             # This is not really a module :/ Hence skip it...
             continue
         try:
+            reset_environment()
             # Tests if we can: import module_name
             module = importlib.import_module(module_name)
+            reset_environment()
             # Tests if we can: from module_name import xxx
             import_first_function(module.__name__)
+
             del module
         except ImportError as e:
-            raise AssertionError(f"Cannot import {module_name}. Error: {e}") from e
+            if "circular import" in str(e):
+                errors.append(f"Circular import detected in {module_name}. Error: {e}")
+                print(f"Circular import detected in {module_name}")
+
+    assert len(errors) == 0, "\n".join(errors)
