### 2025-01-10 a6a220d
fix: leakage correction for log prob batched (#1355)
* Fixing leakage correction inconsistency.

* Improving test to cover batched log_prob on bounded support priors

* Fixing test

* Formating fix

* add type
