+++ sbi/samplers/rejection/rejection.py
@@ -362,6 +362,4 @@ def accept_reject_sample(
         samples.shape[0] == num_samples
     ), "Number of accepted samples must match required samples."
 
-    # NOTE: Restriction prior does currently require a float as return for the
-    # acceptance rate, which is why we for now also return the minimum acceptance rate.
-    return samples, as_tensor(min_acceptance_rate)
+    return samples, as_tensor(acceptance_rate)
