### 2025-01-10 70c88b2
add cpu test for multiround mdn
