### 2025-01-10 46fd891
update and pin ruff and pre-commit.
