+++ pyproject.toml
@@ -64,10 +64,10 @@ doc = [
 dev = [
     "ffmpeg",
     # Lint
-    "pre-commit == 3.5.0",
+    "pre-commit == 4.0.1",
     "pyyaml",
     "pyright",
-    "ruff>=0.3.3",
+    "ruff==0.9.0",
     # Test
     "pytest",
     "pytest-cov",
