+++ sbi/inference/trainers/npe/npe_a.py
@@ -568,9 +568,9 @@ class NPE_A_MDN(ConditionalDensityEstimator):
             Mixture components of the posterior.
         """
         # Remove the batch dimension of `x` (SNPE-A always has a single `x`).
-        assert (
-            x.shape[0] == 1
-        ), f"Batchsize of `x_o` == {x.shape[0]}. SNPE-A only supports a single `x_o`."
+        assert x.shape[0] == 1, (
+            f"Batchsize of `x_o` == {x.shape[0]}. SNPE-A only supports a single `x_o`."
+        )
         x = x.squeeze(dim=0)
 
         # Evaluate the density estimator.
