+++ sbi/analysis/sensitivity_analysis.py
@@ -250,9 +250,9 @@ class ActiveSubspace:
                 prevent exploding gradients. Use `None` for no clipping.
         """
 
-        assert (
-            self._theta is not None and self._emergent_property is not None
-        ), "You must call .add_property() first."
+        assert self._theta is not None and self._emergent_property is not None, (
+            "You must call .add_property() first."
+        )
 
         # Get indices for permutation of the data.
         num_examples = len(self._theta)
@@ -433,9 +433,9 @@ class ActiveSubspace:
         if posterior_log_prob_as_property:
             predictions = self._posterior.potential(thetas, track_gradients=True)
         else:
-            assert (
-                self._regression_net is not None
-            ), "self._regression_net is None, you must call `.train()` first."
+            assert self._regression_net is not None, (
+                "self._regression_net is None, you must call `.train()` first."
+            )
             predictions = self._regression_net.forward(thetas)
         loss = predictions.mean()
         loss.backward()
