### 2025-01-10 3480e0b
apply ruff to tests and tutorials, ignore long lines.
