### 2025-01-10 61a7617
add NaN and Inf check for diagnostics, adapt test.
