### 2025-01-10 0ab3a14
fix #1343 device handling in mog_log_prob (#1356)
* fix device in mog_log_prob, add test, fix tests.

* add cpu test for multiround mdn
