### 2025-01-10 0c6e8e2
increase tutorial test timeout
