+++ tests/tutorials_test.py
@@ -22,7 +22,7 @@ def test_tutorials(notebook_path):
     """Test that all notebooks in the tutorials directory can be executed."""
     with open(notebook_path) as f:
         nb = nbformat.read(f, as_version=4)
-        ep = ExecutePreprocessor(timeout=600, kernel_name='python3')
+        ep = ExecutePreprocessor(timeout=1200, kernel_name='python3')
         print(f"Executing notebook {notebook_path}")
         try:
             ep.preprocess(nb, {'metadata': {'path': os.path.dirname(notebook_path)}})
