### 2025-01-10 e755927
add nan check to _loss method.
