### 2025-01-10 6016f07
fix: tests, embned+maf not working
