### 2025-01-10 90b2836
fix: tests, embned+maf not working
