+++ tests/mnpe_test.py
@@ -110,7 +110,8 @@ def test_batched_sampling(num_simulations: int = 100):
 
 @pytest.mark.parametrize("flow_model", ("mdn", "maf", "nsf", "zuko_nsf"))
 @pytest.mark.parametrize("z_score_theta", ("independent", "none"))
-def test_mnpe_api(flow_model: str, z_score_theta: str):
+@pytest.mark.parametrize("use_embed_net", (True, False))
+def test_mnpe_api(flow_model: str, z_score_theta: str, use_embed_net: bool):
     """Test MNPE API."""
 
     # Generate data
@@ -127,12 +128,12 @@ def test_mnpe_api(flow_model: str, z_score_theta: str):
     x = mixed_param_simulator(theta)
 
     # Build estimator manually
-    theta_embedding = FCEmbedding(2, 2)  # simple embedding net
+    theta_embedding = FCEmbedding(1, 1)  # simple embedding net, 1 continuous parameter
     density_estimator = posterior_nn(
         model="mnpe",
         flow_model=flow_model,
         z_score_theta=z_score_theta,
-        embedding_net=theta_embedding,
+        embedding_net=theta_embedding if use_embed_net else torch.nn.Identity(),
         log_transform_x=False,
     )
     trainer = MNPE(density_estimator=density_estimator)
