+++ sbi/neural_nets/net_builders/mixed_nets.py
@@ -62,6 +62,7 @@ def _build_mixed_density_estimator(
     hidden_layers: int = 2,
     tail_bound: float = 10.0,
     log_transform_x: bool = True,
+    is_mnpe: bool = False,
     **kwargs,
 ) -> MixedDensityEstimator:
     """Base function for building mixed neural density estimators.
@@ -134,17 +135,30 @@ def _build_mixed_density_estimator(
     # Separate continuous and discrete data.
     cont_x, disc_x = _separate_input(batch_x)
 
-    # Set up y-embedding net with z-scoring.
-    z_score_y_bool, structured_y = z_score_parser(z_score_y)
-    if z_score_y_bool:
-        embedding_net = nn.Sequential(
-            standardizing_net(batch_y, structured_y), embedding_net
-        )
-    # Get size of the embedded condition data.
-    print("embedding_net", embedding_net)
-    print("batch_x.shape:", batch_x.shape)
-    print("batch_y.shape:", batch_y.shape)
-    embedded_batch_y = embedding_net(batch_y)
+    # The embedding net is applied to the continuous part of the parameters
+    # MNPE: parameters are in batch_x, MNLE: parameters are in batch_y
+    if is_mnpe:
+        z_score_x_bool, structured_x = z_score_parser(z_score_x)
+        if z_score_x_bool:
+            embedding_net_x = nn.Sequential(
+                standardizing_net(cont_x, structured_x), embedding_net
+            )
+        else:
+            embedding_net_x = embedding_net
+        embedding_net_y = nn.Identity()
+    else:
+        z_score_y_bool, structured_y = z_score_parser(z_score_y)
+        if z_score_y_bool:
+            embedding_net_y = nn.Sequential(
+                standardizing_net(batch_y, structured_y), embedding_net
+            )
+        else:
+            embedding_net_y = embedding_net
+        embedding_net_x = nn.Identity()
+
+    # embed
+    embedded_batch_x = embedding_net_x(cont_x)
+    embedded_batch_y = embedding_net_y(batch_y)
     combined_condition = torch.cat([disc_x, embedded_batch_y], dim=-1)
 
     # Set up a categorical RV neural net for modelling the discrete data.
@@ -155,7 +169,7 @@ def _build_mixed_density_estimator(
         z_score_y="none",
         num_hidden=hidden_features,
         num_layers=hidden_layers,
-        embedding_net=embedding_net,
+        embedding_net=embedding_net_y,
     )
 
     if combined_embedding_net is None:
@@ -172,7 +186,7 @@ def _build_mixed_density_estimator(
     continuous_net = model_builders[flow_model](
         # TODO: add support for optional log-transform in flow builders.
         batch_x=(
-            torch.log(cont_x) if log_transform_x else cont_x
+            torch.log(embedded_batch_x + 1e-10) if log_transform_x else embedded_batch_x
         ),  # log transform manually.
         batch_y=combined_condition,
         z_score_x=z_score_x,
@@ -190,7 +204,7 @@ def _build_mixed_density_estimator(
     return MixedDensityEstimator(
         discrete_net=discrete_net,
         continuous_net=continuous_net,
-        embedding_net=embedding_net,  # pass embedding for continuous condition part.
+        embedding_net=embedding_net_y,  # pass embedding for continuous condition part.
         log_transform_input=log_transform_x,
         input_shape=batch_x[0].shape,
         condition_shape=batch_y[0].shape,
@@ -214,7 +228,9 @@ def build_mnle(
     Returns:
         MixedDensityEstimator for MNLE.
     """
[…]
