### 2025-01-10 597da91
fix: add nan check to _loss methods. (#1361)
* add nan check to _loss method.

* add checks for nres
