### 2025-01-13 e1305b9
style: pin pre commit and ruff to recent versions. (#1358)
* update and pin ruff and pre-commit.

* linting and formatting with new ruff

* apply ruff to tests and tutorials, ignore long lines.
