+++ sbi/diagnostics/tarp.py
@@ -15,7 +15,7 @@ from torch import Tensor
 from sbi.inference.posteriors.base_posterior import NeuralPosterior
 from sbi.utils.diagnostics_utils import (
     get_posterior_samples_on_batch,
-    remove_nans_and_infs,
+    remove_nans_and_infs_in_x,
 )
 from sbi.utils.metrics import l2
 
@@ -65,7 +65,7 @@ def run_tarp(
         alpha: credibility values, see equation 2 of the paper
     """
 
-    thetas, xs = remove_nans_and_infs(thetas, xs)
+    thetas, xs = remove_nans_and_infs_in_x(thetas, xs)
 
     num_tarp_samples, dim_theta = thetas.shape
 
