+++ tests/bm_test.py
@@ -6,17 +6,70 @@ from sbi.utils.metrics import c2st
 
 from .mini_sbibm import get_task
 
-# The probably should be some user control on this
+# Global settings
 SEED = 0
 TASKS = ["two_moons", "linear_mvg_2d", "gaussian_linear", "slcp"]
 NUM_SIMULATIONS = 2000
 EVALUATION_POINTS = 4  # Currently only 3 observation tested for speed
+TRAIN_KWARGS = {}
 
-TRAIN_KWARGS = {
-    # "training_batch_size": 200,  # To speed up training
+# Density estimators to test
+DENSITY_estimators = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
+CLASSIFIERS = ["linear", "mlp", "resnet"]
+NNS = ["mlp", "resnet"]
+SCORE_ESTIMATORS = ["mlp", "ada_mlp"]
+
+# Benchmarking method groups
+ALL_METHODS = [NPE, NRE, NLE, FMPE, NPSE]
+METHOD_GROUPS = {
+    "none": ALL_METHODS,
+    "npe": [NPE],
+    "nle": [NLE],
+    "nre": [NRE],
+    "fmpe": [FMPE],
+    "npse": [NPSE],
+}
+METHOD_PARAMS = {
+    "none": [{}],
+    "npe": [{"density_estimator": de} for de in DENSITY_estimators],
+    "nle": [{"density_estimator": de} for de in ["maf", "nsf"]],
+    "nre": [{"classifier": cl} for cl in CLASSIFIERS],
+    "fmpe": [{"density_estimator": nn} for nn in NNS],
+    "npse": [
+        {"score_estimator": nn, "sde_type": sde}
+        for nn in SCORE_ESTIMATORS
+        for sde in ["ve", "vp"]
+    ],
 }
 
-# Amortized benchmarking
+
+@pytest.fixture
+def method_list(benchmark_mode):
+    name = str(benchmark_mode).lower()
+    if name not in METHOD_GROUPS:
+        raise ValueError(f"Benchmark mode '{benchmark_mode}' is not supported.")
+    return METHOD_GROUPS[name]
+
+
+@pytest.fixture
+def kwargs_list(benchmark_mode):
+    name = str(benchmark_mode).lower()
+    if name not in METHOD_PARAMS:
+        raise ValueError(f"Benchmark mode '{benchmark_mode}' is not supported.")
+    return METHOD_PARAMS[name]
+
+
+# Use pytest.mark.parametrize dynamically
+# Generates a list of methods to test based on the benchmark mode
+def pytest_generate_tests(metafunc):
+    if "inference_method" in metafunc.fixturenames:
+        method_list = metafunc.config.getoption("--bm-mode")
+        method_group = METHOD_GROUPS.get(method_list, [])
+        metafunc.parametrize("inference_method", method_group)
+    if "extra_kwargs" in metafunc.fixturenames:
+        kwargs_list = metafunc.config.getoption("--bm-mode")
+        kwargs_group = METHOD_PARAMS.get(kwargs_list, [])
+        metafunc.parametrize("extra_kwargs", kwargs_group)
 
 
 def standard_eval_c2st_loop(posterior, task):
@@ -36,22 +89,13 @@ def standard_eval_c2st_loop(posterior, task):
     return mean_c2st
 
 
-DENSITY_estimators = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
-DENSITY_estimators = ["maf", "nsf"]  # Fast
-
-
-@pytest.mark.benchmark
-@pytest.mark.parametrize('task_name', TASKS, ids=str)
-@pytest.mark.parametrize('density_estimator', DENSITY_estimators, ids=str)
[…]
