+++ tests/bm_test.py
@@ -2,10 +2,13 @@ import pytest
 import torch
 
 from sbi.inference import FMPE, NLE, NPE, NPSE, NRE
+from sbi.inference.trainers.npe import NPE_C
+from sbi.inference.trainers.nre import BNRE, NRE_A, NRE_B, NRE_C
 from sbi.utils.metrics import c2st
 
 from .mini_sbibm import get_task
 
+# NOTE: This might can be improved...
 # Global settings
 SEED = 0
 TASKS = ["two_moons", "linear_mvg_2d", "gaussian_linear", "slcp"]
@@ -15,19 +18,21 @@ TRAIN_KWARGS = {}
 
 # Density estimators to test
 DENSITY_estimators = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
-CLASSIFIERS = ["linear", "mlp", "resnet"]
+CLASSIFIERS = ["mlp", "resnet"]
 NNS = ["mlp", "resnet"]
 SCORE_ESTIMATORS = ["mlp", "ada_mlp"]
 
 # Benchmarking method groups
-ALL_METHODS = [NPE, NRE, NLE, FMPE, NPSE]
 METHOD_GROUPS = {
-    "none": ALL_METHODS,
+    "none": [NPE, NRE, NLE, FMPE, NPSE],
     "npe": [NPE],
     "nle": [NLE],
-    "nre": [NRE],
+    "nre": [NRE_A, NRE_B, NRE_C, BNRE],
     "fmpe": [FMPE],
     "npse": [NPSE],
+    "snpe": [NPE_C],  # NPE_B not implemented, NPE_A need Gaussian prior
+    "snle": [NLE],
+    "snre": [NRE_A, NRE_B, NRE_C, BNRE],
 }
 METHOD_PARAMS = {
     "none": [{}],
@@ -40,11 +45,23 @@ METHOD_PARAMS = {
         for nn in SCORE_ESTIMATORS
         for sde in ["ve", "vp"]
     ],
+    "snpe": [{}],
+    "snle": [{}],
+    "snre": [{}],
 }
 
 
 @pytest.fixture
-def method_list(benchmark_mode):
+def method_list(benchmark_mode: str) -> list:
+    """
+    Fixture to get the list of methods based on the benchmark mode.
+
+    Args:
+        benchmark_mode (str): The benchmark mode.
+
+    Returns:
+        list: List of methods for the given benchmark mode.
+    """
     name = str(benchmark_mode).lower()
     if name not in METHOD_GROUPS:
         raise ValueError(f"Benchmark mode '{benchmark_mode}' is not supported.")
@@ -52,7 +69,16 @@ def method_list(benchmark_mode):
 
 
 @pytest.fixture
-def kwargs_list(benchmark_mode):
+def kwargs_list(benchmark_mode: str) -> list:
+    """
+    Fixture to get the list of kwargs based on the benchmark mode.
+
+    Args:
+        benchmark_mode (str): The benchmark mode.
+
+    Returns:
+        list: List of kwargs for the given benchmark mode.
+    """
     name = str(benchmark_mode).lower()
     if name not in METHOD_PARAMS:
         raise ValueError(f"Benchmark mode '{benchmark_mode}' is not supported.")
@@ -62,25 +88,38 @@ def kwargs_list(benchmark_mode):
 # Use pytest.mark.parametrize dynamically
 # Generates a list of methods to test based on the benchmark mode
 def pytest_generate_tests(metafunc):
+    """
+    Dynamically generates a list of methods to test based on the benchmark mode.
+
+    Args:
+        metafunc: The metafunc object from pytest.
+    """
     if "inference_method" in metafunc.fixturenames:
[…]

[…]
