### 2025-01-13 cefdaa7
Ruff format compatible -> Ignore string typing
