+++ sbi/samplers/rejection/__init__.py
@@ -1,0 +1,4 @@
+from sbi.samplers.rejection.rejection import (
+    accept_reject_sample,
+    rejection_sample,
+)
