+++ sbi/samplers/vi/__init__.py
@@ -4,10 +4,3 @@ from sbi.samplers.vi.vi_divergence_optimizers import (
 )
 from sbi.samplers.vi.vi_pyro_flows import get_default_flows, get_flow_builder
 from sbi.samplers.vi.vi_quality_control import get_quality_metric
-from sbi.samplers.vi.vi_utils import (
-    adapt_variational_distribution,
-    check_variational_distribution,
-    detach_all_non_leaf_tensors,
-    make_object_deepcopy_compatible,
-    move_all_tensor_to_device,
-)
