+++ sbi/samplers/score/__init__.py
@@ -1,3 +1,3 @@
-from sbi.samplers.score.correctors import Corrector, get_corrector
 from sbi.samplers.score.predictors import Predictor, get_predictor
-from sbi.samplers.score.score import Diffuser
+from sbi.samplers.score.correctors import Corrector, get_corrector
+from sbi.samplers.score.diffuser import Diffuser
