### 2025-01-13 961792d
Ignore types given as string
