+++ sbi/samplers/score/diffuser.py
@@ -7,9 +7,8 @@ import torch
 from torch import Tensor
 from tqdm.auto import tqdm
 
-
-from sbi.samplers.score.predictors import get_predictor, Predictor
-from sbi.samplers.score.correctors import get_corrector, Corrector
+from sbi.samplers.score.correctors import Corrector, get_corrector
+from sbi.samplers.score.predictors import Predictor, get_predictor
 
 
 class Diffuser:
@@ -18,7 +17,7 @@ class Diffuser:
 
     def __init__(
         self,
-        score_based_potential: 'PosteriorScoreBasedPotential',
+        score_based_potential: 'PosteriorScoreBasedPotential',  # noqa: F821
         predictor: Union[str, Predictor],
         corrector: Optional[Union[str, Corrector]] = None,
         predictor_params: Optional[dict] = None,
@@ -63,7 +62,7 @@ class Diffuser:
     def set_predictor(
         self,
         predictor: Union[str, Predictor],
-        score_based_potential: 'PosteriorScoreBasedPotential',
+        score_based_potential: 'PosteriorScoreBasedPotential',  # noqa: F821
         **kwargs,
     ):
         """Set the predictor for the diffusion-based sampler."""
