### 2025-01-13 6210fe2
Also pyright ignore string types
