+++ sbi/samplers/score/predictors.py
@@ -12,7 +12,7 @@ PREDICTORS = {}
 
 def get_predictor(
     name: str,
-    score_based_potential: 'PosteriorScoreBasedPotential',  # noqa: F821
+    score_based_potential: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
     **kwargs,
 ) -> "Predictor":
     """Helper function to get predictor by name.
@@ -52,7 +52,7 @@ class Predictor(ABC):
 
     def __init__(
         self,
-        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821
+        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
     ):
         """Initialize predictor.
 
@@ -92,7 +92,7 @@ class Predictor(ABC):
 class EulerMaruyama(Predictor):
     def __init__(
         self,
-        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821
+        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
         eta: float = 1.0,
     ):
         """Simple Euler-Maruyama discretization of the associated family of reverse
