### 2025-01-13 76c1e1b
fix: add nan handling in diagnostics (#1359)
* add NaN and Inf check for diagnostics, adapt test.

* feedback
