### 2025-01-17 9ff0484
wip: fixed discrete data issue in mcmc_transform with MultipleIndependent; embedding net in mnpe not working yet
