### 2025-01-17 736e3ac
wip: fixed discrete data issue in mcmc_transform with MultipleIndependent; embedding net in mnpe not working yet
