### 2025-01-17 b4f4ab7
wip: remove unnecessary helper function (introduced while working on this PR)
