### 2025-01-17 0934e86
wip: remove unnecessary helper function (introduced while working on this PR)
