+++ sbi/utils/sbiutils.py
@@ -661,80 +661,6 @@ def match_theta_and_x_batch_shapes(theta: Tensor, x: Tensor) -> Tuple[Tensor, Te
     return theta_repeated, x_repeated
 
 
-class CustomConcatTransform(torch_tf.Transform):
-    """Concatenates multiple transforms and applies each to specific dimensions."""
-
-    def __init__(self, transforms: List[TorchTransform], dims: List[int]) -> None:
-        super().__init__()
-
-        self.transforms = transforms
-        self.dims = []
-        current_dim = 0
-        for dim in dims:
-            self.dims.append(slice(current_dim, current_dim + dim))
-            current_dim += dim
-
-        # Cache for inverse transform
-        self._inv = None
-
-    def _call(self, x: Tensor) -> Tensor:
-        """Apply each transform to its respective dimensions."""
-        results = []
-        for transform, dim_slice in zip(self.transforms, self.dims):
-            results.append(transform(x[..., dim_slice]))
-        return torch.cat(results, dim=-1)
-
-    def _inverse(self, y: Tensor) -> Tensor:
-        """Apply inverse of each transform to its respective dimensions."""
-        results = []
-        for transform, dim_slice in zip(self.transforms, self.dims):
-            results.append(transform.inv(y[..., dim_slice]))
-        return torch.cat(results, dim=-1)
-
-    def log_abs_det_jacobian(self, x: Tensor, y: Tensor) -> Tensor:
-        """Sum the log det jacobians of individual transforms."""
-        results = []
-        for transform, dim_slice in zip(self.transforms, self.dims):
-            results.append(
-                transform.log_abs_det_jacobian(x[..., dim_slice], y[..., dim_slice])
-            )
-        return torch.sum(torch.stack(results), dim=0)
-
-    @property
-    def domain(self) -> constraints.Constraint:
-        """Get the domain constraint of the transform."""
-        # For concatenated transforms, we use the most restrictive domain
-        # This is typically constraints.real for standardization transforms
-        return constraints.real
-
-    @property
-    def codomain(self) -> constraints.Constraint:
-        """Get the codomain constraint of the transform."""
-        # For concatenated transforms, we use the most restrictive codomain
-        # This is typically constraints.real for standardization transforms
-        return constraints.real
-
-    def with_cache(self, cache_size: int = 1) -> "CustomConcatTransform":
-        """Returns a transform with caching enabled."""
-        transforms_with_cache = [
-            t.with_cache(cache_size) if hasattr(t, "with_cache") else t
-            for t in self.transforms
-        ]
-        return CustomConcatTransform(
-            transforms_with_cache, [d.stop - d.start for d in self.dims]
-        )
-
-    @property
-    def inv(self) -> "CustomConcatTransform":
-        """Get the inverse transform."""
-        if self._inv is None:
-            self._inv = CustomConcatTransform(
-                [t.inv for t in self.transforms], [d.stop - d.start for d in self.dims]
-            )
-            self._inv._inv = self
-        return self._inv
-
-
 def mcmc_transform(
     prior: Distribution,
     num_prior_samples_for_zscoring: int = 1000,
