### 2025-01-20 fb87e3b
Add better documentation of the test
