+++ tests/circular_import_test.py
@@ -4,11 +4,18 @@ import inspect
 import pkgutil
 import random
 import sys
+from types import ModuleType
 
 
 def import_first_function(module_name: str):
-    # This is a helper which emulates the following:
-    # from sbi.xxx import yyy
+    """This is a helper which imports the first function from a module.
+
+    Performs: `from sbi.module_name import first_function_name`
+
+    Args:
+        module_name: Name of the module.
+    """
+
     module = importlib.import_module(module_name)
     functions = inspect.getmembers(module, inspect.isfunction)
 
@@ -22,13 +29,28 @@ def import_first_function(module_name: str):
         pass
 
 
-def find_submodules(package_name):
-    # This is a helper which finds all modules from which we could import
+def reset_environment():
+    """This is a helper which resets the environment by deleting all sbi modules."""
+    for module_name in list(sys.modules.keys()):
+        if "sbi" in module_name:
+            del sys.modules[module_name]
+
+
+def find_submodules(package_name: str):
+    """This is a helper which finds all submodules of a package.
+
+    Args:
+        package_name: Name of the package.
+    """
     submodules = []
     package = __import__(package_name)
 
-    # Now crawls all submodules
-    def walk_submodules(package):
+    def walk_submodules(package: ModuleType):
+        """This is a recursive helper function which walks all submodules of a package.
+
+        Args:
+            package: The package to crawl through.
+        """
         for _, name, is_pkg in pkgutil.walk_packages(
             package.__path__, package.__name__ + "."
         ):
@@ -42,14 +64,14 @@ def find_submodules(package_name):
     return submodules
 
 
-def reset_environment():
-    # This is a helper which resets the environment
-    for module_name in list(sys.modules.keys()):
-        if "sbi" in module_name:
-            del sys.modules[module_name]
-
-
 def test_for_circular_imports():
+    """This test checks for circular imports in the sbi package.
+
+    In order to do so, it is tested if we can directly import from all submodules i.e.
+    `from sbi.module_name import first_function_name` or `import sbi.module_name`
+    without any import errors.
+
+    """
     modules = find_submodules("sbi")
     # Permute the list of modules
     random.shuffle(modules)
@@ -70,7 +92,9 @@ def test_for_circular_imports():
 
             del module
         except ImportError as e:
+            # NOTE: There might be other errors which are intended
             if "circular import" in str(e):
+                # This is a circular import detected
                 errors.append(f"Circular import detected in {module_name}. Error: {e}")
                 print(f"Circular import detected in {module_name}")
 
