### 2025-01-20 45b8fcf
Fixing suggestions and adding dependency
