+++ tests/bm_test.py
@@ -1,12 +1,15 @@
 import pytest
 import torch
+from pytest_harvest import ResultsBag
 
 from sbi.inference import FMPE, NLE, NPE, NPSE, NRE
+from sbi.inference.posteriors.base_posterior import NeuralPosterior
 from sbi.inference.trainers.npe import NPE_C
 from sbi.inference.trainers.nre import BNRE, NRE_A, NRE_B, NRE_C
 from sbi.utils.metrics import c2st
 
 from .mini_sbibm import get_task
+from .mini_sbibm.base_task import Task
 
 # NOTE: This might can be improved...
 # Global settings
@@ -14,15 +17,17 @@ SEED = 0
 TASKS = ["two_moons", "linear_mvg_2d", "gaussian_linear", "slcp"]
 NUM_SIMULATIONS = 2000
 EVALUATION_POINTS = 4  # Currently only 3 observation tested for speed
+NUM_ROUNDS_SEQUENTIAL = 2
+EVALUATION_POINT_SEQUENTIAL = 1
 TRAIN_KWARGS = {}
 
 # Density estimators to test
-DENSITY_estimators = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
+DENSITY_ESTIMATORS = ["mdn", "made", "maf", "nsf", "maf_rqs"]  # "Kinda exhaustive"
 CLASSIFIERS = ["mlp", "resnet"]
 NNS = ["mlp", "resnet"]
 SCORE_ESTIMATORS = ["mlp", "ada_mlp"]
 
-# Benchmarking method groups
+# Benchmarking method groups i.e. what to run for different --bm-mode
 METHOD_GROUPS = {
     "none": [NPE, NRE, NLE, FMPE, NPSE],
     "npe": [NPE],
@@ -36,7 +41,7 @@ METHOD_GROUPS = {
 }
 METHOD_PARAMS = {
     "none": [{}],
-    "npe": [{"density_estimator": de} for de in DENSITY_estimators],
+    "npe": [{"density_estimator": de} for de in DENSITY_ESTIMATORS],
     "nle": [{"density_estimator": de} for de in ["maf", "nsf"]],
     "nre": [{"classifier": cl} for cl in CLASSIFIERS],
     "fmpe": [{"density_estimator": nn} for nn in NNS],
@@ -106,7 +111,7 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize("extra_kwargs", kwargs_group)
 
 
-def standard_eval_c2st_loop(posterior, task) -> float:
+def standard_eval_c2st_loop(posterior: NeuralPosterior, task: Task) -> float:
     """
     Evaluates the C2ST metric for the given posterior and task.
 
@@ -117,18 +122,23 @@ def standard_eval_c2st_loop(posterior, task) -> float:
     Returns:
         float: The mean C2ST value.
     """
-    metrics = []
+    c2st_scores = []
     for i in range(1, EVALUATION_POINTS):
         c2st_val = eval_c2st(posterior, task, i)
-        metrics.append(c2st_val)
+        c2st_scores.append(c2st_val)
 
-    mean_c2st = sum(metrics) / len(metrics)
+    mean_c2st = sum(c2st_scores) / len(c2st_scores)
     # Convert to float rounded to 3 decimal places
     mean_c2st = float(f"{mean_c2st:.3f}")
     return mean_c2st
 
 
-def eval_c2st(posterior, task, i: int) -> float:
+def eval_c2st(
+    posterior: NeuralPosterior,
+    task: Task,
+    idx_observation: int,
+    num_samples: int = 1000,
+) -> float:
     """
     Evaluates the C2ST metric for a specific observation.
 
[…]
