+++ tests/conftest.py
@@ -36,16 +36,18 @@ def pytest_collection_modifyitems(config, items):
     )
     if not gpu_device_available:
         skip_gpu = pytest.mark.skip(reason="No devices available")
-        skip_bm = pytest.mark.skip(reason="Benchmarking disabled")
+
         for item in items:
             if "gpu" in item.keywords:
                 item.add_marker(skip_gpu)
 
-            if not config.getoption("--bm"):
+    if not config.getoption("--bm"):
+        # Skip marked benchmarking tests
+        skip_bm = pytest.mark.skip(reason="Benchmarking disabled")
+        for item in items:
+            if "benchmark" in item.keywords:
                 item.add_marker(skip_bm)
-
-    # Filter tests to only those with the 'benchmark' marker
-    if config.getoption("--bm"):
+    else:
         # Filter tests to only those with the 'benchmark' marker
         filtered_items = []
         for item in items:
