### 2025-01-20 a163535
fix shape and device bugs.
