### 2025-01-20 80740b2
Fix failing tutorials, change MNLE default for log_transform to False (#1367)
* fix fmpe typo, add nle seed

* fix: change default log transfrom to False.

This was set to True for positive reaction times but does not hold in general.

* small nb fixes
