### 2025-01-21 9ecc68b
fix: do not test batched and iid x
