### 2025-01-22 48747f0
fix coverage and testing bugs
