### 2025-01-22 448cef2
fix #1364: conditional posterior shape and device bugs. (#1373)
* fix shape and device bugs.

* fix: do not test batched and iid x

* fix coverage and testing bugs
