### 2025-01-28 2d97940
change workflow to only the failing test
