### 2025-01-28 6f43a49
reset env at the end of circular import test
