### 2025-01-28 e84f600
hard reset environment
