+++ tests/circular_import_test.py
@@ -28,11 +28,19 @@ def import_first_function(module_name: str):
     else:
         pass
 
-
 def reset_environment():
-    """This is a helper which resets the environment by deleting all sbi modules."""
+    """This is a helper which resets the environment by deleting all modules except required ones."""
+    required_modules = {
+        "sys",
+        "importlib",
+        "inspect",
+        "pkgutil",
+        "contextlib",
+        "random",
+        "types",
+    }
     for module_name in list(sys.modules.keys()):
-        if "sbi" in module_name:
+        if module_name not in required_modules:
             del sys.modules[module_name]
 
 
