### 2025-01-28 71380b8
sampling via diffusion twice
