### 2025-01-28 c5f57b4
strats with sbi reset
