### 2025-01-28 sbi-dev/sbi#1378 @dgedon
Testing speed, mark as slow?
Currently, we have a great set of tests which take a considerable amount of time for PRs. Potentially we want to mark some more tests as "slow" for the hackathon to ensure faster feedback during development. The goal would be to prioritize core functionality tests for PR validation while allowing more time-intensive tests to run separately at a later stage.
