### 2025-01-29 bcea468
batched sampling for score-based posteriors
