+++ sbi/inference/potentials/score_based_potential.py
@@ -85,9 +85,6 @@ class PosteriorScoreBasedPotential(BasePotential):
             x_density_estimator = reshape_to_batch_event(
                 self.x_o, event_shape=self.score_estimator.condition_shape
             )
-            assert x_density_estimator.shape[0] == 1, (
-                "PosteriorScoreBasedPotential supports only x batchsize of 1`."
-            )
             # For large number of evals, we want a high-tolerance flow.
             # This flow will be used mainly for MAP calculations, hence we want to save
             # it instead of rebuilding it every time.
