+++ sbi/samplers/rejection/rejection.py
@@ -190,6 +190,7 @@ def accept_reject_sample(
     proposal: Callable,
     accept_reject_fn: Callable,
     num_samples: int,
+    num_xos: int = 1,
     show_progress_bars: bool = False,
     warn_acceptance: float = 0.01,
     sample_for_correction_factor: bool = False,
@@ -218,6 +219,8 @@ def accept_reject_sample(
             rejected. Must take a batch of parameters and return a boolean tensor which
             indicates which parameters get accepted.
         num_samples: Desired number of samples.
+        num_xos: Number of conditions for batched_sampling (currently only accepting
+            one batch dimension for the condition).
         show_progress_bars: Whether to show a progressbar during sampling.
         warn_acceptance: A minimum acceptance rate under which to warn about slowness.
         sample_for_correction_factor: True if this function was called by
@@ -263,8 +266,6 @@ def accept_reject_sample(
     # But this would require giving the method the condition_shape explicitly...
     if "condition" in proposal_sampling_kwargs:
         num_xos = proposal_sampling_kwargs["condition"].shape[0]
-    else:
-        num_xos = 1
 
     accepted = [[] for _ in range(num_xos)]
     acceptance_rate = torch.full((num_xos,), float("Nan"))
