### 2025-01-29 6d527f7
Improve tests to detect circular imports and resolve all of them (#1357)
* Improve tests to resolve circular imports

* ruff format

* Ruff format compatible -> Ignore string typing

* Ignore types given as string

* Also pyright ignore string types

* Add better documentation of the test

* Test should work

* Fix comment

* torch < 2.6
