+++ sbi/samplers/score/predictors.py
@@ -7,15 +7,13 @@ from typing import Callable, Type
 import torch
 from torch import Tensor
 
-from sbi.inference.potentials.score_based_potential import (
-    PosteriorScoreBasedPotential,
-)
-
 PREDICTORS = {}
 
 
 def get_predictor(
-    name: str, score_based_potential: PosteriorScoreBasedPotential, **kwargs
+    name: str,
+    score_based_potential: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
+    **kwargs,
 ) -> "Predictor":
     """Helper function to get predictor by name.
 
@@ -54,7 +52,7 @@ class Predictor(ABC):
 
     def __init__(
         self,
-        potential_fn: PosteriorScoreBasedPotential,
+        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
     ):
         """Initialize predictor.
 
@@ -94,7 +92,7 @@ class Predictor(ABC):
 class EulerMaruyama(Predictor):
     def __init__(
         self,
-        potential_fn: PosteriorScoreBasedPotential,
+        potential_fn: 'PosteriorScoreBasedPotential',  # noqa: F821 # type: ignore
         eta: float = 1.0,
     ):
         """Simple Euler-Maruyama discretization of the associated family of reverse
