### 2025-01-30 39b1fbb
add test for score batched sampling
