+++ sbi/samplers/score/diffuser.py
@@ -92,20 +92,18 @@ class Diffuser:
         Returns:
             Tensor: _description_
         """
-        # TODO: for iid setting, self.batch_shape.numel() will be the iid-batch. But we
-        # don't want to generate num_obs samples, but only one sample given the the iid
-        # batch.
-        # TODO: the solution will probably be to distinguish between the iid setting and
-        # batched sampling setting with a flag.
         # TODO: this fixes the iid setting shape problems, but iid inference via
         # iid_bridge is not accurate.
-        num_batch = self.batch_shape.numel()
+        num_batch = (
+            1 if self.predictor.potential_fn.x_is_iid else self.batch_shape.numel()
+        )
         init_shape = (num_samples, num_batch) + self.input_shape
         # init_shape = (
         #     num_samples,
         # ) + self.input_shape  # just use num_samples, not num_batch
         # NOTE: for the IID setting we might need to scale the noise with iid batch
         # size, as in equation (7) in the paper.
+        # NOTE: TODO we need some interface for that
         eps = torch.randn(init_shape, device=self.device)
         mean, std, eps = torch.broadcast_tensors(self.init_mean, self.init_std, eps)
         return mean + std * eps
