+++ sbi/inference/posteriors/score_posterior.py
@@ -139,7 +139,8 @@ class ScorePosterior(NeuralPosterior):
 
         x = self._x_else_default_x(x)
         x = reshape_to_batch_event(x, self.score_estimator.condition_shape)
-        self.potential_fn.set_x(x)
+        is_iid = x.ndim > 1 and x.shape[0] > 1
+        self.potential_fn.set_x(x, x_is_iid=is_iid)
 
         num_samples = torch.Size(sample_shape).numel()
 
