### 2025-01-30 9152d28
adding untesed GAUSS
