+++ sbi/inference/potentials/score_fn_iid.py
@@ -7,6 +7,13 @@ from torch.distributions import Distribution
 
 from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
 from sbi.utils.torchutils import ensure_theta_batched
+from sbi.inference.potentials.score_utils import (
+    add_diag_or_dense,
+    denoise,
+    marginalize,
+    mv_diag_or_dense,
+    solve_diag_or_dense,
+)
 
 
 class ScoreFnIID:
@@ -130,5 +137,170 @@ class FNPEScoreFn(ScoreFnIID):
 
         # Accumulate
         score = (1 - N) * prior_score + base_score.sum(-2, keepdim=True)
+
+        return score
+
+
+class AbstractGaussCorrectedScoreFn(ScoreFnIID):
+    def __init__(
+        self,
+        score_estimator: ConditionalScoreEstimator,
+        prior: Distribution,
+    ) -> None:
+        r"""Initializes the AbstractGaussCorrectedScoreFn class.
+
+        Args:
+            score_estimator: The neural network modelling the score.
+            prior: The prior distribution.
+        """
+        super().__init__(score_estimator, prior)
+
+    @abstractmethod
+    def posterior_precision_est_fn(self, x_o: Tensor) -> Tensor:
+        r"""Abstract method to estimate the posterior precision.
+
+        Args:
+            x_o: Observed data.
+
+        Returns:
+            Estimated posterior precision.
+        """
+        pass
+
+    def denoising_prior(self, m: Tensor, std: Tensor, theta: Tensor) -> Distribution:
+        r"""Denoise the prior distribution.
+
+        Args:
+            m: Mean tensor.
+            std: Standard deviation tensor.
+            theta: Parameters tensor.
+
+        Returns:
+            Denoised prior distribution.
+        """
+        return denoise(self.prior, m, std, theta)
+
+    def marginal_prior(self, a: Tensor, theta: Tensor) -> Distribution:
+        r"""Compute the marginal prior distribution.
+
+        Args:
+            a: Auxiliary variable tensor.
+            theta: Parameters tensor.
+
+        Returns:
+            Marginal prior distribution.
+        """
+        m = self.score_estimator.mean_t_fn(a)
+        std = self.score_estimator.std_t_fn(a)
+        return marginalize(self.prior, m, std)
+
+    def marginal_posterior_precision_est_fn(
+        self, a: Tensor, theta: Tensor, x_o: Tensor, N: int
+    ) -> Tensor:
+        r"""Estimates the marginal posterior precision.
+
+        Args:
+            a: Auxiliary variable tensor.
+            theta: Parameters tensor.
+            x_o: Observed data.
+            N: Number of samples.
+
+        Returns:
+            Estimated marginal posterior precision.
+        """
+        precisions_posteriors = self.posterior_precision_est_fn(x_o)
+        precisions_posteriors = torch.atleast_2d(precisions_posteriors)
+
+        # If one constant precision is given, tile it
+        if precisions_posteriors.shape[0] < N:
+            precisions_posteriors = precisions_posteriors.repeat(N, 1)
+
+        # Denoising posterior via Bayes rule
+        m = self.score_estimator.mean_t_fn(a)
+        std = self.score_estimator.std_t_fn(a)
+
+        if precisions_posteriors.ndim == 3:
+            I = torch.eye(precisions_posteriors.shape[-1])
+        else:
+            I = torch.ones_like(precisions_posteriors)
+
+        marginal_precisions = m**2 / std**2 * I + precisions_posteriors
+        return marginal_precisions
+
+    def marginal_prior_score_fn(self, a: Tensor, theta: Tensor) -> Tensor:
+        r"""Computes the score of the marginal prior distribution.
+
[…]
