### 2025-02-05 f18a368
doc: fix input arg dostrings
