### 2025-02-06 b70a198
wip: fixes from PR implemented
