### 2025-02-06 #7754 @editorialbot
[REVIEW]: sbi reloaded: a toolkit for simulation-based inference workflows
**Submitting author:** @janfb (<a href="http://orcid.org/0000-0003-4979-7092">Jan Boelts</a>)
**Repository:** https://github.com/sbi-dev/sbi
**Branch with paper.md** (empty if default branch): joss-submission-2024
**Version:** v0.24.0
