### 2025-02-08 dedb783
wip: pyro mcmc via sbi
