+++ sbi/utils/hierarchical_inference_utils.py
@@ -1,0 +1,243 @@
+from typing import Optional
+
+import numpy as np
+import torch
+import torch.distributions as dist
+from torch import Tensor
+
+from sbi.neural_nets.estimators import ConditionalDensityEstimator
+
+
+def torch_to_numpy(tensor):
+    return tensor.detach().cpu().numpy()
+
+
+def numpy_to_torch(array):
+    return torch.tensor(np.array(array))
+
+
+class TruncatedNormal(dist.TransformedDistribution):
+    def __init__(self, loc, scale, low, high):
+        """
+        Implements a truncated normal using SigmoidTransform.
+
+        Args:
+            loc (Tensor): Mean of the base normal distribution.
+            scale (Tensor): Standard deviation of the base normal.
+            low (Tensor): Lower bound for truncation.
+            high (Tensor): Upper bound for truncation.
+        """
+        # Ensure low and high match loc in shape
+        low, high = torch.broadcast_tensors(low, high)
+        low, high = low.expand_as(loc), high.expand_as(loc)
+
+        base_dist = dist.Normal(loc, scale)
+        sigmoid = dist.transforms.SigmoidTransform()
+        scale_transform = dist.AffineTransform(loc=low, scale=(high - low))
+
+        super().__init__(base_dist, [sigmoid, scale_transform])
+        self.low = low
+        self.high = high
+
+    def sample(self, sample_shape=torch.Size()):
+        """Samples from the truncated normal distribution."""
+        return torch.as_tensor(super().sample(sample_shape))
+
+    def log_prob(self, value: torch.Tensor) -> torch.Tensor:
+        """Computes log probability correctly accounting for truncation."""
+        # Ensure value, low, and high have the same shape
+        value, low, high = torch.broadcast_tensors(value, self.low, self.high)
+
+        log_base = self.base_dist.log_prob(value)
+        log_cdf_low = self.base_dist.cdf(low).log()
+        log_cdf_high = (1 - self.base_dist.cdf(high)).log()
+
+        return log_base - torch.log1p(-torch.exp(log_cdf_low + log_cdf_high))
+
+
+class HierarchicalPrior(dist.Distribution):
+    def __init__(
+        self,
+        mu_prior,
+        sigma_prior,
+        lower_bounds,
+        upper_bounds,
+        num_subjects,
+        validate_args=None,
+    ):
+        """
+        Hierarchical prior as a PyTorch distribution.
+
+        Args:
+            mu_prior (torch.distributions.Distribution): Prior for group means, shape (D,).
+            sigma_prior (torch.distributions.Distribution): Prior for group stds, shape (D,).
+            lower_bounds (torch.Tensor): Lower bounds for subject-level parameters, shape (D,).
+            upper_bounds (torch.Tensor): Upper bounds for subject-level parameters, shape (D,).
+            num_subjects (int): Number of subjects (M).
+        """
+        self.mu_prior = mu_prior
+        self.sigma_prior = sigma_prior
+        self.lower_bounds = lower_bounds
+        self.upper_bounds = upper_bounds
+        self.num_subjects = num_subjects
+        self.num_parameters = lower_bounds.shape[0]
+
+        super().__init__(event_shape=self.event_shape, validate_args=validate_args)
+
+    @property
+    def event_shape(self):
+        """Return the event shape as a property."""
+        return torch.Size([(self.num_subjects + 2) * self.num_parameters])  # (M+2) * D
+
+    def sample(self, sample_shape=torch.Size()) -> Tensor:
+        """
+        Samples from the hierarchical prior.
+
[…]
