+++ tests/hierarchical_inference_test.py
@@ -1,0 +1,357 @@
+import jax.numpy as jnp
+import numpy as np
+import pyro
+import pyro.distributions as dist
+import pytest
+import torch
+from jax import random
+from pyro.infer import MCMC, NUTS
+from torch.distributions import HalfNormal, Normal
+
+from sbi.inference import NLE
+from sbi.simulators.linear_gaussian import true_posterior_linear_gaussian_mvn_prior
+from sbi.utils.hierarchical_inference_utils import (
+    HierarchicalPrior,
+    TruncatedNormal,
+)
+from tests.test_utils import check_c2st
+
+# Ground truth data and posterior
+num_subjects, num_dim = 10, 3  # Subjects, trials, dimensions
+mu_0 = torch.zeros(num_dim)
+sigma_0 = torch.ones(num_dim) * 1.0
+alpha_0 = torch.ones(num_dim) * 2.0
+beta_0 = torch.ones(num_dim)
+sigma_x = 0.5
+num_samples = 1000
+warmup_steps = 1000
+
+
+# fixture for observed data
+@pytest.fixture(scope="module")
+def x_o_and_params():
+    theta, mu, sigma2, x_o = hierarchical_gaussian_model()
+    return x_o, theta, mu, sigma2
+
+
+# fixture for reference posterior, using fixture above.
+@pytest.fixture(scope="module")
+def reference_posterior(x_o_and_params):
+    x_o, *_ = x_o_and_params
+
+    # Run HMC inference
+    nuts_kernel = NUTS(hierarchical_gaussian_model)
+    mcmc = MCMC(nuts_kernel, num_samples=num_samples, num_warmup=warmup_steps)
+    mcmc.run(x_o=x_o)
+
+    return mcmc
+
+
+@pytest.mark.parametrize("sample_shape", [(1000,), (1000, 10)])
+@pytest.mark.parametrize("num_variables", [1, 2])
+def test_truncated_normal(sample_shape, num_variables):
+    loc = torch.tensor([0.0] * num_variables)
+    scale = torch.tensor([1.0] * num_variables)
+    low = torch.tensor([-1.0] * num_variables)
+    high = torch.tensor([1.0] * num_variables)
+    dist = TruncatedNormal(loc, scale, low, high)
+
+    samples = dist.sample(sample_shape)
+    assert samples.shape == (sample_shape + (num_variables,)), "Sample shape incorrect."
+    # test low and high
+    assert (samples >= low).all(), "Sample below lower bound."
+    assert (samples <= high).all(), "Sample above upper bound."
+
+    log_prob = dist.log_prob(samples)
+    # log_prob outputs batch shape
+    assert log_prob.shape == sample_shape + (num_variables,), (
+        "Log probability shape incorrect."
+    )
+
+
+@pytest.mark.parametrize("sample_shape", [(10,), (20, 10)])
+def test_hierarchical_prior(sample_shape):
+    num_subjects = 2
+    lower_bounds = torch.tensor([0.1, 0.1, 0.1])
+    upper_bounds = torch.tensor([3.0, 2.5, 1.0])
+    num_parameters = lower_bounds.shape[0]
+    mu_prior = Normal(torch.zeros(num_parameters), torch.ones(num_parameters) * 2)
+    sigma_prior = HalfNormal(torch.ones(num_parameters) * 0.5)
+
+    hierarchical_prior = HierarchicalPrior(
+        mu_prior, sigma_prior, lower_bounds, upper_bounds, num_subjects=num_subjects
+    )
+
+    samples = hierarchical_prior.sample(sample_shape)
+    assert samples.shape == sample_shape + (num_subjects + 2, num_parameters), (
+        "Sample shape incorrect."
+    )
+
+    log_prob = hierarchical_prior.log_prob(samples)
[…]
