### 2025-02-11 #7754 @francois-rozet
@boisgera I have read the paper and filled the checklist. The improvements are significant and welcome with respect to the first release. I only have two concerns:

1. I could not find in the documentation how to use normalizing flows implemented in Zuko (Rozet, 2023).

2. In the "Related software" section, the reference to LAMPE should cite Rozet, Delaunoy and Miller (2021).

>  The Probabilists (2024) package offers neural posterior and neural ratio estimation, primarily targeting SBI researchers with a low-level API and full flexibility over the training loop (Lampe stopped being maintained in July 2024).

I would suggest the following sentence:

> The `lampe` (Rozet, 2021) package offers neural posterior and neural ratio estimation, primarily targeting SBI researchers with a low-level API and full flexibility over the training loop. Its development has stopped in favor of the `sbi` project in July 2024.
