### 2025-02-11 #7754 @arnauqb
I'm happy with how the repo is structured and I think all the tests and documentation are of very high standard. I only have a few suggestions:

1. I would add a short paragraph discussing how `sbi` fits into the probabilistic programming space. Many people may be aware of Pyro, Numpyro, Turing.jl, etc. and may wonder what is the difference between `sbi` and these packages and how it complements them. This would also help readers to better understand if `sbi` is the right tool to use for their needs.

2. Shameless plug (I am one of the authors), but I will suggest it since the joss editorial bot also highlighted it. Consider adding https://joss.theoj.org/papers/10.21105/joss.05776 to the related work. Obviously only if you see it fit!

3. I tried running the tests locally and even with `pytest -n auto -m "not slow and not gpu"` the tests took a considerable amount of time. I imagine this may not be great when one wants to use the unittests as a quick way to check things didn't break during development.
