### 2025-02-12 #7754 @boisgera
@arnauqb, @francois-rozet

Thanks a lot for these reviews, many great points there!

As @arnauqb mentions, several DOIs are missing from the bibliography. The bibliography should be updated to include DOIs whenever they exist (this is will be checked in post-review by myself and the editor in chief in any case).

Let's allow @janfb to consider this feedback and get back to us.
