### 2025-02-12 44cd90c
update lampe citation and text.
