+++ paper/paper.bib
@@ -1019,14 +1019,11 @@ year={2022},
   note         = {Version 0.2.0},
 }
 
-@misc{lampe,
-  author       = {Probabilists},
-  title        = {LAMPE: Likelihood-free AMortized Posterior Estimation with PyTorch},
-  year         = {2024},
-  publisher    = {GitHub},
-  journal      = {GitHub repository},
-  howpublished = {\url{https://github.com/probabilists/lampe}},
-  note         = {Version 0.9.0},
+@article{rozet_2021_lampe,
+  title={LAMPE: Likelihoodfree Amortized Posterior Estimation},
+  author={Rozet, Fran{\c{c}}ois and Delaunoy, Arnaud and Miller, Benjamin and others},
+  journal={Statistical Software},
+  year={2021}
 }
 
 @article{bayesflow_2023_software,
