### 2025-02-14 e9d9af4
docstrings and cleanup
