+++ sbi/inference/potentials/score_based_potential.py
@@ -1,6 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
-
+from functools import partial
 from typing import Optional, Tuple
 
 import torch
@@ -80,21 +80,20 @@ class PosteriorScoreBasedPotential(BasePotential):
         x_is_iid: Optional[bool] = False,
         rebuild_flow: Optional[bool] = True,
     ):
+        """
+        Set the observed data and whether it is IID.
+        Args:
+        x_o: The observed data.
+        x_is_iid: Whether the observed data is IID (if batch_dim>1).
+        rebuild_flow: Whether to save (overwrrite) a low-tolerance flow model, useful if
+        the flow needs to be evaluated many times (e.g. for MAP calculation).
+        """
         super().set_x(x_o, x_is_iid)
         if rebuild_flow and self._x_o is not None:
-            x_density_estimator = reshape_to_batch_event(
-                self.x_o, event_shape=self.score_estimator.condition_shape
-            )
-            assert x_density_estimator.shape[0] == 1 or not self.x_is_iid, (
-                "PosteriorScoreBasedPotential does not support IID observations`."
-            )
-            # For large number of evals, we want a high-tolerance flow.
+            # By default, we want a high-tolerance flow.
             # This flow will be used mainly for MAP calculations, hence we want to save
             # it instead of rebuilding it every time.
-            flow = self.get_continuous_normalizing_flow(
-                condition=x_density_estimator, atol=1e-2, rtol=1e-3, exact=True
-            )
-            self.flow = flow
+            self.flow = self.rebuild_flow(atol=1e-2, rtol=1e-3, exact=True)
 
     def __call__(
         self,
@@ -123,17 +122,10 @@ class PosteriorScoreBasedPotential(BasePotential):
             theta, theta.shape[1:], leading_is_sample=True
         )
         self.score_estimator.eval()
+        # use rebuild_flow to evaluate log_prob with better precision, without
+        # overwriting self.flow
         if rebuild_flow or self.flow is None:
-            x_density_estimator = reshape_to_batch_event(
-                self.x_o, event_shape=self.score_estimator.condition_shape
-            )
-            assert x_density_estimator.shape[0] == 1, (
-                "PosteriorScoreBasedPotential supports only x batchsize of 1`."
-            )
-
-            flow = self.get_continuous_normalizing_flow(
-                condition=x_density_estimator, atol=atol, rtol=rtol, exact=exact
-            )
+            flow = self.rebuild_flow(atol=atol, rtol=rtol, exact=exact)
         else:
             flow = self.flow
 
@@ -209,6 +201,30 @@ class PosteriorScoreBasedPotential(BasePotential):
         # Use zuko to build the normalizing flow.
         return NormalizingFlow(transform, base=base_density)
 
+    def rebuild_flow(
+        self, atol: float = 1e-5, rtol: float = 1e-6, exact: bool = True
+    ) -> NormalizingFlow:
+        """
+        Rebuilds the continuous normalizing flow. This is used when
+        a new default x is set, or to evaluate the log probs at higher precision.
+        """
+        if self._x_o is None:
+            raise ValueError(
+                "No observed data x_o is available. Please reinitialize \
+                the potential or manually set self._x_o."
+            )
+        x_density_estimator = reshape_to_batch_event(
+            self.x_o, event_shape=self.score_estimator.condition_shape
+        )
+        assert x_density_estimator.shape[0] == 1, (
+            "PosteriorScoreBasedPotential supports only x batchsize of 1`."
+        )
+
+        flow = self.get_continuous_normalizing_flow(
[…]
