### 2025-02-14 d6e4133
docstrings and cleanup
