### 2025-02-14 5195417
messy version of simple gauss, API to sample method
