### 2025-02-14 5bcf427
jac method (still needs feasible Lambda projection to work)
