+++ sbi/inference/potentials/score_based_potential.py
@@ -10,7 +10,7 @@ from zuko.distributions import NormalizingFlow
 from zuko.transforms import FreeFormJacobianTransform
 
 from sbi.inference.potentials.base_potential import BasePotential
-from sbi.inference.potentials.score_fn_iid import FNPEScoreFn, GaussCorrectedScoreFn
+from sbi.inference.potentials.score_fn_iid import get_iid_method
 from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
 from sbi.neural_nets.estimators.shape_handling import (
     reshape_to_batch_event,
@@ -79,7 +79,7 @@ class PosteriorScoreBasedPotential(BasePotential):
         self,
         x_o: Optional[Tensor],
         x_is_iid: Optional[bool] = False,
-        iid_method: Optional[str] = "fnpe",
+        iid_method: str = "fnpe",
         rebuild_flow: Optional[bool] = True,
     ):
         super().set_x(x_o, x_is_iid)
@@ -180,21 +180,10 @@ class PosteriorScoreBasedPotential(BasePotential):
                 )
             else:
                 assert self.prior is not None, "Prior is required for iid methods."
-                # NOTE: Add here different methods for accumulating the score.
-                # TODO: Warn for FNPE -> Kinda needs a "corrector"
-                if self.iid_method == "fnpe":
-                    score_fn_iid = FNPEScoreFn(
-                        self.score_estimator, self.prior, device=self.device
-                    )
-                elif self.iid_method == "gauss":
-                    score_fn_iid = GaussCorrectedScoreFn(
-                        self.score_estimator, self.prior, 2 * torch.ones_like(theta[-1])
-                    )
-                else:
-                    raise NotImplementedError(
-                        f"Method {self.iid_method} for iid score accumulation not \
-                        implemented."
-                    )
+
+                method_iid = get_iid_method(self.iid_method)
+                score_fn_iid = method_iid(self.score_estimator, self.prior)
+
                 score = score_fn_iid(theta, self.x_o, time)
 
         return score
