+++ sbi/inference/potentials/score_fn_iid.py
@@ -1,5 +1,5 @@
 from abc import abstractmethod
-from typing import Callable, Optional
+from typing import Callable, Optional, Type
 
 import torch
 from torch import Tensor
@@ -16,13 +16,32 @@ from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
 from sbi.utils.torchutils import ensure_theta_batched
 
 
+IID_METHODS = {}
+
+
+def get_iid_method(name: str) -> Type["ScoreFnIID"]:
+    if name not in IID_METHODS:
+        raise NotImplementedError(
+            f"Method {name} for iid score accumulation not implemented."
+        )
+    return IID_METHODS[name]
+
+
+def register_iid_method(name: str) -> Callable:
+    def decorator(cls):
+        IID_METHODS[name] = cls
+        return cls
+
+    return decorator
+
+
 class ScoreFnIID:
     def __init__(
         self,
-        score_estimator: ConditionalScoreEstimator,
+        score_estimator: "ConditionalScoreEstimator",
         prior: Distribution,
         device: str = "cpu",
-    ):
+    ) -> None:
         r"""
         Initializes the ScoreFnIID class.
 
@@ -72,15 +91,15 @@ class ScoreFnIID:
             )[0]
         return prior_score
 
-
+@register_iid_method("fnpe")
 class FNPEScoreFn(ScoreFnIID):
     def __init__(
         self,
-        score_estimator: ConditionalScoreEstimator,
+        score_estimator: "ConditionalScoreEstimator",
         prior: Distribution,
         device: str = "cpu",
         prior_score_weight: Optional[Callable[[Tensor], Tensor]] = None,
-    ):
+    ) -> None:
         r"""
         Initializes the FNPEScoreFn class.
 
@@ -123,7 +142,7 @@ class FNPEScoreFn(ScoreFnIID):
         if time is None:
             time = torch.tensor([self.score_estimator.t_min])
 
-        # NOTE: If this always works needs to be testd
+        # NOTE: If this always works needs to be tested
 
         N = inputs.shape[-2]
 
@@ -144,7 +163,7 @@ class FNPEScoreFn(ScoreFnIID):
 class AbstractGaussCorrectedScoreFn(ScoreFnIID):
     def __init__(
         self,
-        score_estimator: ConditionalScoreEstimator,
+        score_estimator: "ConditionalScoreEstimator",
         prior: Distribution,
     ) -> None:
         r"""Initializes the AbstractGaussCorrectedScoreFn class.
@@ -263,6 +282,7 @@ class AbstractGaussCorrectedScoreFn(ScoreFnIID):
         m = self.score_estimator.mean_t_fn(time)
         std = self.score_estimator.std_fn(time)
         p_denoise = self.denoising_prior(m, std, inputs)
+        # TODO: If multivariate prior this will break
         return 1 / p_denoise.variance
 
     def __call__(
@@ -278,12 +298,14 @@ class AbstractGaussCorrectedScoreFn(ScoreFnIID):
         Returns:
             Corrected score function.
         """
+        # TODO We can assume here a fixed 3dim shape of inputs [b,N,d]
+        # TODO We can assume here a fixed format of conditions [N,...]
         N = conditions.shape[0]
         base_score = self.score_estimator(inputs, conditions, time, **kwargs)
         prior_score = self.marginal_prior_score_fn(time, inputs)
 
         print(f"base_score: {base_score.shape}")
-        print(f"prior_score: {prior_score.shape}")
+        # print(f"prior_score: {prior_score.shape}")
 
         # Marginal prior precision
         prior_precision = self.marginal_denoising_prior_precision_fn(time, inputs)
@@ -297,7 +319,7 @@ class AbstractGaussCorrectedScoreFn(ScoreFnIID):
 
         # Total precision
[…]
