+++ paper/paper.bib
@@ -71,7 +71,8 @@
   volume={9},
   pages={e1516},
   year={2023},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi = {10.7717/peerj-cs.1516}
 }
 
 @misc{delfi-repo,
@@ -168,7 +169,8 @@
   author={Lueckmann, Jan-Matthis and Goncalves, Pedro J and Bassetto, Giacomo and {\"O}cal, Kaan and Nonnenmacher, Marcel and Macke, Jakob H},
   journal={Advances in neural information processing systems},
   volume={30},
-  year={2017}
+  year={2017},
+  doi = {10.48550/arXiv.1711.01861}
 }
 
 @article{wood2010statistical,
@@ -179,7 +181,8 @@
   number={7310},
   pages={1102--1104},
   year={2010},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/nature09319}
 }
 
 @article{beaumont2002approximate,
@@ -190,7 +193,8 @@
   number={4},
   pages={2025--2035},
   year={2002},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/genetics/162.4.2025},
 }
 
 @inproceedings{chen2021neural,
@@ -260,7 +264,8 @@ year={2021},
   pages = {2505},
   author = {Tejero-Cantero, Alvaro and Boelts, Jan and Deistler, Michael and Lueckmann, Jan-Matthis and Durkan, Conor and Goncalves, Pedro J. and Greenberg, David S. and Macke, Jakob H.},
   title = {sbi: A toolkit for simulation-based inference},
-  journal = {Journal of Open Source Software}
+  journal = {Journal of Open Source Software},
+  doi = {10.21105/joss.02505},
 }
 
 @inproceedings{lueckmann2021benchmarking,
@@ -269,7 +274,8 @@ year={2021},
   booktitle={International Conference on Artificial Intelligence and Statistics},
   pages={343--351},
   year={2021},
-  organization={PMLR}
+  organization={PMLR},
+  doi = {10.48550/arXiv.2101.04653}
 }
 
 @inproceedings{lueckmann2019likelihood,
@@ -278,7 +284,8 @@ year={2021},
   booktitle={Symposium on Advances in Approximate Bayesian Inference},
   pages={32--53},
   year={2019},
-  organization={PMLR}
+  organization={PMLR},
+  doi = {10.48550/arXiv.1805.09294}
 }
 
 @article{miller2021truncated,
@@ -286,7 +293,8 @@ year={2021},
   author={Miller, Benjamin and Cole, Alex and Forr{\'e}, Patrick and Louppe, Gilles and Weniger, Christoph},
   journal={Advances in Neural Information Processing Systems},
   volume={34},
-  year={2021}
+  year={2021},
+  doi = {10.48550/arXiv.2107.01214}
 }
 
 @inproceedings{gloeckler2022variational,
@@ -294,6 +302,7 @@ title={Variational methods for simulation-based inference},
 author={Manuel Gloeckler and Michael Deistler and Jakob H. Macke},
 booktitle={International Conference on Learning Representations},
 year={2022},
+doi = {10.48550/arXiv.2203.04176}
 }
 
 @article{gonccalves2020training,
@@ -303,7 +312,8 @@ year={2022},
   volume={9},
   pages={e56261},
   year={2020},
-  publisher={eLife Sciences Publications Limited}
+  publisher={eLife Sciences Publications Limited},
+  doi = {10.7554/eLife.56261}
 }
 
[…]
