+++ paper/paper.bib
@@ -62,6 +62,7 @@
   volume    = {20},
   pages     = {28:1--28:6},
   year      = {2019},
+  doi = {10.48550/arXiv.1810.09538}
 }
 
 @article{abril2023pymc,
@@ -209,7 +210,8 @@ year={2021},
   author={Papamakarios, George and Murray, Iain},
   journal={Advances in neural information processing systems},
   volume={29},
-  year={2016}
+  year={2016},
+  doi = {10.48550/arXiv.1605.06376},
 }
 
 @article{papamakarios2021normalizing,
@@ -219,7 +221,8 @@ year={2021},
   volume={22},
   number={57},
   pages={1--64},
-  year={2021}
+  year={2021},
+  doi = {10.48550/arXiv.1912.02762},
 }
 
 @inproceedings{papamakarios2019sequential,
@@ -228,7 +231,8 @@ year={2021},
   booktitle={The 22nd International Conference on Artificial Intelligence and Statistics},
   pages={837--848},
   year={2019},
-  organization={PMLR}
+  organization={PMLR},
+  doi = {10.48550/arXiv.1805.07226},
 }
 
 @article{papamakarios2017masked,
@@ -253,7 +257,8 @@ year={2021},
   booktitle={International Conference on Machine Learning},
   pages={2404--2414},
   year={2019},
-  organization={PMLR}
+  organization={PMLR},
+  doi = {10.48550/arXiv.1905.07488}
 }
 
 @article{tejerocantero2020sbi,
@@ -404,7 +409,8 @@ doi = {10.48550/arXiv.2203.04176}
   booktitle={International Conference on Machine Learning},
   pages={4239--4248},
   year={2020},
-  organization={PMLR}
+  organization={PMLR},
+  doi = {10.48550/arXiv.1903.04057},
 }
 
 @article{jiang2017learning,
@@ -450,7 +456,8 @@ doi = {10.48550/arXiv.2203.04176}
 	Author = {Bishop, C M},
 	Journal = {Technical Report. Aston University, Birmingham},
 	Title = {Mixture density networks},
-	Year = {1994}}
+	Year = {1994},
+  }
 
 @article{sisson2007sequential,
 	Author = {Sisson, Scott A and Fan, Yanan and Tanaka, Mark M},
@@ -523,14 +530,15 @@ author={Joeri Hermans and Arnaud Delaunoy and Fran{\c{c}}ois Rozet and Antoine W
 journal={Transactions on Machine Learning Research},
 issn={2835-8856},
 year={2022},
-note={}
+doi = {10.48550/arXiv.2110.06581},
 }
 
 @article{talts2018validating,
   title={Validating Bayesian inference algorithms with simulation-based calibration},
   author={Talts, Sean and Betancourt, Michael and Simpson, Daniel and Vehtari, Aki and Gelman, Andrew},
   journal={arXiv preprint arXiv:1804.06788},
-  year={2018}
+  year={2018},
+  doi = {10.48550/arXiv.1804.06788},
 }
 
 @article{jarvenpaa2019efficient,
@@ -552,7 +560,8 @@ note={}
   volume={11},
   pages={e77220},
   year={2022},
-  publisher={eLife Sciences Publications Limited}
+  publisher={eLife Sciences Publications Limited},
+  doi = {10.7554/eLife.77220}
 }
 
 @article{deistler2022energy,
@@ -609,6 +618,7 @@ author={Michael Deistler and Pedro J. Goncalves and Jakob H. Macke},
 booktitle={Advances in Neural Information Processing Systems},
 editor={Alice H. Oh and Alekh Agarwal and Danielle Belgrave and Kyunghyun Cho},
 year={2022},
[…]
