+++ paper/paper.bib
@@ -1178,3 +1178,50 @@ doi = {10.48550/arXiv.2206.07570}
   publisher={Elsevier},
   doi = {10.1016/j.neunet.2023.03.040}
 }
+
+@article{phan2019composable,
+  title={Composable effects for flexible and accelerated probabilistic programming in NumPyro},
+  author={Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
+  journal={arXiv preprint arXiv:1912.11554},
+  year={2019},
+  doi = {10.48550/arXiv.1912.11554},
+}
+
+@article{gelman2015stan,
+  title={Stan: A probabilistic programming language for Bayesian inference and optimization},
+  author={Gelman, Andrew and Lee, Daniel and Guo, Jiqiang},
+  journal={Journal of Educational and Behavioral Statistics},
+  volume={40},
+  number={5},
+  pages={530--543},
+  year={2015},
+  publisher={Sage Publications Sage CA: Los Angeles, CA},
+  doi = {10.3102/1076998615606113},
+}
+
+@inproceedings{ge2018t,
+  author    = {Hong Ge and
+               Kai Xu and
+               Zoubin Ghahramani},
+  title     = {Turing: a language for flexible probabilistic inference},
+  booktitle = {International Conference on Artificial Intelligence and Statistics,
+               {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
+               Spain},
+  pages     = {1682--1690},
+  year      = {2018},
+  url       = {http://proceedings.mlr.press/v84/ge18b.html},
+  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+}
+
+@article{quera-bofarull2023,
+    doi = {10.21105/joss.05776},
+    url = {https://doi.org/10.21105/joss.05776},
+    year = {2023},
+    publisher = {The Open Journal},
+    volume = {8},
+    number = {89},
+    pages = {5776},
+    author = {Arnau Quera-Bofarull and Joel Dyer and Anisoara Calinescu and J. Doyne Farmer and Michael Wooldridge},
+    title = {BlackBIRDS: Black-Box Inference foR Differentiable Simulators},
+    journal = {Journal of Open Source Software},
+}
