### 2025-02-17 #7754 @janfb
@francois-rozet thanks again for your review!

> 1. I could not find in the documentation how to use normalizing flows implemented in Zuko (Rozet, 2023). This feature is mentioned in the paper.
Good catch! We update the documentation accordingly, see here: https://github.com/sbi-dev/sbi/pull/1387

> 3. In the "Related software" section, the reference to LAMPE should cite Rozet, Delaunoy and Miller (2021).>    > The Probabilists (2024) package offers neural posterior and neural ratio estimation, primarily targeting SBI researchers with a low-level API and full flexibility over the training loop (Lampe stopped being maintained in July 2024).
>
>
>    I would suggest the following sentence:
>    > The `lampe` (Rozet, 2021) package offers neural posterior and neural ratio estimation, primarily targeting SBI researchers with a low-level API and full flexibility over the training loop. Its development has stopped in favor of the `sbi` project in July 2024.

Good point, thanks. We updated the paper and references accordingly.

> In addition, I would suggest not to provide the source code directly in the documentation, but to link to the code repository instead. This is possible with Sphinx, I don't know if it is with mkdocs.

Thanks for your feedback! Just to clarify—are you referring to the tutorial notebooks on the website? We use MkDocs and nbconvert to display the latest version of the notebooks for each release (as selected on the site). Plus, each notebook includes links to the corresponding file in the GitHub repository. Does that address your concern about the source code in the documentation?
