### 2025-02-17 #7754 @janfb
> I'm happy with how the repo is structured and I think all the tests and documentation are of very high standard. I only have a few suggestions:

@arnauqb thanks again for your review!

> 1. I would add a short paragraph discussing how `sbi` fits into the probabilistic programming space. Many people may be aware of Pyro, Numpyro, Turing.jl, etc. and may wonder what is the difference between `sbi` and these packages and how it complements them. This would also help readers to better understand if `sbi` is the right tool to use for their needs.

Very good point! We added a paragraph at the beginning of the related software section:

"Simulation-based inference methods implemented in the `sbi` package require only access
to simulated data, which can also be generated offline in other programming languages or
frameworks. This sets `sbi` apart from toolboxes for traditional Bayesian inference,
such as MCMC-based methods [@abril2023pymc; @bingham2019pyro; @gelman2015stan], which
rely on likelihood evaluations, and from probabilistic programming languages (e.g., Pyro
[@bingham2019pyro], NumPyro [@phan2019composable], Stan [@gelman2015stan], or Turing.jl
[@ge2018t]), which typically require the simulator to be differentiable and implemented
within their respective frameworks [@quera-bofarull2023]."

> 3. Shameless plug (I am one of the authors), but I will suggest it since the joss editorial bot also highlighted it. Consider adding https://joss.theoj.org/papers/10.21105/joss.05776 to the related work. Obviously only if you see it fit!

It makes sense to add this paper as well, thanks! We added it to the new paragraph (see above).

> 5. I tried running the tests locally and even with `pytest -n auto -m "not slow and not gpu"` the tests took a considerable amount of time. I imagine this may not be great when one wants to use the unittests as a quick way to check things didn't break during development.

Another good point, thanks! We are currently discussion how to make the tests faster, e.g., by improving the test suite itself, and by reducing the set of core tests used during CI, (see https://github.com/sbi-dev/sbi/issues/1378).

Does this address all the comments and concerns you had?
