### 2025-02-17 #7754 @francois-rozet
Thank you for addressing my comments. Everything seems good to me. Great work!

> Just to clarify—are you referring to the tutorial notebooks on the website?

I am referring to the source code present in the API section of the website. These code blocks are not very legible without code highlighting, and the code lines are not aligned with the line numbers on the left.

![Image](https://github.com/user-attachments/assets/533855c7-1bcf-41bc-90da-8f6625bbc4eb)

Another issue with these source code blocks is that they don't allow to search the code base (where is this function defined, what does it do, what are the arguments, ...). Linking to the source code in the repo is much more valuable.

For example, in the Zuko library, the code `</>` button

![Image](https://github.com/user-attachments/assets/e9e8c670-5460-411c-86e9-ddbd6766ea9d)

links to https://github.com/probabilists/zuko/blob/469a68e2dd161c2c7e01745549f56baca7c503bb/zuko/flows/autoregressive.py#L213-L308

I don't know if this is possible in mkdocs, but if it is, you should consider it.
