### 2025-02-17 sbi-dev/sbi#1389 @janfb
fix code display in API refs and add link to source code
Our API Refs are cluttered and difficult to navigate. To fix this we should

1. have a look at the documentation of https://github.com/mkdocstrings/mkdocstrings or similar plugins to improve the API reference setup
2. find a way to link directly to the corresponding source code.

See also https://github.com/openjournals/joss-reviews/issues/7754#issuecomment-2662947696 for context.
