### 2025-02-17 #7754 @janfb
Thanks for clarifying! Agreed. I created an [issue](https://github.com/sbi-dev/sbi/issues/1389) and we will work on this during the upcoming hackathon where a major re-work of the documentation is planned.
