### 2025-02-17 436542d
fix: initialize HYPER_PARAMETERS in VI subclasses.
