### 2025-02-17 e09a4b4
mark pymc nuts api as failing.
