+++ sbi/inference/potentials/score_based_potential.py
@@ -1,5 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+
 from functools import partial
 from typing import Optional, Tuple
 
