### 2025-02-18 #7754 @boisgera
> [@boisgera](https://github.com/boisgera) I pushed the edits in the main text and the references (fixed DOIs) to the submission branch (https://github.com/sbi-dev/sbi/tree/joss-submission-2024).

Great, thank you!

> We have question regarding the formatting of the author list: We have several groups of authors who contributed equally at different levels. E.g., Michael Deistler and me are both maintainers and equally contributing first authors. The following four authors on the list are "core contributors" and the next five are "major contributors", setting them apart from regular "contributors". Is there a way to format the groups with superscripts in the JOSS markdown template (similar to how we did it on arvix, https://arxiv.org/pdf/2411.17337)?

Ah, that's a good question. I have never edited a paper with this need before, let me ask my fellow co-editors and get back to you!
