### 2025-02-19 b0b8b41
A bunch of auto marginalize and denois methods
