+++ sbi/inference/potentials/score_fn_iid.py
@@ -1,5 +1,5 @@
-import functools
 from abc import abstractmethod
+import functools
 from typing import Callable, Optional, Type
 
 import torch
@@ -14,7 +14,9 @@ from sbi.inference.potentials.score_utils import (
     solve_diag_or_dense,
 )
 from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
+from sbi.samplers.score.diffuser import Diffuser
 from sbi.utils.torchutils import ensure_theta_batched
+
 
 IID_METHODS = {}
 
@@ -28,7 +30,7 @@ def get_iid_method(name: str) -> Type["ScoreFnIID"]:
 
 
 def register_iid_method(name: str) -> Callable:
-    def decorator(cls: Type["ScoreFnIID"]) -> Type["ScoreFnIID"]:
+    def decorator(cls):
         IID_METHODS[name] = cls
         return cls
 
@@ -50,12 +52,16 @@ class ScoreFnIID:
             prior: The prior distribution.
             device: The device on which to evaluate the potential. Defaults to "cpu".
         """
+
         self.score_estimator = score_estimator.to(device)
         self.prior = prior
         self.score_estimator.eval()
 
     @abstractmethod
-    def __call__(self, theta: Tensor) -> Tensor:
+    def __call__(
+        self,
+        theta: Tensor,
+    ) -> Tensor:
         r"""
         Abstract method to be implemented by subclasses to compute the score function.
 
@@ -87,7 +93,6 @@ class ScoreFnIID:
             )[0]
         return prior_score
 
-
 @register_iid_method("fnpe")
 class FNPEScoreFn(ScoreFnIID):
     def __init__(
@@ -114,7 +119,7 @@ class FNPEScoreFn(ScoreFnIID):
         if prior_score_weight is None:
             t_max = score_estimator.t_max
 
-            def prior_score_weight_fn(t: Tensor) -> Tensor:
+            def prior_score_weight_fn(t):
                 return (t_max - t) / t_max
 
         self.prior_score_weight_fn = prior_score_weight_fn
@@ -139,6 +144,8 @@ class FNPEScoreFn(ScoreFnIID):
         if time is None:
             time = torch.tensor([self.score_estimator.t_min])
 
+        # NOTE: If this always works needs to be tested
+
         N = inputs.shape[-2]
 
         # Compute the per-sample score
@@ -160,27 +167,21 @@ class AbstractGaussCorrectedScoreFn(ScoreFnIID):
         self,
         score_estimator: "ConditionalScoreEstimator",
         prior: Distribution,
-        ensure_valid_lam: bool = True,
-        precission_nugget: float = 0.05,
     ) -> None:
         r"""Initializes the AbstractGaussCorrectedScoreFn class.
 
         Args:
             score_estimator: The neural network modelling the score.
             prior: The prior distribution.
-            ensure_valid_lam: Flag to ensure the precision matrix is positive definite.
-            precission_nugget: Nugget value to ensure positive definiteness.
         """
         super().__init__(score_estimator, prior)
-        self.ensure_valid_lam = ensure_valid_lam
-        self.precission_nugget = precission_nugget
 
     @abstractmethod
     def posterior_precision_est_fn(self, conditions: Tensor) -> Tensor:
         r"""Abstract method to estimate the posterior precision.
 
         Args:
-            conditions: Observed data.
+            x_o: Observed data.
 
         Returns:
             Estimated posterior precision.
@@ -193,7 +194,7 @@ class AbstractGaussCorrectedScoreFn(ScoreFnIID):
         Args:
             m: Mean tensor.
             std: Standard deviation tensor.
-            inputs: Parameters tensor.
+            theta: Parameters tensor.
 
         Returns:
             Denoised prior distribution.
[…]
