+++ sbi/inference/potentials/score_utils.py
@@ -7,8 +7,11 @@ from torch.distributions import (
     Normal,
     MixtureSameFamily,
     Categorical,
+    Uniform,
+    constraints,
 )
 from sbi.utils.sbiutils import ImproperEmpirical
+from sbi.utils.torchutils import BoxUniform
 
 # Automatic denoising -----------------------------------------------------
 
@@ -100,6 +103,11 @@ def denoise_multivariate_gaussian(
     posterior_mean = torch.matmul(posterior_cov, term1 + term2).squeeze(-1)
     return MultivariateNormal(posterior_mean, covariance_matrix=posterior_cov)
 
+
+def denoise_uniform(p: Uniform, m: Tensor, s: Tensor, x_t: Tensor):
+    return UniformNormalPosterior(p.low, p.high, m, s, x_t)
+
+
 def denoise_empirical(
     p: Distribution,
     m: Tensor,
@@ -155,6 +163,8 @@ def marginalize(p: Distribution, m: Tensor, s: Tensor) -> Distribution:
         return marginalize_gaussian(p, m, s)
     elif isinstance(p, MultivariateNormal):
         return marginalize_multivariate_gaussian(p, m, s)
+    elif isinstance(p, (Uniform, BoxUniform)):
+        return marginalize_uniform(p, m, s)
     else:
         return marginalize_empirical(p, m, s)
 
@@ -171,6 +181,23 @@ def marginalize_independent(p: Independent, m: Tensor, s: Tensor) -> Independent
         The marginal independent distribution.
     """
     return Independent(marginalize(p.base_dist, m, s), p.reinterpreted_batch_ndims)
+
+def marginalize_mixture(
+    p: MixtureSameFamily, m: Tensor, s: Tensor
+) -> MixtureSameFamily:
+    """Marginalize a MixtureSameFamily distribution.
+
+    Args:
+        p: The prior MixtureSameFamily distribution.
+        m: The scaling factor m.
+        s: The standard deviation of the noise s.
+
+    Returns:
+        The marginal MixtureSameFamily distribution.
+    """
+    return MixtureSameFamily(
+        p.mixture_distribution, marginalize(p.component_distribution, m, s)
+    )
 
 
 def marginalize_gaussian(p: Normal, m: Tensor, s: Tensor) -> Normal:
@@ -217,10 +244,30 @@ def marginalize_multivariate_gaussian(
 
     marginal_mean = m * mean_0
     marginal_cov = (m**2) * cov_0 + s**2 * torch.eye(
-        mean_0.shape[-1], dtype=cov_0.dtype, device=cov_0.device
+        mean_0.shape[-1],
+        dtype=cov_0.dtype,  # type: ignore
+        device=cov_0.device,  # type: ignore
     )
 
     return MultivariateNormal(marginal_mean, covariance_matrix=marginal_cov)
+
+
+def marginalize_uniform(
+    p: Uniform | BoxUniform, m: Tensor, s: Tensor
+) -> 'UniformNormalConvolution':
+    """Marginalize a uniform distribution.
+
+    Args:
+        p: The prior uniform distribution.
+        m: The scaling factor.
+        s: The standard deviation of the noise.
+
+    Returns:
+        The marginal uniform distribution.
+    """
+
+    return UniformNormalConvolution(p.low, p.high, m, s)  # type: ignore
+
 
 def marginalize_empirical(
     p: Distribution, m: Tensor, s: Tensor, num_particles: int = 5000
@@ -238,6 +285,92 @@ def marginalize_empirical(
     # Less exact but more well bahaved score
     # p = Independent(Normal(marginal_mean.mean(0), marginal_std.mean(0)), 1)
     return p
+
+# Special distributions:
[…]
