### 2025-02-19 0d0991b
general prior with GMM approx
