+++ sbi/utils/score_utils.py
@@ -1,3 +1,4 @@
+from sklearn.mixture import GaussianMixture
 import torch
 from torch import Tensor
 from torch.distributions import (
@@ -12,6 +13,7 @@ from torch.distributions import (
 )
 from sbi.utils.sbiutils import ImproperEmpirical
 from sbi.utils.torchutils import BoxUniform
+from functools import lru_cache
 
 # Automatic denoising -----------------------------------------------------
 
@@ -34,10 +36,14 @@ def denoise(p: Distribution, m: Tensor, s: Tensor, x_t: Tensor) -> Distribution:
     """
     if isinstance(p, Independent):
         return denoise_independent(p, m, s, x_t)
+    elif isinstance(p, MixtureSameFamily):
+        return denoise_mixture(p, m, s, x_t)
     elif isinstance(p, Normal):
         return denoise_gaussian(p, m, s, x_t)
     elif isinstance(p, MultivariateNormal):
         return denoise_multivariate_gaussian(p, m, s, x_t)
+    elif isinstance(p, (Uniform, BoxUniform)):
+        return denoise_uniform(p, m, s, x_t)
     else:
         return denoise_empirical(p, m, s, x_t)
 
@@ -87,11 +93,11 @@ def denoise_multivariate_gaussian(
 ) -> MultivariateNormal:
     mean0 = p.loc
     cov0 = p.covariance_matrix
-    n = cov0.size(-1)
+    n = cov0.size(-1)  # type: ignore
     # Support batch dimensions by expanding the identity matrix to match cov0's
     # batch shape
-    batch_shape = cov0.shape[:-2]
-    id_matrix = torch.eye(n, dtype=cov0.dtype, device=cov0.device)
+    batch_shape = cov0.shape[:-2]  # type: ignore
+    id_matrix = torch.eye(n, dtype=cov0.dtype, device=cov0.device)  # type: ignore
     id_matrix = id_matrix.expand(*batch_shape, n, n)
     precision_prior = torch.linalg.inv(cov0)
     # Reshape m and s so the operation broadcasts correctly over batch dims
@@ -104,7 +110,34 @@ def denoise_multivariate_gaussian(
     return MultivariateNormal(posterior_mean, covariance_matrix=posterior_cov)
 
 
-def denoise_uniform(p: Uniform, m: Tensor, s: Tensor, x_t: Tensor):
+def denoise_mixture(
+    p: MixtureSameFamily, m: Tensor, s: Tensor, x_t: Tensor
+) -> MixtureSameFamily:
+    """Denoise a MixtureSameFamily distribution.
+
+    Args:
+        p: The prior MixtureSameFamily distribution.
+        m: The scaling factor m.
+        s: The standard deviation of the noise.
+        x_t: The observed data.
+
+    Returns:
+        The posterior MixtureSameFamily distribution.
+    """
+    mixture_dist = p.mixture_distribution
+    component_dist = p.component_distribution
+    denoised_components = denoise(component_dist, m, s, x_t)
+    mixture_logits = mixture_dist.logits
+    # Update the logits to reflect the new likelihoods
+    component_loglikelihood = denoised_components.log_prob(x_t)
+    # Update the logits to reflect the new likelihoods
+    denoised_logits = mixture_logits[None, ...] + component_loglikelihood
+    # Normalize the logits
+    denoised_logits = denoised_logits - denoised_logits.logsumexp(dim=-1, keepdims=True)
+    return MixtureSameFamily(Categorical(logits=denoised_logits), denoised_components)
+
+
+def denoise_uniform(p: Uniform | BoxUniform, m: Tensor, s: Tensor, x_t: Tensor):
[…]
