### 2025-02-20 f22467d
Bunch of reffactorings and customizability
