+++ sbi/utils/score_utils.py
@@ -1,19 +1,21 @@
-from sklearn.mixture import GaussianMixture
+from functools import lru_cache
+from typing import Optional
+
 import torch
+from sklearn.mixture import GaussianMixture
 from torch import Tensor
 from torch.distributions import (
+    Categorical,
     Distribution,
     Independent,
+    MixtureSameFamily,
     MultivariateNormal,
     Normal,
-    MixtureSameFamily,
-    Categorical,
     Uniform,
     constraints,
 )
-from sbi.utils.sbiutils import ImproperEmpirical
+
 from sbi.utils.torchutils import BoxUniform
-from functools import lru_cache
 
 # Automatic denoising -----------------------------------------------------
 
@@ -91,6 +93,17 @@ def denoise_gaussian(p: Normal, m: Tensor, s: Tensor, x_t: Tensor) -> Normal:
 def denoise_multivariate_gaussian(
     p: MultivariateNormal, m: Tensor, s: Tensor, x_t: Tensor
 ) -> MultivariateNormal:
+    """Denoise a multivariate Gaussian distribution.
+
+    Args:
+        p: The prior multivariate Gaussian distribution.
+        m: The scaling factor.
+        s: The standard deviation of the noise.
+        x_t: The observed data.
+
+    Returns:
+        The posterior multivariate Gaussian distribution.
+    """
     mean0 = p.loc
     cov0 = p.covariance_matrix
     n = cov0.size(-1)  # type: ignore
@@ -137,7 +150,20 @@ def denoise_mixture(
     return MixtureSameFamily(Categorical(logits=denoised_logits), denoised_components)
 
 
-def denoise_uniform(p: Uniform | BoxUniform, m: Tensor, s: Tensor, x_t: Tensor):
+def denoise_uniform(
+    p: Uniform | BoxUniform, m: Tensor, s: Tensor, x_t: Tensor
+) -> 'UniformNormalPosterior':
+    """Denoise a uniform distribution.
+
+    Args:
+        p: The prior uniform distribution.
+        m: The scaling factor m.
+        s: The standard deviation of the noise s.
+        x_t: The observed data.
+
+    Returns:
+        The posterior uniform distribution.
+    """
     return UniformNormalPosterior(p.low, p.high, m, s, x_t)
 
 
@@ -146,33 +172,21 @@ def denoise_empirical(
     m: Tensor,
     s: Tensor,
     x_t: Tensor,
-    num_particles: int = 5000,
 ) -> MixtureSameFamily:
+    """Denoise an empirical distribution.
+
+    Args:
+        p: The prior empirical distribution.
+        m: The scaling factor m.
+        s: The standard deviation of the noise s.
+        x_t: The observed data.
+        num_particles: Number of particles to sample. Defaults to 5000.
+
+    Returns:
+        The posterior empirical distribution.
+    """
     gmm = fit_gmm(p)
     return denoise_mixture(gmm, m, s, x_t)
-    raise NotImplementedError("Denoising empirical distributions is not supported.")
-    # NOTE: Thats not that nice
-    particles = p.sample((num_particles,))  # Sample from prior
-    s = torch.clip(s, 1e-1, 1e6)
-
-    # Compute posterior mean and standard deviation
-    posterior_mean = (m * particles + s**2 * x_t) / (m**2 + s**2)
-    posterior_std = torch.sqrt((s**2) / (m**2 + s**2))
-    posterior_std = torch.broadcast_to(posterior_std, posterior_mean.shape)
-
-    # Compute unnormalized log-likelihoods as logits
-    # Gaussian likelihood
-    logits = -0.5 * torch.sum(((x_t - m * particles) / s) ** 2, axis=-1)  # type: ignore
[…]

[…]
