### 2025-02-20 8d69e9c
Update tests/bm_test.py
Co-authored-by: Jan <janfb@users.noreply.github.com>
