### 2025-02-20 #7754 @boisgera
> [@janfb](https://github.com/janfb) Wrt to the superscript for authors issue:
>
> I have not received any answers (yet) from my co-editors. Therefore I suspect that it's not something that's been asked before. Let me have a look at the document processor (https://github.com/openjournals/inara) and I'll get back to you.

@janfb

AFAICT there is a bug in inara when you try to add a custom footnote for a corresponding author; see https://github.com/openjournals/inara/issues/105. I advise you to follow this issue. If no solution in a reasonable time frame, then we'll reconsider your options. Is that a plan that works for you?
