+++ sbi/utils/nn_utils.py
@@ -2,6 +2,11 @@
 from typing import Optional
 from warnings import warn
 
+import nflows.nn.nde.made as made
+import numpy as np
+import torch
+import torch.nn.functional as F
+from pyknos.nflows import distributions as distributions_
 from torch import Tensor, nn
 
 
@@ -62,3 +67,127 @@ def check_net_device(
         return net.to(device)
     else:
         return net
+
+
+"""
+Temporary Patches to fix nflows MADE bug. Remove once upstream bug is fixed.
+"""
+
+
+class MADEWrapper(made.MADE):
+    """Implementation of MADE.
+
+    It can use either feedforward blocks or residual blocks (default is residual).
+    Optionally, it can use batch norm or dropout within blocks (default is no).
+    """
+
+    def __init__(
+        self,
+        features,
+        hidden_features,
+        context_features=None,
+        num_blocks=2,
+        output_multiplier=1,
+        use_residual_blocks=True,
+        random_mask=False,
+        activation=F.relu,
+        dropout_probability=0.0,
+        use_batch_norm=False,
+    ):
+        if use_residual_blocks and random_mask:
+            raise ValueError("Residual blocks can't be used with random masks.")
+        super().__init__(
+            features + 1,
+            hidden_features,
+            context_features,
+            num_blocks,
+            output_multiplier,
+            use_residual_blocks,
+            random_mask,
+            activation,
+            dropout_probability,
+            use_batch_norm,
+        )
+
+    def forward(self, inputs, context=None):
+        # add dummy input to ensure all dims conditioned on context.
+        dummy_input = torch.zeros((inputs.shape[:-1] + (1,)))
+        concat_input = torch.cat((dummy_input, inputs), dim=-1)
+        outputs = super().forward(concat_input, context)
+        return outputs[..., self.output_multiplier :]  # remove dummy input
+
+
+"""
+Temporary Patches to fix nflows MADE bug. Remove once upstream bug is fixed.
+"""
+
+
+class MADEMoGWrapper(distributions_.MADEMoG):
+    def __init__(
+        self,
+        features,
+        hidden_features,
+        context_features,
+        num_blocks=2,
+        num_mixture_components=1,
+        use_residual_blocks=True,
+        random_mask=False,
+        activation=F.relu,
+        dropout_probability=0.0,
+        use_batch_norm=False,
+        custom_initialization=False,
+    ):
+        super().__init__(
+            features + 1,
+            hidden_features,
+            context_features,
+            num_blocks,
+            num_mixture_components,
+            use_residual_blocks,
+            random_mask,
+            activation,
+            dropout_probability,
+            use_batch_norm,
+            custom_initialization,
+        )
+
+    def _log_prob(self, inputs, context=None):
+        dummy_input = torch.zeros((inputs.shape[:-1] + (1,)))
+        concat_inputs = torch.cat((dummy_input, inputs), dim=-1)
+
+        outputs = self._made.forward(concat_inputs, context=context)
+        outputs = outputs.reshape(
+            *concat_inputs.shape, self._made.num_mixture_components, 3
+        )
+
+        logits, means, unconstrained_stds = (
+            outputs[..., 0],
+            outputs[..., 1],
+            outputs[..., 2],
+        )
+        # remove first dimension of means, unconstrained_stds
+        logits = logits[..., 1:, :]
+        means = means[..., 1:, :]
+        unconstrained_stds = unconstrained_stds[..., 1:, :]
+
+        log_mixture_coefficients = torch.log_softmax(logits, dim=-1)
+        stds = F.softplus(unconstrained_stds) + self._made.epsilon
+
+        log_prob = torch.sum(
+            torch.logsumexp(
+                log_mixture_coefficients
[…]
