### 2025-02-20 aa05585
fix: incorrect num chains warn text in batched sampling (#1391)
