### 2025-02-20 16436e6
feat: refactoring and new features for NPSE (#1370)
* npse MAP

* set default enable_Transform to True

* sampling via diffusion twice

* batched sampling for score-based posteriors

* add test for score batched sampling

* better convergence checks
