### 2025-02-20 a7392f9
improve PR and issue templates
