+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,23 +1,36 @@
 ---
 name: Bug report
-about: Create a report to help us improve
+about: Report an issue to help us improve SBI
 title: ''
 labels: bug
 assignees: ''
 
 ---
 
-**Describe the bug**
-A clear and concise description of what the bug is.
+## 🐛 Bug Description
 
-**To Reproduce**
-Please add a minimal code example that reproduces the problem:
-1. Name Python version and SBI version
-2. Minimal code example
-3. [Optional]: error message
+A clear and concise description of the bug. What is happening, and how does it differ from expected behavior?
 
-**Expected behavior**
-A clear and concise description of what you expected to happen.
+...
 
-**Additional context**
-Add any other context about the problem here.
+## 🔄 Steps to Reproduce
+
+Please provide a minimal example to help us reproduce the issue:
+
+1. **Python version** and **SBI version** (run `python --version` and `pip show sbi`)
+2. **Minimal code example** that triggers the bug
+3. **[Optional] Error message or logs** (if applicable, paste the full traceback)
+
+...
+
+## ✅ Expected Behavior
+
+What did you expect to happen instead?
+
+...
+
+## 📌 Additional Context
+
+Any other information that might help us debug the issue (e.g., system info, dependencies, related issues).
+
+...
