### 2025-02-20 sbi-dev/sbi#1378 @janfb
another idea would be caching test results across CI runs within a feature branch and then executing only the previously failed tests with every new push:

something like:

```
    - name: Cache pytest results
      uses: actions/cache@v4
      with:
        path: .pytest_cache
        key: ${{ runner.os }}-pytest-${{ github.ref }}
        restore-keys: |
          ${{ runner.os }}-pytest-

    - name: Run only failing tests from last CI run
      run: |
        pytest --lf -v -x -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/ || \
        pytest -v -x -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
```

note the `||` between the `pytest` calls, which will execute all tests only if the any recently failed tests fail again.
