### 2025-02-20 sbi-dev/sbi#1378 @janfb
yet another idea to speed up CI would be to call only the tests of the changed files, e.g., using https://testmon.org/

this could look something like:

- collect tests affected by the current files changes into `.testmondata`
- cache this file across CI runs
- run pytest with `testmon` option

```
- name: Cache testmon data
  uses: actions/cache@v4
  with:
    path: .testmondata
    key: testmon-${{ github.ref }}-${{ github.sha }}
    restore-keys: |
      testmon-${{ github.ref }}-
      testmon-

- name: Run the fast CPU tests with testmon
  run: |
    pytest --testmon -v -x -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
```
