### 2025-02-20 68ee1a7
fix: temporary wrappers to fix MADE (#1398)
* add temporary MADE wrappers

* test MADEMoG
