### 2025-02-20 9a781bd
doc: fix docstring shapes
