### 2025-02-21 d6c57ff
add uv to CD and type checks
