+++ .github/workflows/cd.yml
@@ -25,26 +25,19 @@ jobs:
         fetch-depth: 0
         lfs: false
 
-    - name: Set up Python
-      uses: actions/setup-python@v5
+    - name: Install uv and set the python version
+      uses: astral-sh/setup-uv@v5
       with:
         python-version: '3.9'
+        enable-cache: true
 
-    - name: Cache dependency
-      id: cache-dependencies
-      uses: actions/cache@v4
-      with:
-        path: ~/.cache/pip
-        key: ${{ runner.os }}-pip
-
-    - name: Install dependencies
+    - name: Install dependencies with uv
       run: |
-        python -m pip install --upgrade pip
-        pip install -e .[dev]
+        uv pip install -e .[dev]
 
     - name: Run the fast and the slow CPU tests with coverage
       run: |
-        pytest -v -x -n auto -m "not gpu" --cov=sbi --cov-report=xml tests/
+        uv run pytest -v -x -n auto -m "not gpu" --cov=sbi --cov-report=xml tests/
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4-beta
