### 2025-02-21 bfe6df2
Remove assert that IID data is not supported
