### 2025-02-21 fd0f964
Ruff linting
