+++ .github/workflows/ci.yml
@@ -40,9 +40,18 @@ jobs:
         uv pip install torch --extra-index-url https://download.pytorch.org/whl/cpu
         uv pip install -e .[dev]
 
+    - name: Cache testmon data
+      uses: actions/cache@v4
+      with:
+        path: .testmondata
+        key: testmon-${{ github.ref }}-${{ github.sha }}
+        restore-keys: |
+          testmon-${{ github.ref }}-
+          testmon-
+
     - name: Run the fast CPU tests with coverage
       run: |
-        uv run pytest -v -x -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
+        uv run pytest --testmon -v -x -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
