### 2025-02-21 53173ff
allow testmon and markers
