+++ .github/workflows/ci.yml
@@ -40,18 +40,26 @@ jobs:
         uv pip install torch --extra-index-url https://download.pytorch.org/whl/cpu
         uv pip install -e .[dev]
 
+    - name: Cache uv dependencies
+      uses: actions/cache@v4
+      with:
+        path: ~/.cache/uv
+        key: uv-${{ runner.os }}-${{ matrix.python-version }}-${{ hashFiles('pyproject.toml') }}
+        restore-keys: |
+          uv-${{ runner.os }}-${{ matrix.python-version }}-
+
     - name: Cache testmon data
       uses: actions/cache@v4
       with:
         path: .testmondata
-        key: testmon-${{ github.ref }}-${{ github.sha }}
+        key: testmon-${{ runner.os }}-${{ matrix.python-version }}-${{ github.ref }}
         restore-keys: |
-          testmon-${{ github.ref }}-
-          testmon-
+          testmon-${{ runner.os }}-${{ matrix.python-version }}-
 
     - name: Run the fast CPU tests with coverage
       run: |
-        uv run pytest --testmon-forceselect -v -x -n auto	-m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
+        uv run pytest --testmon-forceselect -v -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
+
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4
