+++ sbi/inference/posteriors/mcmc_posterior.py
@@ -404,6 +404,7 @@ class MCMCPosterior(NeuralPosterior):
             Samples from the posteriors of shape (*sample_shape, B, *input_shape)
         """
 
+        # TODO: adding this comment to test the CI with testmon.
         # Replace arguments that were not passed with their default.
         method = self.method if method is None else method
         thin = self.thin if thin is None else thin
