+++ sbi/inference/posteriors/mcmc_posterior.py
@@ -404,7 +404,9 @@ class MCMCPosterior(NeuralPosterior):
             Samples from the posteriors of shape (*sample_shape, B, *input_shape)
         """
 
-        # TODO: adding this comment to test the CI with testmon.
+        # TODO: testing testmon by changing this file.
+        print("sample_shape", sample_shape)
+        print("x", x)
         # Replace arguments that were not passed with their default.
         method = self.method if method is None else method
         thin = self.thin if thin is None else thin
