+++ .github/workflows/ci.yml
@@ -39,12 +39,13 @@ jobs:
       run: |
         uv pip install torch --extra-index-url https://download.pytorch.org/whl/cpu
         uv pip install -e .[dev]
+        uv sync --generate-lockfile  # <-- Generate uv.lock
 
     - name: Cache uv dependencies
       uses: actions/cache@v4
       with:
         path: ~/.cache/uv
-        key: uv-${{ runner.os }}-${{ matrix.python-version }}-${{ hashFiles('pyproject.toml') }}
+        key: uv-${{ runner.os }}-${{ matrix.python-version }}-${{ hashFiles('uv.lock') }}
         restore-keys: |
           uv-${{ runner.os }}-${{ matrix.python-version }}-
 
