### 2025-02-21 fc6ca5d
fix uv setup
