### 2025-02-21 b6991cf
fix uv setup
