### 2025-02-21 a4d7bd9
fix uv setup
