### 2025-02-21 828385e
fix uv setup
