### 2025-02-21 64784b7
remove pip from ci
