+++ tests/mnle_test.py
@@ -41,7 +41,7 @@ def mixed_simulator(theta: Tensor, stimulus_condition: Union[Tensor, float] = 2.
 
 
 def mixed_simulator_with_conditions(
-    theta_and_condition: Tensor, last_idx_parameters: int = 2
+    theta_and_condition: Tensor, last_idx_parameters: int = 3
 ) -> Tensor:
     """Simulator for mixed data with experimental conditions."""
     # simulate with experiment conditions
@@ -155,6 +155,9 @@ def test_mnle_accuracy_with_different_samplers_and_trials(
         [
             Gamma(torch.tensor([1.0]), torch.tensor([0.5])),
             Beta(torch.tensor([2.0]), torch.tensor([2.0])),
+            Beta(
+                torch.tensor([2.0]), torch.tensor([2.0])
+            ),  # tests discrete dims > 1 works
         ],
         validate_args=False,
     )
@@ -238,11 +241,11 @@ class BinomialGammaPotential(BasePotential):
         batch_size = theta.shape[0]
         num_trials = self.x_o.shape[0]
         theta = theta.reshape(batch_size, 1, -1)
-        beta, rho = theta[:, :, :1], theta[:, :, 1:]
+        beta, rhos = theta[:, :, :1], theta[:, :, 1:]
 
         # vectorized
-        logprob_choices = Binomial(probs=rho).log_prob(
-            self.x_o[:, 1:].reshape(1, num_trials, -1)
+        logprob_choices = torch.stack(
+            [Binomial(probs=rho).log_prob(self.x_o[:, 1:]) for rho in rhos],
         )
 
         logprob_rts = InverseGamma(
@@ -250,7 +253,7 @@ class BinomialGammaPotential(BasePotential):
             rate=beta,
         ).log_prob(self.x_o[:, :1].reshape(1, num_trials, -1))
 
-        joint_likelihood = (logprob_choices + logprob_rts).squeeze()
+        joint_likelihood = torch.sum(logprob_choices, dim=-1) + logprob_rts.squeeze()
 
         assert joint_likelihood.shape == torch.Size([theta.shape[0], self.x_o.shape[0]])
         return joint_likelihood.sum(1)
@@ -273,6 +276,9 @@ def test_mnle_with_experimental_conditions(mcmc_params_accurate: dict):
         [
             Gamma(torch.tensor([1.0]), torch.tensor([0.5])),
             Beta(torch.tensor([2.0]), torch.tensor([2.0])),
+            Beta(
+                torch.tensor([2.0]), torch.tensor([2.0])
+            ),  # tests discrete dims > 1 works
             BoxUniform(torch.tensor([0.0]), torch.tensor([1.0])),
         ],
         validate_args=False,
@@ -280,14 +286,14 @@ def test_mnle_with_experimental_conditions(mcmc_params_accurate: dict):
 
     theta = proposal.sample((num_simulations,))
     x = mixed_simulator_with_conditions(theta)
-    assert x.shape == (num_simulations, 2)
+    assert x.shape == (num_simulations, 3)
 
     num_trials = 10
     theta_and_condition = proposal.sample((num_trials,))
     # use only a single parameter (iid trials)
-    theta_o = theta_and_condition[:1, :2].repeat(num_trials, 1)
+    theta_o = theta_and_condition[:1, :3].repeat(num_trials, 1)
     # but different conditions
-    condition_o = theta_and_condition[:, 2:]
+    condition_o = theta_and_condition[:, 3:]
     theta_and_conditions_o = torch.cat((theta_o, condition_o), dim=1)
 
     x_o = mixed_simulator_with_conditions(theta_and_conditions_o)
[…]
