### 2025-02-21 9f04c91
restore old testmon setup
