### 2025-02-21 5096a1a
rebuild flow for each new x, but not in __call__. fix map test
