+++ sbi/inference/potentials/score_based_potential.py
@@ -1,7 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
-from functools import partial
 from typing import Optional, Tuple
 
 import torch
@@ -79,42 +78,36 @@ class PosteriorScoreBasedPotential(BasePotential):
         self,
         x_o: Optional[Tensor],
         x_is_iid: Optional[bool] = False,
-        rebuild_flow: Optional[bool] = False,
+        atol: float = 1e-5,
+        rtol: float = 1e-6,
+        exact: bool = True,
     ):
         """
         Set the observed data and whether it is IID.
+
+        Rebuids the continuous normalizing flow if the observed data is set.
+
         Args:
-        x_o: The observed data.
-        x_is_iid: Whether the observed data is IID (if batch_dim>1).
-        rebuild_flow: Whether to rebuild (overwrite) a higher-tolerance (faster) flow
-            model. Useful if the flow needs to be evaluated many times (e.g. for
-            MAP calculation).
+            x_o: The observed data.
+            x_is_iid: Whether the observed data is IID (if batch_dim>1).
+            atol: Absolute tolerance for the ODE solver.
+            rtol: Relative tolerance for the ODE solver.
+            exact: Whether to use the exact ODE solver.
         """
         super().set_x(x_o, x_is_iid)
-        if rebuild_flow and self._x_o is not None:
-            # By default, we want a high-tolerance flow (see below).
-            # Here, we create a faster high-tolerance flow, eg., for MAP calculations
-            # that will be built only once.
-            self.flow = self.rebuild_flow(atol=1e-2, rtol=1e-3, exact=True)
+        if self._x_o is not None:
+            self.flow = self.rebuild_flow(atol=atol, rtol=rtol, exact=exact)
 
     def __call__(
         self,
         theta: Tensor,
         track_gradients: bool = True,
-        rebuild_flow: bool = True,
-        atol: float = 1e-5,
-        rtol: float = 1e-6,
-        exact: bool = True,
     ) -> Tensor:
         """Return the potential (posterior log prob) via probability flow ODE.
 
         Args:
             theta: The parameters at which to evaluate the potential.
             track_gradients: Whether to track gradients.
-            rebuild_flow: Whether to rebuild the CNF for accurate log_prob evaluation.
-            atol: Absolute tolerance for the ODE solver.
-            rtol: Relative tolerance for the ODE solver.
-            exact: Whether to use the exact ODE solver.
 
         Returns:
             The potential function, i.e., the log probability of the posterior.
@@ -124,15 +117,9 @@ class PosteriorScoreBasedPotential(BasePotential):
             theta, theta.shape[1:], leading_is_sample=True
         )
         self.score_estimator.eval()
-        # use rebuild_flow to evaluate log_prob with better precision, without
-        # overwriting self.flow
-        if rebuild_flow or self.flow is None:
-            flow = self.rebuild_flow(atol=atol, rtol=rtol, exact=exact)
-        else:
-            flow = self.flow
 
         with torch.set_grad_enabled(track_gradients):
-            log_probs = flow.log_prob(theta_density_estimator).squeeze(-1)
+            log_probs = self.flow.log_prob(theta_density_estimator).squeeze(-1)
             # Force probability to be zero outside prior support.
             in_prior_support = within_support(self.prior, theta)
 
@@ -313,9 +300,8 @@ class CallableDifferentiablePotentialFunction:
         self.posterior_score_based_potential = posterior_score_based_potential
 
     def __call__(self, input):
-        prepared_potential = partial(
-            self.posterior_score_based_potential.__call__, rebuild_flow=False
-        )
[…]
