### 2025-02-21 67bb7fc
rebuild flow for each new x, but not in __call__. fix map test
