+++ tutorials/example_01_utils.py
@@ -42,11 +42,11 @@ class BinomialGammaPotential(BasePotential):
         batch_size = theta.shape[0]
         num_trials = self.x_o.shape[0]
         theta = theta.reshape(batch_size, 1, -1)
-        beta, rho = theta[:, :, :1], theta[:, :, 1:]
+        beta, rhos = theta[:, :, :1], theta[:, :, 1:]
 
         # vectorized
-        logprob_choices = Binomial(probs=rho).log_prob(
-            self.x_o[:, 1:].reshape(1, num_trials, -1)
+        logprob_choices = torch.stack(
+            [Binomial(probs=rho).log_prob(self.x_o[:, 1:]) for rho in rhos],
         )
 
         logprob_rts = InverseGamma(
@@ -54,7 +54,7 @@ class BinomialGammaPotential(BasePotential):
             rate=beta,
         ).log_prob(self.x_o[:, :1].reshape(1, num_trials, -1))
 
-        joint_likelihood = torch.sum(logprob_choices, dim=2) + logprob_rts.squeeze()
+        joint_likelihood = torch.sum(logprob_choices, dim=-1) + logprob_rts.squeeze()
 
         assert joint_likelihood.shape == torch.Size([theta.shape[0], self.x_o.shape[0]])
         return joint_likelihood.sum(1)
