### 2025-02-24 031910a
fix: protocol and refactor for custom potential
