+++ sbi/inference/potentials/base_potential.py
@@ -1,9 +1,8 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
-import inspect
 from abc import ABCMeta, abstractmethod
-from typing import Callable, Optional
+from typing import Optional, Protocol
 
 import torch
 from torch import Tensor
@@ -83,12 +82,18 @@ class BasePotential(metaclass=ABCMeta):
         return self._x_o
 
 
-class CallablePotentialWrapper(BasePotential):
+class CustomPotential(Protocol):
+    """Protocol for custom potential functions."""
+
+    def __call__(self, theta: Tensor, x_o: Tensor) -> Tensor: ...
+
+
+class CustomPotentialWrapper(BasePotential):
     """If `potential_fn` is a callable it gets wrapped as this."""
 
     def __init__(
         self,
-        potential_fn: Callable,
+        potential_fn: CustomPotential,
         prior: Optional[Distribution],
         x_o: Optional[Tensor] = None,
         device: str = "cpu",
@@ -96,8 +101,7 @@ class CallablePotentialWrapper(BasePotential):
         """Wraps a callable potential function.
 
         Args:
-            potential_fn: Callable potential function, must have `theta` and `x_o` as
-                arguments.
+            potential_fn: Custom callable potential function.
             prior: Prior distribution.
             x_o: Observed data.
             device: Device on which to evaluate the potential function.
@@ -105,14 +109,6 @@ class CallablePotentialWrapper(BasePotential):
         """
         super().__init__(prior, x_o, device)
 
-        kwargs_of_callable = list(inspect.signature(potential_fn).parameters.keys())
-        required_keys = ["theta", "x_o"]
-        for key in required_keys:
-            assert key in kwargs_of_callable, (
-                "If you pass a `Callable` as `potential_fn` then it must have "
-                "`theta` and `x_o` as inputs, even if some of these keyword "
-                "arguments are unused."
-            )
         self.potential_fn = potential_fn
 
     def __call__(self, theta, track_gradients: bool = True):
@@ -121,4 +117,4 @@ class CallablePotentialWrapper(BasePotential):
         Note, x_o is re-used from the initialization of the potential function.
         """
         with torch.set_grad_enabled(track_gradients):
-            return self.potential_fn(theta=theta, x_o=self.x_o)
+            return self.potential_fn(theta, self.x_o)
