### 2025-02-25 234dcee
add tolerance options in log_prob; fix map test.
