### 2025-02-25 e063c58
tests: add pytest testmon plugin to speed up CI (#1402)
* add pytest testmon plugin

* test changes to check testmon

* allow testmon and markers

* further improvements for uv and testmon

* add test comment

* fix caching, test changes in mcmc posterior

* fix uv setup

* fix uv setup

* fix uv setup

* fix uv setup

* another fix

* remove pip from ci

* undo changes to test testmon

* load testmon cache first

* restore old testmon setup

* reorder jobs

* test comment

* back to uv pip

* test change

* seems to be working :tada:
