### 2025-02-25 dd4aef7
fix: rebuild flow logic in score methods (#1404)
* change rebuild_flow default to False

* rebuild flow for each new x, but not in __call__. fix map test

* remove old kwargs

* add tolerance options in log_prob; fix map test.

* fix tests
