### 2025-02-25 e13f7b3
chore: comments, doc, kwargs, last cleanup
