### 2025-02-25 5914851
feat:  add CategoricalMADE for estimating multiple discrete dimension (#1269)
* wip: first draft on categorical made

* wip: forward, log_prob, sample working

* wip: CategoricalMassEstimator can be build and MixedDensityEstimator too. log_prob has shape issues tho

* wip: sampling and log_prob works for categorical_made. working on getting mixed_density estimator log_probs and sample to work as well

* wip: build_mnle works and trains without log_transform. Add made as arg to categorical_model

* fix: categorical_made now trains in 1D MNLE

* fix: change net kwarg

* fix: verify ND training i
[…]
