### 2025-02-26 1a1a403
add pbar for sequential init params.
