+++ tests/mnle_test.py
@@ -171,14 +171,9 @@ def test_mnle_accuracy_with_different_samplers_and_trials(
     )
     trainer = MNLE(prior, density_estimator=density_estimator)
     trainer.append_simulations(theta, x).train(training_batch_size=200)
-    posterior = trainer.build_posterior()
 
     theta_o = prior.sample((1,))
     x_o = mixed_simulator(theta_o.repeat(num_trials, 1))
-
-    mcmc_kwargs = dict(
-        method="slice_np_vectorized", init_strategy="proposal", **mcmc_params_accurate
-    )
 
     # True posterior samples
     transform = mcmc_transform(prior)
@@ -186,10 +181,12 @@ def test_mnle_accuracy_with_different_samplers_and_trials(
         BinomialGammaPotential(prior, atleast_2d(x_o)),
         theta_transform=transform,
         proposal=prior,
-        **mcmc_kwargs,
+        **mcmc_params_accurate,
     ).sample((num_samples,), show_progress_bars=False)
 
-    posterior = trainer.build_posterior(prior=prior, sample_with=sampler)
+    posterior = trainer.build_posterior(
+        prior=prior, sample_with=sampler, mcmc_parameters=mcmc_params_accurate
+    )
     posterior.set_default_x(x_o)
     if sampler == "vi":
         posterior.train()
@@ -197,7 +194,7 @@ def test_mnle_accuracy_with_different_samplers_and_trials(
     mnle_posterior_samples = posterior.sample(
         sample_shape=(num_samples,),
         show_progress_bars=True,
-        **mcmc_kwargs if sampler == "mcmc" else {},
+        **mcmc_params_accurate if sampler == "mcmc" else {},
     )
 
     check_c2st(
@@ -241,21 +238,33 @@ class BinomialGammaPotential(BasePotential):
         batch_size = theta.shape[0]
         num_trials = self.x_o.shape[0]
         theta = theta.reshape(batch_size, 1, -1)
+        # We assume the InverseGamma to be in the first position of theta.
+        # And potentially multiple Binomials in the rest.
         beta, rhos = theta[:, :, :1], theta[:, :, 1:]
 
-        # vectorized
-        logprob_choices = torch.stack(
-            [Binomial(probs=rho).log_prob(self.x_o[:, 1:]) for rho in rhos],
+        # evaluate vectorized across batch of thetas.
+        logprob_choices = (
+            torch.stack(
+                [
+                    Binomial(probs=rhos[:, :, rho_idx]).log_prob(
+                        self.x_o[:, 1 + rho_idx]
+                    )
+                    for rho_idx in range(rhos.shape[-1])
+                ],
+            )
+            .transpose(0, 1)
+            .transpose(1, 2)
         )
 
         logprob_rts = InverseGamma(
-            concentration=self.concentration_scaling * torch.ones_like(beta),
-            rate=beta,
+            concentration=self.concentration_scaling * torch.ones_like(beta), rate=beta
         ).log_prob(self.x_o[:, :1].reshape(1, num_trials, -1))
 
+        # sum across parameter dimensions.
         joint_likelihood = torch.sum(logprob_choices, dim=-1) + logprob_rts.squeeze()
 
         assert joint_likelihood.shape == torch.Size([theta.shape[0], self.x_o.shape[0]])
+        assert joint_likelihood.isfinite().all()
         return joint_likelihood.sum(1)
 
 
@@ -271,61 +280,42 @@ def test_mnle_with_experimental_conditions(mcmc_params_accurate: dict):
     """
     num_simulations = 10000
     num_samples = 1000
[…]
