### 2025-02-26 f9de96c
fix batch size bug in sir
