### 2025-02-26 #7754 @boisgera
> Hi [@boisgera](https://github.com/boisgera) thanks a lot for looking into this. Yes, I will have a look at the issue for editing the author list and let you know.
>
> Thanks!

Hi @janfb,

Given [the issue comment](https://github.com/openjournals/inara/issues/105#issuecomment-2682005652) you'd better try a workaround. @arfon suggests that you don't use the `corresponding: true` flag, but instead manage directly all the foonotes manually (including the info that would have been generated by `corresponding : true`).

Would that work for you?
