### 2025-02-26 583c548
chore: drop Python 3.9 support, update workflows and docs.
