+++ .github/workflows/lint.yml
@@ -22,7 +22,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: '3.9'
+          python-version: '3.10'
       - uses: pre-commit/action@v3.0.1
         with:
           extra_args: --all-files --show-diff-on-failure
@@ -40,14 +40,11 @@ jobs:
       - name: Install uv and set the python version
         uses: astral-sh/setup-uv@v5
         with:
-          python-version: '3.9'
+          python-version: '3.10'
           enable-cache: true
 
       - name: Install dependencies with uv
-        run: |
-          uv pip install torch --extra-index-url https://download.pytorch.org/whl/cpu
-          uv pip install -e .[dev]
+        run: uv sync --all-extras --dev
 
       - name: Check types with pyright
-        run: |
-          uv run pyright sbi
+        run: uv run pyright sbi
