+++ .github/workflows/cd.yml
@@ -28,16 +28,15 @@ jobs:
     - name: Install uv and set the python version
       uses: astral-sh/setup-uv@v5
       with:
-        python-version: '3.9'
+        python-version: '3.10'
         enable-cache: true
+        cache-dependency-glob: "pyproject.toml"
 
     - name: Install dependencies with uv
-      run: |
-        uv pip install -e .[dev]
+      run: uv sync --all-extras --dev
 
     - name: Run the fast and the slow CPU tests with coverage
-      run: |
-        uv run pytest -v -x -n auto -m "not gpu" --cov=sbi --cov-report=xml tests/
+      run: uv run pytest -v -x -n auto -m "not gpu" --cov=sbi --cov-report=xml tests/
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v4-beta
