### 2025-02-26 12c034f
lint: checks and linting with py3.10.
