### 2025-02-26 d6aa95f
remove uv lock, update docs action
