+++ pyproject.toml
@@ -38,11 +38,11 @@ dependencies = [
     "pyknos>=0.16.0",
     "pyro-ppl>=1.3.1",
     "scikit-learn",
-    "scipy<1.13",
+    "scipy",
     "tensorboard",
     "torch>=1.13.0, <2.6.0",
     "tqdm",
-    "pymc>=5.21.0",
+    "pymc>=5.0.0",
     "zuko>=1.2.0",
 ]
 
