+++ .github/workflows/build_docs.yml
@@ -5,6 +5,7 @@ on:
       - main
   release:
     types: [ published ]
+  workflow_dispatch:
 
 jobs:
   docs:
