### 2025-02-27 c5bffae
load with weights only False to have pytorch 2.6 compatibility
