### 2025-02-27 b7e5a34
torch 2.6 compatibility
