### 2025-03-04 0e45d36
fix python 3.13 induced errors
