+++ pyproject.toml
@@ -33,14 +33,14 @@ dependencies = [
     "joblib>=1.0.0",
     "matplotlib",
     "notebook <= 6.4.12",
-    "numpy<2.0.0",
+    "numpy",
     "pillow",
     "pyknos>=0.16.0",
     "pyro-ppl>=1.3.1",
     "scikit-learn",
     "scipy",
     "tensorboard",
-    "torch>=1.13.0, <2.6.0",
+    "torch>=1.13.0",
     "tqdm",
     "pymc>=5.0.0",
     "zuko>=1.2.0",
