### 2025-03-04 f0ab7d2
fix torch deps and plot pyright
