### 2025-03-05 6e9a309
fix: fix mnle tests; pbar for MCMC init, fix sir batch size. (#1410)
* add pbar for sequential init params.

* better mcmc args for mnle notebook

* fix mnle tests.

* fix batch size bug in sir

* remove outputs from notebooks, update figures.
