### 2025-03-05 c2ff942
fix: mnle tests (#1415)
