### 2025-03-06 d51e9fb
add numfocus badge and credits in readme
