+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: Tests
+name: Continuous Integration
 
 on: [pull_request, workflow_dispatch]
 
@@ -12,7 +12,7 @@ concurrency:
 
 jobs:
   ci:
-    name: CI
+    name: Continuous Integration
     runs-on: ubuntu-latest
     if: |
       github.event_name == 'push' ||
@@ -20,7 +20,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ['3.10', '3.11', '3.12', '3.13']
+        python-version: ['3.10']
 
     steps:
     - name: Checkout
@@ -57,7 +57,7 @@ jobs:
         restore-keys: |
           testmon-${{ runner.os }}-${{ matrix.python-version }}-
 
-    - name: Run the fast CPU tests with coverage
+    - name: Run fast CPU tests with coverage
       run: uv run pytest --testmon-forceselect -v -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
 
     - name: Upload coverage to Codecov
