+++ tests/lc2st_test.py
@@ -189,7 +189,7 @@ def test_lc2st_true_positiv_rate(method):
 
 @pytest.mark.slow
 @pytest.mark.parametrize("method", (LC2ST, LC2ST_NF))
-def test_lc2st_false_positiv_rate(method):
+def test_lc2st_false_positiv_rate(method, set_seed):
     """Tests the false positiv rate of the LC2ST-(NF) test:
     for a "good" estimator, the LC2ST-(NF) should not reject the null hypothesis."""
     num_runs = 100
@@ -198,7 +198,6 @@ def test_lc2st_false_positiv_rate(method):
     # use big num_train and num_epochs to obtain "good" estimator
     # (convergence of the estimator)
     num_train = 5_000
-    num_epochs = 200
 
     num_cal = 1_000
     num_eval = 10_000
@@ -215,7 +214,7 @@ def test_lc2st_false_positiv_rate(method):
     # Train the neural posterior estimators
     inference = NPE(prior, density_estimator='maf')
     inference = inference.append_simulations(theta=theta_train, x=x_train)
-    npe = inference.train(training_batch_size=100, max_num_epochs=num_epochs)
+    npe = inference.train(training_batch_size=100)
 
     thetas = prior.sample((num_cal,))
     xs = simulator(thetas)
