+++ .github/workflows/cd.yml
@@ -18,6 +18,7 @@ jobs:
     name: Continuous Deployment
     runs-on: ubuntu-latest
     strategy:
+      fail-fast: false
       matrix:
         python-version: ['3.10', '3.11', '3.12', '3.13']
 
