### 2025-03-10 a72d750
build: drop py3.9, fix deps, fix types, lc2st tests (#1412)
* chore: drop Python 3.9 support, update workflows and docs.

* lint: checks and linting with py3.10.

* remove uv lock, update docs action

* fix pymc x-failing tests

* fix python 3.13 induced errors

* fix torch deps and plot pyright

* Fix circular import test

* test more python versions

* add numfocus badge and credits in readme

* remove testmon temporarily

* add testmon adain

* test python version in CD

* change locally passing but ci failing test

* turn off fast fail

* remote test still failing, increas n

---------

Co
[…]
