+++ pyproject.toml
@@ -24,7 +24,7 @@ classifiers = [
     "Programming Language :: Python :: 3",
     "Development Status :: 3 - Alpha",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dynamic = ["version"]
 readme = "README.md"
 keywords = ["Bayesian inference", "simulation-based inference", "PyTorch"]
@@ -33,12 +33,12 @@ dependencies = [
     "joblib>=1.0.0",
     "matplotlib",
     "notebook <= 6.4.12",
-    "numpy<2.0.0",
+    "numpy",
     "pillow",
     "pyknos>=0.16.0",
     "pyro-ppl>=1.3.1",
     "scikit-learn",
-    "scipy<1.13",
+    "scipy",
     "tensorboard",
     "torch>=1.13.0, <2.6.0",
     "tqdm",
@@ -140,7 +140,7 @@ xfail_strict = true
 [tool.pyright]
 include = ["sbi"]
 exclude = ["**/__pycache__", "**/__node_modules__", ".git", "docs", "tutorials", "tests"]
-python_version = "3.9"
+python_version = "3.10"
 reportUnsupportedDunderAll = false
 reportGeneralTypeIssues = false
 reportInvalidTypeForm = false
