### 2025-03-10 7b6cab3
ci: fix docs workflow (#1419)
* fix docs workflow

* fix uv extra syntax.

* fix numfocus badge
