### 2025-03-11 51da431
Update install.md
Updated python version to be >=3.10
