### 2025-03-11 6e9a9dc
Updated with working xdist support now. Simplified API. Saves result human-readable
