### 2025-03-11 928932b
Update docs/docs/contribute.md
Co-authored-by: Jan <janfb@users.noreply.github.com>
