+++ sbi/inference/potentials/score_fn_iid.py
@@ -1,6 +1,6 @@
 import functools
 import math
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 from typing import Callable, Optional, Type
 
 import torch
@@ -20,7 +20,7 @@ from sbi.utils.torchutils import ensure_theta_batched
 IID_METHODS = {}
 
 
-def get_iid_method(name: str) -> Type["ScoreFnIID"]:
+def get_iid_method(name: str) -> Type["IIDScoreFunction"]:
     r"""
     Retrieves the IID method by name.
 
@@ -48,14 +48,14 @@ def register_iid_method(name: str) -> Callable:
         A decorator function to register the IID method class.
     """
 
-    def decorator(cls: Type["ScoreFnIID"]) -> Type["ScoreFnIID"]:
+    def decorator(cls: Type["IIDScoreFunction"]) -> Type["IIDScoreFunction"]:
         IID_METHODS[name] = cls
         return cls
 
     return decorator
 
 
-class ScoreFnIID:
+class IIDScoreFunction(ABC):
     def __init__(
         self,
         score_estimator: "ConditionalScoreEstimator",
@@ -63,7 +63,17 @@ class ScoreFnIID:
         device: str = "cpu",
     ) -> None:
         r"""
-        Initializes the ScoreFnIID class.
+        This is a abstract base class wrapper for score estimators. Subclasses are used
+        to implement different methods for factorized distributions. For example, in
+        the IID setting the posterior for N observations can be represented as a product
+        of N "local" posteriors, divided by N-1 prior terms.
+
+        This allows to efficiently extend "single" observation score estimators to a
+        sequence of IID observtions. Unfortunatly, this is not as simple as just summing
+        the scores minus the prior score, as in diffusion models the we also need to
+        represent the "marginal" posterior scores at time $t$. Even if the true
+        posterior factorizes, the marginal true posterior at time $t>0$ does not and
+        this requires some adjustments.
 
         Args:
             score_estimator: The neural network modeling the score.
@@ -71,25 +81,107 @@ class ScoreFnIID:
             device: The device on which to evaluate the potential. Defaults to "cpu".
         """
 
-        self.score_estimator = score_estimator.to(device)
+        self.score_estimator = score_estimator.to(device).eval()
         self.prior = prior
-        self.score_estimator.eval()
 
     @abstractmethod
     def __call__(
         self,
-        theta: Tensor,
+        inputs: Tensor,
+        conditions: Tensor,
+        time: Optional[Tensor] = None,
     ) -> Tensor:
         r"""
         Abstract method to be implemented by subclasses to compute the score function.
 
         Args:
-            theta: The parameters at which to evaluate the potential.
+            inputs: The parameters at which to evaluate the potential of size [b,iid,d]
+            conditions: The sequence of observations of size [iid,...]
+            time: The time in the diffusion process to specify the target marginal.
+
+        Returns:
+            The score of the inputs given N observations of shape (batch,d).
+        """
+        pass
+
+
+@register_iid_method("fnpe")
+class FNPEScoreFunction(IIDScoreFunction):
+    def __init__(
+        self,
+        score_estimator: "ConditionalScoreEstimator",
+        prior: Distribution,
+        device: str = "cpu",
+        prior_score_weight: Optional[Callable[[Tensor], Tensor]] = None,
+    ) -> None:
+        r"""
+        The FNPEScoreFunction implments the "Factorized Neural Posterior Estimation"
+        method for score-based models [1]. This method does not apply the necessary
+        corrections for the score function, but instead uses a simple weighting of the
[…]
