+++ sbi/utils/score_utils.py
@@ -21,17 +21,17 @@ from sbi.utils.torchutils import BoxUniform
 
 
 def denoise(p: Distribution, m: Tensor, s: Tensor, x_t: Tensor) -> Distribution:
-    """Given the prior distribution p(X), scaling factor m, standard deviation of the
-    noise s, and observation X_t, return the posterior distribution p(X | X_t = x_t).
+    """Returns the denoised distribution p(X | X_t = x_t).
+
+    This return the posterior distribution p(X | X_t = x_t) given the prior
+    distribution p(X) and the following generative process X_t = m * X + s*ε, where
+    ε ~ N(0, 1).
 
     Args:
         p: The prior distribution p(X).
         m: The scaling factor m.
         s: The standard deviation of the noise s.
         x_t: The observation X_t.
-
-    Raises:
-        NotImplementedError: If the distribution is not supported.
 
     Returns:
         The posterior distribution p(X | X_t = x_t).
@@ -47,7 +47,7 @@ def denoise(p: Distribution, m: Tensor, s: Tensor, x_t: Tensor) -> Distribution:
     elif isinstance(p, (Uniform, BoxUniform)):
         return denoise_uniform(p, m, s, x_t)
     else:
-        return denoise_empirical(p, m, s, x_t)
+        return denoise_general(p, m, s, x_t)
 
 
 def denoise_independent(
@@ -167,13 +167,21 @@ def denoise_uniform(
     return UniformNormalPosterior(p.low, p.high, m, s, x_t)  # type: ignore
 
 
-def denoise_empirical(
+def denoise_general(
     p: Distribution,
     m: Tensor,
     s: Tensor,
     x_t: Tensor,
 ) -> MixtureSameFamily:
-    """Denoise an empirical distribution.
+    """Denoise a general distribution.
+
+    This is an approximate method intended for general distributions. It fits a
+    Gaussian Mixture Model (GMM) to the empirical distribution, which can then be
+    denoised analytically using the denoise_mixture function.
+
+    NOTE: Why GMM not KDE or MC methods? This is because in one intended use case we
+    require access to the score function (i.e. the gradient of the log density), and
+    this gradient is rather ill-behave for particle-based approximations.
 
     Args:
         p: The prior empirical distribution.
@@ -361,7 +369,15 @@ class UniformNormalPosterior(Distribution):
     }  # type: ignore
     has_rsample: bool = False
 
-    def __init__(self, low, high, m, s, x_t, validate_args=None):
+    def __init__(
+        self,
+        low: Tensor | float,
+        high: Tensor | float,
+        m: Tensor | float,
+        s: Tensor | float,
+        x_t: Tensor | float,
+        validate_args: Optional[bool] = None,
+    ) -> None:
         self.low = torch.as_tensor(low)
         self.high = torch.as_tensor(high)
         self.m = torch.as_tensor(m)
@@ -391,7 +407,7 @@ class UniformNormalPosterior(Distribution):
         super().__init__(batch_shape, event_shape, validate_args=validate_args)
 
     @property
-    def mean(self):
+    def mean(self) -> Tensor:
         # Standard normal pdf values
         phi_alpha = torch.exp(self.standard_normal.log_prob(self.alpha))
         phi_beta = torch.exp(self.standard_normal.log_prob(self.beta))
@@ -399,7 +415,7 @@ class UniformNormalPosterior(Distribution):
         return self.mu + self.sigma * (phi_alpha - phi_beta) / self.Z
 
     @property
-    def variance(self):
+    def variance(self) -> Tensor:
         phi_alpha = torch.exp(self.standard_normal.log_prob(self.alpha))
[…]
