### 2025-03-11 2032a53
Refactored tests with fixtures!
