### 2025-03-12 8b8179f
Fix FNPE intitialization - not that nice
