+++ sbi/inference/potentials/score_based_potential.py
@@ -171,7 +171,6 @@ class PosteriorScoreBasedPotential(BasePotential):
                 assert self.prior is not None, "Prior is required for iid methods."
 
                 iid_method = get_iid_method(self.iid_method)
-                # Always creating a new object every call is not efficient...
                 score_fn_iid = iid_method(
                     self.score_estimator, self.prior, **(self.iid_params or {})
                 )
