### 2025-03-12 a9978a5
Fix negative interaction with normal pytest xdist
