### 2025-03-12 0ea6f63
Mini sbibm (#1335)
* Start of mini_sbibm

* Working prototype

* Rename to get globally discovered

* ruff

* extended to something reasonable

* remove example

* different modes

* Added docs and formating

* colored numbers

* formating and stuff

* formating?

* Fixing suggestions and adding dependency

* Using taskt dict

* licence note

* Update tests/bm_test.py

Co-authored-by: Jan <janfb@users.noreply.github.com>

* num_eval_obs fix, remove xdist comment

* Clarifing the fixture

* remove comment

* Adding xdist cache to gitignore

* load with weights only False to have pytorch 2.6 com
[…]
