+++ tests/conftest.py
@@ -1,13 +1,20 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+import re
+import shutil
+from pathlib import Path
+from shutil import rmtree
 
+import pandas as pd
 import pytest
 import torch
+from pytest_harvest import get_session_results_df, get_xdist_worker_id, is_main_process
 
 from sbi.utils.sbiutils import seed_all_backends
 
 # Seed for `set_seed` fixture. Change to random state of all seeded tests.
 seed = 1
+harvested_fixture_data = None
 
 
 # Use seed automatically for every test function.
@@ -29,9 +36,174 @@ def pytest_collection_modifyitems(config, items):
     )
     if not gpu_device_available:
         skip_gpu = pytest.mark.skip(reason="No devices available")
+
         for item in items:
             if "gpu" in item.keywords:
                 item.add_marker(skip_gpu)
+
+    if not config.getoption("--bm"):
+        # Skip marked benchmarking tests
+        skip_bm = pytest.mark.skip(reason="Benchmarking disabled")
+        for item in items:
+            if "benchmark" in item.keywords:
+                item.add_marker(skip_bm)
+    else:
+        # Filter tests to only those with the 'benchmark' marker
+        filtered_items = []
+        for item in items:
+            if item.get_closest_marker("benchmark"):
+                filtered_items.append(item)
+
+        items[:] = filtered_items  # Inplace!
+
+
+# Run mini-benchmark tests with `pytest --print-harvest`
+def pytest_addoption(parser):
+    parser.addoption(
+        "--bm",
+        action="store_true",
+        default=False,
+        help="Run mini-benchmark tests with specified mode",
+    )
+    parser.addoption(
+        "--bm-mode",
+        action="store",
+        default=None,
+        help="Run mini-benchmark tests with specified mode",
+    )
+
+
+@pytest.fixture
+def benchmark_mode(request):
+    """Fixture to access the --bm value in test files."""
+    return request.config.getoption("--bm-mode")
+
+
+@pytest.fixture(scope="session", autouse=True)
+def finalize_fixture_store(request, fixture_store):
+    # The code before `yield` runs at the start of the session (before tests).
+    yield
+    # The code after `yield` runs after all tests have completed.
+    # At this point, fixture_store should have all the harvested data.
+    global harvested_fixture_data
+    harvested_fixture_data = dict(fixture_store)
+
+
+def strip_ansi_escape_codes(text):
+    ansi_escape = re.compile(r'\x1b\[.*?m')
+    return ansi_escape.sub('', text)
+
+
+# Function to center text with ANSI colors, adjusting for escape codes
+def center_colored_text(text, width):
+    visible_length = len(strip_ansi_escape_codes(text))
+    padding = max(0, (width - visible_length) // 2)
+    return " " * padding + text + " " * (width - visible_length - padding)
+
+
+def pytest_terminal_summary(terminalreporter, exitstatus, config):
+    """
+    Custom pytest terminal summary to display mini SBIBM results with relative coloring
+    per task.
+
+    This function is called after the test session ends and generates a summary
+    of the results if the `--bm` option is specified. It displays the results
+    in a formatted table with methods as rows and tasks as columns, applying
+    relative coloring to metrics based on their performance within each task.
+    """
+    if config.getoption("--bm"):
+        terminal_width = shutil.get_terminal_size().columns
+        summary_text = " mini SBIBM results "
+        centered_line = summary_text.center(terminal_width, '=')
+        colored_line = f"\033[96m{centered_line}\033[0m"
+        terminalreporter.write_line(colored_line)
+
[…]
