### 2025-03-13 7a87864
Fixed MD040 in README.md, provided bash specification for Install section code blocks
