### 2025-03-13 95a9c23
Fixed MD012/no-multiple-blanks: Multiple consecutive blank lines in README.md, removed double blank line
