### 2025-03-13 642e613
Fix MD034/no-bare-urls i Highlists section of CHANGELOG.md
