### 2025-03-13 0d7c3f7
Fix MD034/no-bare-urls in New Contributors section of CHANGELOG.md
