### 2025-03-13 60f672f
Fix MD004/ul-style for coherent unordered list marked (dash) in CHANGELOG.md
