### 2025-03-13 4663b45
Fix MD034/no-bare-urls in v0.23.0 Maintaince and other changes of CHANGELOG.md
