### 2025-03-13 39ae551
Fix MD045/no-alt-text for goal.png in docs/index.md, now goal.png has an alt text for accessibility and SEO
