### 2025-03-13 6f3895b
Fix MD040/fenced-code-language in docs/install.md for bash blocks
