### 2025-03-13 66c2be6
Fix MD028/no-blanks-blockquote in docs/credits.md Licence section
