### 2025-03-13 27aaff8
Fix MD045/no-alt-text for logo_bmbf.svg in docs/credits.md, now the BMFM logo image has an alt text for Accessibility and SEO
