### 2025-03-14 c36fc5b
sbi diagram more meaningful alt description in docs/index.md
Co-authored-by: Jan <janfb@users.noreply.github.com>
