### 2025-03-14 efe2760
docs: markdown formatting according to markdown linter (solves #1434) (#1443)
* Fixed MD040 in README.md, provided bash specification for Install section code blocks

* Fixed MD004/ul-style Unordered list style [Expected: dash; Actual: asterisk] in Tutorials section

* Fixed MD012/no-multiple-blanks: Multiple consecutive blank lines in README.md, removed double blank line

* Fixed MD034/no-bare-urls in CODE_OF_CONDUCT for email reference in Enforcement section

* Fixed MD032/blanks-around-lists in faq/question_01, divided an unordered list item with its following code block

* Fix MD034/no-bar
[…]
