### 2025-03-14 61d801f
update changelog, fix gitignore
