+++ CHANGELOG.md
@@ -1,11 +1,58 @@
-# v0.23.3
+# Changelog
 
-## Highlights 🤩
+## v0.24.0
+
+### ✨ Highlights
+
+- Add `CategoricalMADE` by @jnsbck in https://github.com/sbi-dev/sbi/pull/1269  **(Major New Feature)**
+- tests: `mini-sbibm` by @manuelgloeckler in https://github.com/sbi-dev/sbi/pull/1335 **(Major New Feature)**
+- Drop python3.9 support, fix ci by @janfb in https://github.com/sbi-dev/sbi/pull/1412 **(Python Version Support Change)**
+- additional features for NPSE by @gmoss13 in https://github.com/sbi-dev/sbi/pull/1370 **(Enhancement)**
+
+### 🐛 Bug Fixes
+
+- #1350 leakage correction breaks consistency of log prob vs log prob batched by @manuelgloeckler in https://github.com/sbi-dev/sbi/pull/1355
+- fix #1343 device handling in mog_log_prob by @janfb in https://github.com/sbi-dev/sbi/pull/1356
+- Fix failing tutorials, change MNLE default for log_transform to False by @janfb in https://github.com/sbi-dev/sbi/pull/1367
+- Fix conditional posterior shape and device bugs. by @janfb in https://github.com/sbi-dev/sbi/pull/1373
+- fix: type fix in VI subclasses. xfail pymc tests. by @janfb in https://github.com/sbi-dev/sbi/pull/1390
+- Temporary Wrappers to fix MADE by @gmoss13 in https://github.com/sbi-dev/sbi/pull/1398
+- Fix mnle tests, MCMC init pbar, sir batch size. by @janfb in https://github.com/sbi-dev/sbi/pull/1410
+- fix mnle tests by @janfb in https://github.com/sbi-dev/sbi/pull/1415
+- fix: protocol and refactor for custom potential by @janfb in https://github.com/sbi-dev/sbi/pull/1409
+- fix docs workflow by @janfb in https://github.com/sbi-dev/sbi/pull/1419
+
+### 🛠️ Maintenance & Improvements
+
+- increase tutorial test timeout by @janfb in https://github.com/sbi-dev/sbi/pull/1360
+- add nan check to _loss method. by @janfb in https://github.com/sbi-dev/sbi/pull/1361
+- Update and pin pre commit and ruff to recent version. by @janfb in https://github.com/sbi-dev/sbi/pull/1358
+- add nan handling in diagnostics by @janfb in https://github.com/sbi-dev/sbi/pull/1359
+- Improve tests to detect circular imports and resolve all of them by @manuelgloeckler in https://github.com/sbi-dev/sbi/pull/1357
+- docs: add details about zuko density estimators by @janfb in https://github.com/sbi-dev/sbi/pull/1387
+- fix: duplication of 'large number' in warning by @turnmanh in https://github.com/sbi-dev/sbi/pull/1391
+- improve PR and issue templates by @janfb in https://github.com/sbi-dev/sbi/pull/1399
+- perf: speed up CI with uv by @janfb in https://github.com/sbi-dev/sbi/pull/1400
+- tests: add pytest testmon plugin to speed up CI by @janfb in https://github.com/sbi-dev/sbi/pull/1402
+- small fixes to score methods by @janfb in https://github.com/sbi-dev/sbi/pull/1404
+- docs: Corrected python version in installation document (Revised) by @VijaySamant4368 in https://github.com/sbi-dev/sbi/pull/1423
+- docs: Markdown formatting compliant to Markdown Linter (solves #1434) by @nMaax in https://github.com/sbi-dev/sbi/pull/1443
+
+### 🎉 New Contributors
+
[…]
