### 2025-03-14 5b6b143
add note on AI code to contribute.md
