+++ docs/docs/contribute.md
@@ -296,3 +296,31 @@ cd docs
 jupyter nbconvert --to markdown ../tutorials/*.ipynb --output-dir docs/tutorials/
 mkdocs serve
 ```
+
+### Using AI Coding Assistants
+
+We understand that AI coding assistants (like GitHub Copilot, ChatGPT, etc.) can be
+helpful tools.  You are welcome to use them when contributing to this project, but *with
+caution and responsibility*.
+
+- **Understand the Code:** Do *not* blindly accept suggestions from AI assistants. You
+    are responsible for ensuring that any code you submit (whether written by you or
+    generated by an AI) is correct, efficient, secure, and follows all project
+    guidelines.
+- **Thoroughly Review AI-Generated Code:** Treat AI-generated code as you would any
+  other code you didn't write yourself: review it carefully, line by line.  Understand
+  what it does, how it does it, and why it's the best approach.
+- **Test Extensively:** AI assistants can make mistakes (hallucinations, subtle bugs,
+  inefficient code).  Write comprehensive unit tests to verify the correctness of any
+  AI-generated code.  Don't rely solely on the AI to test its own code.
+- **Attribution:** If a significant portion of the code was generated by an AI
+  assistant, briefly mention this in your commit message or pull request description
+  (e.g., "Implemented feature X with assistance from GitHub Copilot"). This helps
+  reviewers understand the origin of the code.  However, you remain responsible for the
+  code's correctness.
+- **Maintainability:** Make sure the code is well-formatted, commented and does follow
+  our code style.
+
+**In essence: Use AI assistants as a *tool* to enhance your productivity, but *you* are
+the programmer.  You are ultimately responsible for the quality and correctness of the
+code you contribute.**
