### 2025-03-14 8fc7ec6
Raise error on potential. Make BaseClass abstract and remove redundant arguments.
