### 2025-03-14 7d43073
fix: gpu-handling for made and acceptance rate (#1448)
* fix: gpu-handling for made and leakage correction

* Fix buffers

---------

Co-authored-by: manuelgloeckler <manu.gloeckler@hotmail.de>
