+++ sbi/inference/potentials/score_based_potential.py
@@ -108,8 +108,8 @@ class PosteriorScoreBasedPotential(BasePotential):
         super().set_x(x_o, x_is_iid)
         self.iid_method = iid_method
         self.iid_params = iid_params
-        # NOTE: Once IID potential evaluation is supported. This might needs to be
-        # adapted. See #1450.
+        # NOTE: Once IID potential evaluation is supported. This needs to be adapted.
+        # See #1450.
         if not x_is_iid and (self._x_o is not None):
             self.flow = self.rebuild_flow(atol=atol, rtol=rtol, exact=exact)
 
@@ -130,9 +130,10 @@ class PosteriorScoreBasedPotential(BasePotential):
 
         if self.x_is_iid:
             raise NotImplementedError(
-                "Potential function evaluation in the IID setting not yet supported for\
-                score-based methods. Sampling does however work `.sample`. If you  \
-                insetad like to evaluate multiple posteriors use `log_prob_batched`."
+                "Potential function evaluation in the IID setting is not yet supported"
+                " for score-based methods. Sampling does however work via `.sample`. "
+                "If you intended to evaluate the posterior given a batch of (non-iid) "
+                "x use `log_prob_batched`."
             )
 
         theta = ensure_theta_batched(torch.as_tensor(theta))
