### 2025-03-14 1ee577e
feat: score-based iid sampling (#1381)
* npse MAP

* set default enable_Transform to True

* ruff formatting version change

* sampling via diffusion twice

* batched sampling for score-based posteriors

* iid api integration

* new ruff

* adding corrector back

* adding untesed GAUSS

* All other methods

* reformat

* messy version of simple gauss, API to sample method

* jac method (still needs feasible Lambda projection to work)

* Only jac left

* Formating and so on

* Adding correct tests

* Add empirical support - but this doesnt work that well

* A bunch of auto marginalize and denoi
[…]
