+++ sbi/utils/score_utils.py
@@ -1,0 +1,620 @@
+from functools import lru_cache
+from typing import Optional
+
+import torch
+from sklearn.mixture import GaussianMixture
+from torch import Tensor
+from torch.distributions import (
+    Categorical,
+    Distribution,
+    Independent,
+    MixtureSameFamily,
+    MultivariateNormal,
+    Normal,
+    Uniform,
+    constraints,
+)
+
+from sbi.utils.torchutils import BoxUniform
+
+# Automatic denoising ------------------------------------------------------
+
+
+def denoise(p: Distribution, m: Tensor, s: Tensor, x_t: Tensor) -> Distribution:
+    """Returns the denoised distribution p(X | X_t = x_t).
+
+    This return the posterior distribution p(X | X_t = x_t) given the prior
+    distribution p(X) and the following generative process X_t = m * X + s*ε, where
+    ε ~ N(0, 1).
+
+    Args:
+        p: The prior distribution p(X).
+        m: The scaling factor m.
+        s: The standard deviation of the noise s.
+        x_t: The observation X_t.
+
+    Returns:
+        The posterior distribution p(X | X_t = x_t).
+    """
+    if isinstance(p, Independent):
+        return denoise_independent(p, m, s, x_t)
+    elif isinstance(p, MixtureSameFamily):
+        return denoise_mixture(p, m, s, x_t)
+    elif isinstance(p, Normal):
+        return denoise_gaussian(p, m, s, x_t)
+    elif isinstance(p, MultivariateNormal):
+        return denoise_multivariate_gaussian(p, m, s, x_t)
+    elif isinstance(p, (Uniform, BoxUniform)):
+        return denoise_uniform(p, m, s, x_t)
+    else:
+        return denoise_general(p, m, s, x_t)
+
+
+def denoise_independent(
+    p: Independent, m: Tensor, s: Tensor, x_t: Tensor
+) -> Independent:
+    """Denoise an independent distribution.
+
+    Args:
+        p: The prior independent distribution.
+        m: The scaling factor m.
+        s: The standard deviation of the noise s.
+        x_t: The observation X_t.
+
+    Returns:
+        The posterior independent distribution.
+    """
+    return Independent(denoise(p.base_dist, m, s, x_t), p.reinterpreted_batch_ndims)
+
+
+def denoise_gaussian(p: Normal, m: Tensor, s: Tensor, x_t: Tensor) -> Normal:
+    """Denoise a Gaussian distribution.
+
+    Args:
+        p: The prior Gaussian distribution.
+        m: The scaling factor.
+        s: The standard deviation of the noise.
+        x_t: The observed data.
+
+    Returns:
+        The posterior Gaussian distribution.
+    """
+    mean_0 = p.loc
+    std_0 = p.scale
+
+    # Calculate the posterior mean and variance
+    posterior_variance = 1 / (1 / std_0**2 + m**2 / s**2)
+    posterior_mean = posterior_variance * (mean_0 / std_0**2 + m * x_t / s**2)
+    posterior_std = torch.sqrt(posterior_variance)
+
+    return Normal(posterior_mean, posterior_std)
+
+
+def denoise_multivariate_gaussian(
+    p: MultivariateNormal, m: Tensor, s: Tensor, x_t: Tensor
+) -> MultivariateNormal:
+    """Denoise a multivariate Gaussian distribution.
+
+    Args:
+        p: The prior multivariate Gaussian distribution.
+        m: The scaling factor.
+        s: The standard deviation of the noise.
+        x_t: The observed data.
+
+    Returns:
+        The posterior multivariate Gaussian distribution.
+    """
+    mean0 = p.loc
+    cov0 = p.covariance_matrix
+    n = cov0.size(-1)  # type: ignore
+    # Support batch dimensions by expanding the identity matrix to match cov0's
[…]
