+++ sbi/inference/potentials/score_fn_iid.py
@@ -1,0 +1,744 @@
+import functools
+import math
+from abc import ABC, abstractmethod
+from typing import Callable, Optional, Type
+
+import torch
+from torch import Tensor
+from torch.distributions import Distribution
+
+from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
+from sbi.utils.score_utils import (
+    add_diag_or_dense,
+    denoise,
+    marginalize,
+    mv_diag_or_dense,
+    solve_diag_or_dense,
+)
+from sbi.utils.torchutils import ensure_theta_batched
+
+IID_METHODS = {}
+
+
+def get_iid_method(name: str) -> Type["IIDScoreFunction"]:
+    r"""
+    Retrieves the IID method by name.
+
+    Args:
+        name: The name of the IID method.
+
+    Returns:
+        The IID method class.
+    """
+    if name not in IID_METHODS:
+        raise NotImplementedError(
+            f"Method {name} for iid score accumulation not implemented."
+        )
+    return IID_METHODS[name]
+
+
+def register_iid_method(name: str) -> Callable:
+    r"""
+    Registers an IID method.
+
+    Args:
+        name: The name of the IID method.
+
+    Returns:
+        A decorator function to register the IID method class.
+    """
+
+    def decorator(cls: Type["IIDScoreFunction"]) -> Type["IIDScoreFunction"]:
+        IID_METHODS[name] = cls
+        return cls
+
+    return decorator
+
+
+class IIDScoreFunction(ABC):
+    def __init__(
+        self,
+        score_estimator: "ConditionalScoreEstimator",
+        prior: Distribution,
+        device: str = "cpu",
+    ) -> None:
+        r"""
+        This is a abstract base class wrapper for score estimators.
+
+        Subclasses are used to implement different methods for factorized distributions.
+        For example, in the IID setting the posterior for N observations can be
+        represented as a product of N "local" posteriors, divided by N-1 prior terms.
+        This allows to efficiently extend "single" observation score estimators to a
+        sequence of IID observtions. Unfortunatly, this is not as simple as just summing
+        the scores minus the prior score, as in diffusion models the we also need to
+        represent the "marginal" posterior scores at time $t$. Even if the true
+        posterior factorizes, the marginal true posterior at time $t>0$ does not and
+        this requires some adjustments.
+
+        Args:
+            score_estimator: The neural network modeling the score.
+            prior: The prior distribution.
+            device: The device on which to evaluate the potential. Defaults to "cpu".
+        """
+
+        self.score_estimator = score_estimator.to(device).eval()
+        self.prior = prior
+
+    @abstractmethod
+    def __call__(
+        self,
+        inputs: Tensor,
+        conditions: Tensor,
+        time: Optional[Tensor] = None,
+    ) -> Tensor:
+        r"""
+        Abstract method to be implemented by subclasses to compute the score function.
+
+        Args:
+            inputs: The parameters at which to evaluate the potential of size [b,iid,d]
+            conditions: The sequence of observations of size [iid,...]
+            time: The time in the diffusion process to specify the target marginal.
+
+        Returns:
+            The score of the inputs given N observations of shape (batch,d).
+        """
+        pass
+
+
+@register_iid_method("fnpe")
+class FNPEScoreFunction(IIDScoreFunction):
+    def __init__(
+        self,
+        score_estimator: "ConditionalScoreEstimator",
+        prior: Distribution,
+        device: str = "cpu",
+        prior_score_weight: Optional[Callable[[Tensor], Tensor]] = None,
+    ) -> None:
+        r"""
+        The FNPEScoreFunction implments the "Factorized Neural Posterior Estimation"
+        method for score-based models [1].
+
+        This method does not apply the necessary corrections for the score function, but
[…]
