+++ sbi/inference/posteriors/score_posterior.py
@@ -105,6 +105,8 @@ class ScorePosterior(NeuralPosterior):
         corrector_params: Optional[Dict] = None,
         steps: int = 500,
         ts: Optional[Tensor] = None,
+        iid_method: str = "auto_gauss",
+        iid_params: Optional[Dict] = None,
         max_sampling_batch_size: int = 10_000,
         sample_with: Optional[str] = None,
         show_progress_bars: bool = True,
@@ -123,6 +125,19 @@ class ScorePosterior(NeuralPosterior):
             steps: Number of steps to take for the Euler-Maruyama method.
             ts: Time points at which to evaluate the diffusion process. If None, a
                 linear grid between t_max and t_min is used.
+            iid_method: Which method to use for computing the score in the iid setting.
+                We currently support "fnpe", "gauss", "auto_gauss", "jac_gauss". The
+                fnpe method is simple and generally applicable. However, it can become
+                inaccurate already for quite a few iid samples (as it based on heuristic
+                approximations), and should be used at best only with a `corrector`. The
+                "gauss" methods are more accurate, by aiming for an efficient
+                approximation of the correct marginal score in the iid case. This
+                however requires estimating some hyperparamters, which is done in a
+                systematic way in the "auto_gauss" (initial overhead) and "jac_gauss"
+                (iterative jacobian computations are expensive). We default to
+                "auto_gauss" for these reasons.
+            iid_params: Additional parameters passed to the iid method. See the specific
+                `IIDScoreFunction` child class for details.
             max_sampling_batch_size: Maximum batch size for sampling.
             sample_with: Deprecated - use `.build_posterior(sample_with=...)` prior to
                 `.sample()`.
@@ -138,7 +153,10 @@ class ScorePosterior(NeuralPosterior):
 
         x = self._x_else_default_x(x)
         x = reshape_to_batch_event(x, self.score_estimator.condition_shape)
-        self.potential_fn.set_x(x, x_is_iid=True)
+        is_iid = x.shape[0] > 1
+        self.potential_fn.set_x(
+            x, x_is_iid=is_iid, iid_method=iid_method, iid_params=iid_params
+        )
 
         num_samples = torch.Size(sample_shape).numel()
 

[…]
