+++ sbi/inference/potentials/score_fn_iid.py
@@ -648,7 +648,7 @@ class JacCorrectedScoreFn(BaseGaussCorrectedScoreFunction):
     """
 
     def posterior_precision_est_fn(self, conditions: Tensor) -> Tensor:
-        raise ValueError("This method does not use the posterior precision estimation.")
+        pass
 
     def marginal_denoising_posterior_precision_est_fn(
         self,
