+++ sbi/inference/potentials/score_fn_iid.py
@@ -648,7 +648,9 @@ class JacCorrectedScoreFn(BaseGaussCorrectedScoreFunction):
     """
 
     def posterior_precision_est_fn(self, conditions: Tensor) -> Tensor:
-        pass
+        raise NotImplementedError(
+            "This method does not use the posterior precision estimation."
+        )
 
     def marginal_denoising_posterior_precision_est_fn(
         self,
