### 2025-03-14 8aecbe4
Value error is better
