+++ sbi/inference/potentials/score_fn_iid.py
@@ -648,9 +648,7 @@ class JacCorrectedScoreFn(BaseGaussCorrectedScoreFunction):
     """
 
     def posterior_precision_est_fn(self, conditions: Tensor) -> Tensor:
-        raise ValueError(
-            "This method does not use the posterior precision estimation."
-        )
+        raise ValueError("This method does not use the posterior precision estimation.")
 
     def marginal_denoising_posterior_precision_est_fn(
         self,
