### 2025-03-14 e878cd0
This might help testmon to not fail after first run
