+++ .github/workflows/ci.yml
@@ -57,6 +57,10 @@ jobs:
         restore-keys: |
           testmon-${{ runner.os }}-${{ matrix.python-version }}-
 
+    - name: Fix file permissions for testmondata
+      run: |
+        [ -f .testmondata ] && chmod u+w .testmondata || true
+
     - name: Run fast CPU tests with coverage
       run: uv run pytest --testmon-forceselect -v -n auto -m "not slow and not gpu" --cov=sbi --cov-report=xml tests/
 
