### 2025-03-14 f732971
Fix abstract class error (#1451)
* Fix

* Instead raise error

* Value error is better

* formatting
