### 2025-03-17 #7754 @francois-rozet
Hi @boisgera, there are no performance claims in the paper, so I checked off this item, as instructed.

> If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
