### 2025-03-17 #7754 @boisgera
> I guess this is a typo in the instruction, so I now checked the item. I would recommend to update the wording.

Thanks @francois-rozet! I believe that in English "check off" in this context means "mark as completed" (and the opposite would be "uncheck") and therefore that the actual wording is correct. (But as a French native speaker, this idiom also throws me off!).
