### 2025-03-17 b2dfc1a
remove old stuff and make things better
