### 2025-03-17 #7754 @francois-rozet
Yes, I guessed that "check off" meant uncheck, as other items use "check".
