### 2025-03-17 c8b1f2f
add author contributions in main text
