### 2025-03-17 #7754 @janfb
Hi @boisgera

as the formatting of the author list into multiple contributions categories was not possible I now added a section in the main text, just before the `Acknowledgments` section. It's pushed to the `joss-submission-2024` branch.

This would resolve this problem from our side.

Is there anything else we would need to do?

Thanks
Jan
