### 2025-03-17 f74511a
reuse npse_trained_model fixure for test_npse_map
