### 2025-03-17 f59a023
fix: sde_type and prior_type are now the part of the trained model fixture
