+++ docs/reference/sbi.models.rst
@@ -1,0 +1,10 @@
+Models
+======
+
+.. autofunction:: sbi.neural_nets.factory.posterior_nn
+.. autofunction:: sbi.neural_nets.factory.likelihood_nn
+.. autofunction:: sbi.neural_nets.factory.classifier_nn
+.. autofunction:: sbi.neural_nets.factory.flowmatching_nn
+.. autofunction:: sbi.neural_nets.factory.posterior_score_nn
+.. autoclass:: sbi.neural_nets.estimators.ConditionalDensityEstimator
+   :members:
