+++ docs/advanced_tutorials.rst
@@ -1,0 +1,53 @@
+.. _advanced_tutorials:
+
+
+Advanced tutorials
+==================
+
+.. toctree::
+   :maxdepth: 2
+
+Advanced
+--------
+
+.. toctree::
+   :maxdepth: 1
+
+   tutorials/02_multiround_inference.ipynb
+   tutorials/09_sampler_interface.ipynb
+   tutorials/03_density_estimators.ipynb
+   tutorials/04_embedding_networks.ipynb
+   tutorials/12_iid_data_and_permutation_invariant_embeddings.ipynb
+   tutorials/06_restriction_estimator.ipynb
+   tutorials/08_crafting_summary_statistics.ipynb
+   tutorials/15_importance_sampled_posteriors.ipynb
+
+Diagnostics
+-----------
+
+.. toctree::
+   :maxdepth: 1
+
+   tutorials/10_diagnostics_posterior_predictive_checks.ipynb
+   tutorials/11_diagnostics_simulation_based_calibration.ipynb
+   tutorials/13_diagnostics_lc2st.ipynb
+   tutorials/14_mcmc_diagnostics_with_arviz.ipynb
+
+Analysis
+--------
+
+.. toctree::
+   :maxdepth: 1
+
+   tutorials/05_conditional_distributions.ipynb
+   tutorials/07_sensitivity_analysis.ipynb
+   tutorials/17_plotting_functionality.ipynb
+
+Examples
+--------
+
+.. toctree::
+   :maxdepth: 1
+
+   tutorials/Example_00_HodgkinHuxleyModel.ipynb
+   tutorials/Example_01_DecisionMakingModel.ipynb
