+++ docs/contributing.md
@@ -1,0 +1,327 @@
+(contributing)=
+# How to contribute
+
+!!! important
+
+    By participating in the `sbi` community, all members are expected to comply with
+    our [Code of Conduct](code_of_conduct.md). This ensures a positive and inclusive
+    environment for everyone involved.
+
+## User experiences, bugs, and feature requests
+
+If you are using `sbi` to infer the parameters of a simulator, we would be
+delighted to know how it worked for you. If it didn't work according to plan,
+please open up an [issue](https://github.com/sbi-dev/sbi/issues) or
+[discussion](https://github.com/sbi-dev/sbi/discussions) and tell us more about
+your use case: the dimensionality of the input parameters and of the output,
+as well as the setup you used to run inference (i.e., number of simulations,
+number of rounds, etc.).
+
+To report bugs and suggest features -- including better documentation --
+please equally head over to [issues on GitHub](https://github.com/sbi-dev/sbi/issues)
+and tell us everything.
+
+## Contributing code
+
+Contributions to the `sbi` package are always welcome! The preferred way to do
+it is via pull requests onto our [main repository](https://github.com/sbi-dev/sbi).
+To give credit to contributors, we consider adding contributors who repeatedly
+and substantially contributed to `sbi` to the list of authors of the package at
+the end of every year. Additionally, we mention all contributors in the releases.
+
+!!! note
+    To avoid doing duplicated work, we strongly suggest that you go take
+    a look at our current [open issues](https://github.com/sbi-dev/sbi/issues) and
+    [pull requests](https://github.com/sbi-dev/sbi/pulls) to see if someone else is
+    already doing it. Also, in case you're planning to work on something that has not
+    yet been proposed by others (e.g. adding a new feature, adding a new example),
+    it is preferable to first open a new issue explaining what you intend to
+    propose and then working on your pull request after getting some feedback from
+    others.
+
+### Contribution workflow
+
+The following steps describe all parts of the workflow for doing a contribution
+such as installing locally `sbi` from source, creating a `conda` environment,
+setting up your `git` repository, etc. We've taken strong inspiration from the
+contribution guides of
+[`scikit-learn`](https://scikit-learn.org/stable/developers/contributing.html)
+and [`mne`](https://mne.tools/stable/development/contributing.html):
+
+**Step 1**: [Create an account](https://github.com/) on GitHub if you do not
+already have one.
+
+**Step 2**: Fork the [project repository](https://github.com/sbi-dev/sbi): click
+on the ‘Fork’ button near the top of the page. This will create a copy of the
+`sbi` codebase under your GitHub user account. See more details on how to fork
+a repository [here](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/fork-a-repo).
+
+**Step 3**: Clone your fork of the `sbi` repo from your GitHub account to your
+local disk:
+
+```bash
+git clone git@github.com:$USERNAME/sbi.git
+cd sbi
+```
+
+**Step 4**: Install a recent version of Python (we currently recommend 3.10)
+for instance using [`miniforge`](https://github.com/conda-forge/miniforge). We
+strongly recommend you create a specific `conda` environment for doing
+development on `sbi` as per:
+
+```bash
+conda create -n sbi_dev python=3.10
+conda activate sbi_dev
+```
+
+**Step 5**: Install `sbi` in editable mode with
+
+```bash
+pip install -e ".[dev]"
+```
+
+This installs the `sbi` package into the current environment by creating a
+link to the source code directory (instead of copying the code to pip’s `site_packages`
+directory, which is what normally happens). This means that any edits you make
[…]
