+++ pyproject.toml
@@ -60,6 +60,13 @@ doc = [
     "nbconvert",
     "nbformat",
     "traitlets <= 5.9.0",
+    "sphinx",
+    "sphinx-autobuild",
+    "sphinx_autodoc_typehints",
+    "sphinx-math-dollar",
+    "myst-nb",
+    "jupytext",
+    "sphinx-book-theme",
 ]
 dev = [
     "ffmpeg",
@@ -108,13 +115,13 @@ ignore = [
 [tool.ruff.lint.extend-per-file-ignores]
 "__init__.py" = ["E402", "F401", "F403"]  # allow unused imports and undefined names
 "test_*.py" = ["F403", "F405"]
-"tutorials/*.ipynb" = ["E501"]  # allow long lines in notebooks
+"docs/tutorials/*.ipynb" = ["E501"]  # allow long lines in notebooks
 
 [tool.ruff.lint.isort]
 case-sensitive = true
 lines-between-types = 0  # like isort default.
 relative-imports-order = "furthest-to-closest"  # like isort default.
-known-first-party = ["sbi", "tests", "tutorials"]
+known-first-party = ["sbi", "tests", "docs/tutorials"]
 
 [tool.ruff.format]
 exclude = ["*.ipynb"]
@@ -141,7 +148,7 @@ xfail_strict = true
 # Pyright configuration
 [tool.pyright]
 include = ["sbi"]
-exclude = ["**/__pycache__", "**/__node_modules__", ".git", "docs", "tutorials", "tests"]
+exclude = ["**/__pycache__", "**/__node_modules__", ".git", "docs", "mkdocs", "tests"]
 python_version = "3.10"
 reportUnsupportedDunderAll = false
 reportGeneralTypeIssues = false
