+++ docs/reference/sbi.inference.rst
@@ -1,0 +1,39 @@
+Inference
+=========
+
+.. autoclass:: sbi.inference.trainers.npe.npe_a.NPE_A
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.npe.npe_c.NPE_C
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.fmpe.fmpe.FMPE
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.npse.npse.NPSE
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.nle.nle_a.NLE_A
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.nre.nre_a.NRE_A
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.nre.nre_b.NRE_B
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.nre.nre_c.NRE_C
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.trainers.nre.bnre.BNRE
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.abc.mcabc.MCABC
+   :members:
+   :inherited-members:
+.. autoclass:: sbi.inference.abc.smcabc.SMCABC
+   :members:
+   :inherited-members:
+.. autofunction:: sbi.inference.trainers.base.simulate_for_sbi
+.. autofunction:: sbi.utils.user_input_checks.process_prior
+.. autofunction:: sbi.utils.user_input_checks.process_simulator
