### 2025-03-17 ac26ab6
shift VectorFieldNet type from estimators/flowmatching_estimator.py to utils/vector_field_utils.py
