### 2025-03-17 d90d0e7
shift VectorFieldNet type from estimators/flowmatching_estimator.py to utils/vector_field_utils.py
