+++ sbi/neural_nets/estimators/flowmatching_estimator.py
@@ -1,5 +1,4 @@
 import math
-from abc import ABC, abstractmethod
 from typing import Optional, Tuple
 
 import torch
@@ -12,12 +11,7 @@ from zuko.transforms import FreeFormJacobianTransform
 from zuko.utils import broadcast
 
 from sbi.neural_nets.estimators.base import ConditionalDensityEstimator
-
-
-# abstract class to ensure forward signature for flow matching networks
-class VectorFieldNet(nn.Module, ABC):
-    @abstractmethod
-    def forward(self, theta: Tensor, x: Tensor, t: Tensor) -> Tensor: ...
+from sbi.utils.vector_field_utils import VectorFieldNet
 
 
 class FlowMatchingEstimator(ConditionalDensityEstimator):
