### 2025-03-17 8555fc6
shift VectorFieldNet type from estimators/flowmatching_estimator.py to utils/vector_field_utils.py
