### 2025-03-17 92b79a2
shift VectorFieldNet type from estimators/flowmatching_estimator.py to utils/vector_field_utils.py
