### 2025-03-17 e503c88
logit transform changes
