### 2025-03-17 eee9f67
wip: speed batched sampling test
