+++ tests/density_estimator_test.py
@@ -243,8 +243,10 @@ def test_correctness_of_density_estimator_log_prob(
 @pytest.mark.parametrize(
     "density_estimator_build_fn", model_builders + flowmatching_build_functions
 )
-@pytest.mark.parametrize("input_event_shape", ((1,), (4,)))
-@pytest.mark.parametrize("condition_event_shape", ((1,), (7,)))
+@pytest.mark.parametrize(
+    "input_event_shape", ((1,), pytest.param((2,), marks=pytest.mark.slow))
+)
+@pytest.mark.parametrize("condition_event_shape", ((1,), (3, 3)))
 @pytest.mark.parametrize("sample_shape", ((1000,), (500, 2)))
 def test_correctness_of_batched_vs_seperate_sample_and_log_prob(
     density_estimator_build_fn: Callable,
@@ -267,11 +269,17 @@ def test_correctness_of_batched_vs_seperate_sample_and_log_prob(
     samples = density_estimator.sample(sample_shape, condition=condition)
     samples = samples.reshape(-1, batch_dim, *input_event_shape)  # Flat for comp.
 
+    # Flatten sample_shape to (B*E,) if it is (B, E)
+    if len(sample_shape) > 1:
+        flat_sample_shape = (torch.prod(torch.tensor(sample_shape)).item(),)
+    else:
+        flat_sample_shape = sample_shape
+
     samples_separate1 = density_estimator.sample(
-        (1000,), condition=condition[0][None, ...]
+        flat_sample_shape, condition=condition[0][None, ...]
     )
     samples_separate2 = density_estimator.sample(
-        (1000,), condition=condition[1][None, ...]
+        flat_sample_shape, condition=condition[1][None, ...]
     )
 
     # Check if means are approx. same
@@ -310,12 +318,14 @@ def _build_density_estimator_and_tensors(
     """Helper function for all tests that deal with shapes of density
     estimators."""
 
+    batch_size = 1000
     # Use positive random values for continuous dims (log transform)
-    batch_input = torch.rand((1000, *input_event_shape), dtype=torch.float32) * 10.0
+    batch_input = (
+        torch.rand((batch_size, *input_event_shape), dtype=torch.float32) * 10.0
+    )
     # make last dim discrete for mixed density estimators
-    batch_input[:, -1] = torch.randint(0, 4, (1000,))
-    batch_condition = torch.randn((1000, *condition_event_shape))
-
+    batch_input[:, -1] = torch.randint(0, 4, (batch_size,))
+    batch_condition = torch.randn((batch_size, *condition_event_shape))
     if len(condition_event_shape) > 1:
         embedding_net = CNNEmbedding(condition_event_shape, kernel_size=1)
         z_score_y = "structured"
