### 2025-03-17 #7754 @boisgera
> Hi [@boisgera](https://github.com/boisgera)
>
> as the formatting of the author list into multiple contributions categories was not possible I now added a section in the main text, just before the `Acknowledgments` section. It's pushed to the `joss-submission-2024` branch.
>
> This would resolve this problem from our side.

👍

> Is there anything else we would need to do?

The review step is over now ; thanks a lot to @arnauqb and @francois-rozet for their very helpful reviews and to you @janfb for a great submission and your reactivity during this review process! 🙏

We both (you as the author, me as the editor) still have a few mundane tasks to do in the post-review step ([checklist here](https://github.com/openjournals/joss-reviews/issues/7754#issuecomment-2730792331)) before I can recommend the paper for acceptance to the editor in chief.
