### 2025-03-18 7b5d57e
chore: add pytest-split plugin to dev dependencies, use split in ci workflow based on test durations in .test_durations
