+++ pyproject.toml
@@ -74,6 +74,7 @@ dev = [
     "pytest-testmon",
     "pytest-xdist",
     "pytest-harvest",
+    "pytest-split",
     "torchtestcase",
 ]
 
