### 2025-03-18 935323c
Add ResNet embedding model
