+++ sbi/neural_nets/embedding_nets/resnet.py
@@ -1,0 +1,312 @@
+from functools import partial
+from typing import Callable, Type
+
+import torch
+import torch.nn as nn
+
+
+def zero_pad(x: torch.tensor, c_out: int) -> torch.tensor:
+    """
+    Add additional channels to the input tensor by filling it with zeros.
+
+    Args:
+        x:      Input tensor
+        c_out:  Number of output channels
+
+    Returns:
+        x_aug:      Input tensor with additional channels filled with zeros
+    """
+
+    # Check if it is a image like iput
+    assert len(x.shape) == 4
+
+    # Check if the number of channels is increased
+    assert c_out >= x.shape[1]
+
+    c_in = x.shape[1]
+    a = torch.zeros(x.shape[0], c_out - c_in, x.shape[2], x.shape[3], device=x.device)
+
+    x_aug = torch.cat([x, a], dim=1)
+
+    return x_aug
+
+
+def construct_simple_conv_net(
+    c_in: int,
+    c_out: int,
+    c_hidden: int = 32,
+    activation: Type[nn.Module] = nn.ReLU,
+    downsample: int = False,
+) -> nn.Module:
+    """
+    Default implementation for a simple convolutional network with 3 layers.
+
+    Args:
+        c_in:       Number of input channels
+        c_out:      Number of output channels
+        c_hidden:   Number of hidden channels
+        activation: Constructor for a activation function
+        downsample: Apply a strided convolution to halve
+                    the height and width of the input
+
+    Returns:
+        layers:     A sequential container with the convolutional layers
+    """
+
+    # Define the convolutional layers
+    stride = 2 if downsample else 1
+    layers = nn.Sequential(
+        nn.Conv2d(c_in, c_hidden, kernel_size=3, stride=stride, padding=1),
+        activation(),
+        nn.Conv2d(c_hidden, c_hidden, kernel_size=3, stride=1, padding=1),
+        activation(),
+        nn.Conv2d(c_hidden, c_out, kernel_size=3, stride=1, padding=1),
+    )
+
+    return layers
+
+
+class ResidualBLockConv(nn.Module):
+    def __init__(
+        self,
+        c_in: int,
+        c_out: int | None = None,
+        change_c_mode: str = "conv",
+        downsample: bool = False,
+        construct_mapping: Callable = construct_simple_conv_net,
+        construct_mapping_kwargs: dict | None = None,
+        activation: Type[nn.Module] = nn.ReLU,
+    ) -> None:
+        """
+        Single residual block for image like data.
+
+        Args:
+            c_in: Number of input channels
+            c_out: Number of output channels. If None, use c_in.
+            change_c_mode: Mode to change the number of channels.
+                           Options are "conv" and "zeros". If "conv" is selected,
+                           1x1 convolutions are applied to change the number
+                           of channels. If "zeros" is selected, the required additional
+                           channels are filled with zeros.
+            downsample: Apply a strided convolution to halve the height and
+                        width of the input.
+            construct_mapping: Constructor for the mapping function. The mapping
+                               function is applied to the input tensor before the
+                               residual connection is added.
+            construct_mapping_kwargs: Additional keyword arguments for the mapping
+                                      function.
+            activation: Constructor for the activation function.
+        """
+
+        super(ResidualBLockConv, self).__init__()
+
+        if construct_mapping_kwargs is None:
+            construct_mapping_kwargs = {}
+
+        ###############################################################################
+        # Change number of channels
+        ###############################################################################
+
+        # Preserve the input dimensionality if the output is not specified
+        if c_out is None:
+            c_out = c_in
+
+        # Preserv the dimensionality of the input and the output
+        if c_in == c_out:
+            self.residual = nn.Identity()
+
[…]
