### 2025-03-18 6412964
bug fix with normalization when using embedding nets
