+++ sbi/neural_nets/embedding_nets/__init__.py
@@ -3,3 +3,6 @@ from sbi.neural_nets.embedding_nets.fully_connected import FCEmbedding
 from sbi.neural_nets.embedding_nets.permutation_invariant import (
     PermutationInvariantEmbedding,
 )
+from sbi.neural_nets.embedding_nets.resnet import (
+    ResNet,
+)
