### 2025-03-18 0be2365
init LRU implementation based on Orvieto et al
