+++ sbi/neural_nets/embedding_nets/lru.py
@@ -3,14 +3,11 @@ from typing import Callable
 import lru
 import torch
 from torch import Tensor, nn
+import numpy as np
 
 
 class LRUEmbedding(nn.Module):
     """Embedding network backed by a Linear Recurrent Unit (LRU).
-
-    As suggested in https://github.com/forgi86/sysid-pytorch-lru, we use GLU
-    activation function after each layer.
-
     See also:
         https://arxiv.org/pdf/2303.06349
     """
@@ -19,9 +16,15 @@ class LRUEmbedding(nn.Module):
         self,
         input_dim: int,
         output_dim: int = 10,
+        state_dim: int = 20,
         hidden_dim: int = 20,
+        mlp_dim: int = 80,
         num_layers: int = 2,
-        aggregation_fn: Callable[[Tensor], Tensor] = torch.mean,
+        r_min=0.0,
+        r_max=1.0,
+        max_phase=2 * np.pi,
+        bidirectional: bool = False,
+        dropout: float = 0.0,
     ):
         """Fully-connected multi-layer neural network to be used as embedding network.
 
@@ -33,43 +36,29 @@ class LRUEmbedding(nn.Module):
         """
         super().__init__()
 
-        assert aggregation_fn in (torch.mean, torch.sum), (
-            "Only torch.mean and torch.sum are supported as aggregation_fn"
-        )
-        self.aggregation_fn = aggregation_fn
-
         # The first layer is defined by the observations' input dimension.
+        self.embedding = nn.Linear(input_dim, hidden_dim)
+
         layers = [
-            lru.linear.LRU(
-                in_features=input_dim,
-                out_features=hidden_dim,
-                state_features=hidden_dim,
-            ),
-            nn.GLU(),
+            LRUBlock(
+                input_dim=hidden_dim,
+                output_dim=hidden_dim,
+                state_dim=hidden_dim,
+                mlp_dim=mlp_dim,
+                r_min=r_min,
+                r_max=r_max,
+                max_phase=max_phase,
+                bidirectional=bidirectional,
+                dropout=dropout,
+            )
+            for _ in range(num_layers)
         ]
 
-        # The number of hidden layers is num_layers - 2.
-        for _ in range(num_layers - 2):
-            layers.append(
-                lru.linear.LRU(
-                    in_features=hidden_dim,
-                    out_features=hidden_dim,
-                    state_features=hidden_dim,
-                )
-            )
-            layers.append(nn.GLU())
+        # Output layer.
+        # TODO: only readout one timestep
+        self.output = nn.Linear(hidden_dim, output_dim)
 
-        # The layer is defined by the observations' output dimension.
-        layers.append(
-            lru.linear.LRU(
-                in_features=hidden_dim,
-                out_features=output_dim,
-                state_features=hidden_dim,
-            )
-        )
-        layers.append(nn.GLU())
-
-        self.net = nn.Sequential(*layers)
+        self.layers = nn.Sequential(*layers)
 
     def forward(self, x: Tensor) -> Tensor:
         """Embed a batch of 2-dim observations, e.g. a multi-dimensional
@@ -81,9 +70,104 @@ class LRUEmbedding(nn.Module):
         Returns:
             Network output (batch_size, output_dim).
         """
+        x = self.embedding(x)
         x_scan = self.net(x)  # (batch_size, len_sequence, output_dim)
 
         # Aggregate the features of the time dimension.
[…]
