### 2025-03-18 48f0e63
revert unnecessary gpu handling things. Now MultipleIndependent does not allow gpu handling yet though
