+++ sbi/utils/user_input_checks_utils.py
@@ -8,8 +8,6 @@ import torch
 from torch import Tensor, float32
 from torch.distributions import Distribution, constraints
 
-from sbi.utils.torchutils import BoxUniform, process_device
-
 
 class CustomPriorWrapper(Distribution):
     def __init__(
@@ -175,21 +173,18 @@ class MultipleIndependent(Distribution):
         arg_constraints: Optional[Dict[str, constraints.Constraint]] = None,
         device: Optional[str] = None,
     ):
+        if device is not None:
+            raise NotImplementedError("device is not supported yet")
+
         self._check_distributions(dists)
         if validate_args is not None:
             [d.set_default_validate_args(validate_args) for d in dists]
 
-        # Device handling
-        device = dists[0].sample().device.type if device is None else device
-        device = process_device(device)
-
-        # Move all distributions to the specified device
-        # self.dists = dists
-        self.dists = [self._move_dist_to_device(d, device) for d in dists]
-
+        self.dists = dists
         # numel() instead of event_shape because for all dists both is possible,
         # event_shape=[1] or batch_shape=[1]
         self.dims_per_dist = [d.sample().numel() for d in self.dists]
+
         self.ndims = int(torch.sum(torch.as_tensor(self.dims_per_dist)).item())
         self.custom_arg_constraints = arg_constraints or {}
         self.validate_args = validate_args
@@ -201,103 +196,6 @@ class MultipleIndependent(Distribution):
             ]),  # Event shape is the sum of all ndims.
             validate_args=validate_args,
         )
-
-    def _move_dist_to_device(self, dist: Distribution, device: str) -> Distribution:
-        """
-        Helper to move a distribution's parameters to the specified device.
-        This is quite ugly but pytorch does not has a simple way to move parameters
-        of all distributions to a device.
-        """
-        try:
-            # Handle Independent distributions (including BoxUniform)
-            if isinstance(dist, torch.distributions.Independent):
-                # Handle other Independent distributions
-                if isinstance(dist, BoxUniform):
-                    return BoxUniform(
-                        low=dist.base_dist.low,
-                        high=dist.base_dist.high,
-                        reinterpreted_batch_ndims=dist.reinterpreted_batch_ndims,
-                        device=device,
-                    )
-                else:
-                    return self._move_dist_to_device(dist.base_dist, device)
-
-            # Try to get parameters through the Distribution interface
-            params = {}
-            # Map of distribution types to their required parameter names
-            dist_param_map = {
-                'Beta': ['concentration0', 'concentration1'],
-                'Bernoulli': ['probs'],
-                'Binomial': ['total_count', 'probs'],
-                'Categorical': ['probs'],
-                'Cauchy': ['loc', 'scale'],
-                'Chi2': ['df'],
-                'Dirichlet': ['concentration'],
-                'Exponential': ['rate'],
-                'Gamma': ['concentration', 'rate'],
-                'Geometric': ['probs'],
-                'Gumbel': ['loc', 'scale'],
-                'HalfNormal': ['scale'],
-                'Laplace': ['loc', 'scale'],
-                'LogNormal': ['loc', 'scale'],
-                'Multinomial': ['total_count', 'probs'],
-                'MultivariateNormal': ['loc', 'covariance_matrix'],
-                'Normal': ['loc', 'scale'],
-                'Pareto': ['scale', 'alpha'],
-                'Poisson': ['rate'],
-                'StudentT': ['df', 'loc', 'scale'],
-                'Uniform': ['low', 'high'],
-                'VonMises': ['loc', 'concentration'],
-                'Weibull': ['scale', 'concentration'],
-            }
-
-            # Get parameter names specific to this distribution type
-            dist_type = type(dist).__name__
[…]
