### 2025-03-18 7b43583
Moving and renaming. Refactoring IID function to reduce code redundancy
