+++ sbi/inference/potentials/score_fn_util.py
@@ -18,6 +18,7 @@ from sbi.utils.score_utils import (
 from sbi.utils.torchutils import ensure_theta_batched
 
 IID_METHODS = {}
+GUIDANCE_METHODS = {}
 
 
 def get_iid_method(name: str) -> Type["IIDScoreFunction"]:
@@ -36,6 +37,39 @@ def get_iid_method(name: str) -> Type["IIDScoreFunction"]:
         )
     return IID_METHODS[name]
 
+def get_guidance_method(name: str) -> Type["ScoreAdaptation"]:
+    r"""
+    Retrieves the guidance method by name.
+
+    Args:
+        name: The name of the guidance method.
+
+    Returns:
+        The guidance method class.
+    """
+    if name not in GUIDANCE_METHODS:
+        raise NotImplementedError(
+            f"Method {name} for guidance not implemented."
+        )
+    return GUIDANCE_METHODS[name]
+
+def register_guidance_method(name: str) -> Callable:
+    r"""
+    Registers a guidance method.
+
+    Args:
+        name: The name of the guidance method.
+
+    Returns:
+        A decorator function to register the guidance method class.
+    """
+
+    def decorator(cls: Type["ScoreAdaptation"]) -> Type["ScoreAdaptation"]:
+        GUIDANCE_METHODS[name] = cls
+        return cls
+
+    return decorator
+
 
 def register_iid_method(name: str) -> Callable:
     r"""
@@ -53,6 +87,90 @@ def register_iid_method(name: str) -> Callable:
         return cls
 
     return decorator
+
+
+class ScoreAdaptation(ABC):
+    def __init__(
+        self,
+        score_estimator: ConditionalScoreEstimator,
+        prior: Optional[Distribution],
+        device: str = "cpu",
+    ):
+        """ This class manages manipulating the score estimator to impose additional
+        constraints on the posterior via guidance.
+
+        Args:
+            score_estimator: The score estimator.
+            prior: The prior distribution.
+            device: The device on which to evaluate the potential.
+        """
+        self.score_estimator = score_estimator
+        self.prior = prior
+        self.device = device
+
+    @abstractmethod
+    def __call__(self, theta: Tensor, x_o: Tensor, time: Optional[Tensor] = None):
+        pass
+
+@register_guidance_method("classifier_free")
+class ClassifierFreeGuidance(ScoreAdaptation):
+    def __init__(
+        self,
+        score_estimator: ConditionalScoreEstimator,
+        prior: Optional[Distribution],
+        prior_scale: float | Tensor,
+        prior_shift: float | Tensor,
+        likelihood_scale: float | Tensor,
+        likelihood_shift: float | Tensor,
+        device: str = "cpu",
+    ):
+        """ This class manages manipulating the score estimator to temper of shift the
+        prior and likelihood.
+
+        This is usually known as classifier-free guidance. And works by decomposing the
+        posterior score into a prior and likelihood component. These can then be scaled
+        and shifted to impose additional constraints on the posterior.
+
+        Args:
+            score_estimator: The score estimator.
+            prior: The prior distribution.
+            device: The device on which to evaluate the potential.
+        """
+
+        if prior is None:
+            raise ValueError("Prior is required for classifier-free guidance, please"
+                            " provide as least an improper empirical prior.")
+
+        self.prior_scale = prior_scale
+        self.prior_shift = prior_shift
+        self.likelihood_scale = likelihood_scale
+        self.likelihood_shift = likelihood_shift
+
+        super().__init__(score_estimator, prior, device)
+
+    def marginal_prior_score(self, theta: Tensor,  time: Tensor):
+        """ Computes the marginal prior score analyticaly (or approximatly)
+        """
[…]
