### 2025-03-18 ab6b7b4
Moving and renaming. Refactoring IID function to reduce code redundancy
