### 2025-03-18 5fdb087
enh: mark default as NOT slow
