+++ sbi/neural_nets/estimators/__init__.py
@@ -1,4 +1,7 @@
-from sbi.neural_nets.estimators.base import ConditionalDensityEstimator
+from sbi.neural_nets.estimators.base import (
+    ConditionalDensityEstimator,
+    UnconditionalDensityEstimator,
+)
 from sbi.neural_nets.estimators.categorical_net import (
     CategoricalMADE,
     CategoricalMassEstimator,
@@ -7,4 +10,4 @@ from sbi.neural_nets.estimators.flowmatching_estimator import FlowMatchingEstima
 from sbi.neural_nets.estimators.mixed_density_estimator import MixedDensityEstimator
 from sbi.neural_nets.estimators.nflows_flow import NFlowsFlow
 from sbi.neural_nets.estimators.score_estimator import ConditionalScoreEstimator
-from sbi.neural_nets.estimators.zuko_flow import ZukoFlow
+from sbi.neural_nets.estimators.zuko_flow import ZukoFlow, ZukoUnconditionalFlow
