+++ sbi/neural_nets/estimators/base.py
@@ -117,7 +117,7 @@ class ConditionalEstimator(nn.Module, ABC):
 
 
 class ConditionalDensityEstimator(ConditionalEstimator):
-    r"""Base class for density estimators.
+    r"""Base class for conditional density estimators.
 
     The density estimator class is a wrapper around neural networks that allows to
     evaluate the `log_prob`, `sample`, and provide the `loss` of $\theta,x$ pairs. Here
@@ -126,7 +126,7 @@ class ConditionalDensityEstimator(ConditionalEstimator):
     Note:
         We assume that the input to the density estimator is a tensor of shape
         (sample_dim, batch_dim, *input_shape), where input_shape is the dimensionality
-        of the input. The condition is a tensor of shape (batch_size, *condition_shape),
+        of the input. The condition is a tensor of shape (batch_dim, *condition_shape),
         where condition_shape is the shape of the condition tensor.
 
     """
@@ -224,11 +224,12 @@ class ConditionalDensityEstimator(ConditionalEstimator):
 
 
 class ConditionalVectorFieldEstimator(ConditionalEstimator):
-    r"""Base class for vector field (e.g., score and ODE flow) estimators.
+    r"""Base class for conditional vector field (e.g., score and ODE flow) estimators.
 
-    The vector field estimator class is a wrapper around neural networks that allows to
-    evaluate the `vector_field`, and provide the `loss` of $\theta,x$ pairs. Here
-    $\theta$ would be the `input` and $x$ would be the `condition`.
+    The conditional vector field estimator class is a wrapper around neural
+    networks that allows to evaluate the `vector_field`, and provide the `loss`
+    of $\theta,x$ pairs. Here $\theta$ would be the `input` and $x$ would be
+    the `condition`.
 
     Note:
         We assume that the input to the density estimator is a tensor of shape
@@ -262,3 +263,162 @@ class ConditionalVectorFieldEstimator(ConditionalEstimator):
         Raises:
             NotImplementedError: This method should be implemented by sub-classes.
         """
+
+
+class UnconditionalEstimator(nn.Module, ABC):
+    r"""Base class for unconditional estimators that estimate properties of
+    distributions without conditioning on an input.
+
+    For example, this can be:
+    - A density estimator of the pdf $p(x)$.
+    - A vector field estimator e.g. $\nabla_x \log p(x)$.
+
+    Subclasses of UnconditionalEstimator should implement the ``loss(input)``
+    method to be compatible with sbi's training procedures.
+    """
+
+    def __init__(self, input_shape: Tuple) -> None:
+        r"""Construct an unconditional estimator given shapes.
+
+        Args:
+            input_shape: Event shape of the input at which the density is being
+                evaluated (and which is also the event_shape of samples).
+        """
+        super().__init__()
+        self._input_shape = torch.Size(input_shape)
+
+    @property
+    def input_shape(self) -> torch.Size:
+        r"""Return the input shape."""
+        return self._input_shape
+
+    @abstractmethod
+    def loss(self, input: Tensor, **kwargs) -> Tensor:
+        r"""Return the loss for training the estimator.
+
+        Args:
+            input: Inputs to evaluate the loss on of shape
+                `(batch_dim, *input_event_shape)`.
+
+        Returns:
+            Loss of shape (batch_dim,)
+        """
+        pass
+
+    def _check_input_shape(self, input: Tensor):
+        r"""This method checks whether the input has the correct shape.
+
+        Args:
+            input: Inputs to evaluate the log probability on of shape
+                    `(sample_dim_input, batch_dim_input, *event_shape_input)`.
+
+        Raises:
+            ValueError: If the input has a dimensionality that does not match
+                        the expected input dimensionality.
+            ValueError: If the shape of the input does not match the expected
+                        input dimensionality.
[…]
