### 2025-03-18 c014fdf
implementation snpe-b method + adaptation of snpe tests
