### 2025-03-18 227ca03
Bug fix for one channel input
