### 2025-03-18 b357760
Add test for ResNet embedding model for image like data
