### 2025-03-18 c41c8bc
modify readme to add snpe-b
