### 2025-03-18 aea9a1e
Fix gloeckler2023adversarial bib entry DOI
