+++ paper/paper.bib
@@ -940,7 +940,7 @@ doi = {10.48550/arXiv.2305.17161},
   pages={11493--11524},
   year={2023},
   organization={PMLR},
-  doi = {10.48550/arVix.2305.14984}
+  doi = {10.48550/arXiv.2305.14984}
 }
 
 @inproceedings{
