### 2025-03-18 cf7df9a
modify docs to add (S)NPE-B
