### 2025-03-18 a396040
chore: pytest split for ci workflow  (#1465)
- add pytest-split plugin to dev dependencies
- use split in ci workflow based on test durations in .test_durations
