### 2025-03-18 sbi-dev/sbi#1476 @boisgera
Duplicate Bibliography Item and Missing DOI
## 🐛 Bug Description

There is a duplicate item in the [JOSS submission bibliography](https://github.com/sbi-dev/sbi/blob/joss-submission-2024/paper/paper.bib) that triggers the "missing DOI" JOSS warning message.

## 🔄 Steps to Reproduce

```
@article{BeaumontZhang_02,
	Author = {Beaumont, M and Zhang, W and Balding, D J},
	Journal = {Genetics},
	Number = {4},
	Title = {Approximate Bayesian computation in population genetics},
	Volume = {162},
	Year = {2002}}

@article{beaumont2002approximate,
  title={Approximate Bayesian computation in population genetics},
  author={Beaumont, Mark A and Zhang, Wenyang and Balding, David J},
  journal={Genetics},
  volume={162},
  number={4},
  pages={2025--2035},
  year={2002},
  publisher={Oxford University Press},
  doi = {10.1093/genetics/162.4.2025},
}
```

## ✅ Expected Behavior

Have a single entry (with DOI) for this bibliography item (and corresponding citation in `paper.md`).
