### 2025-03-18 sbi-dev/sbi#1476 @janfb
thanks for spotting this @boisgera!

I will update the bib accordingly.
