### 2025-03-18 #7754 @boisgera
Hi @janfb,

Please consider the items

  - The issue https://github.com/sbi-dev/sbi/issues/1476
  - The PR https://github.com/sbi-dev/sbi/pull/1474

that should address the following DOI-related warnings

```
❌ MISSING DOIs

- 10.1093/genetics/162.4.2025 may be a valid DOI for title: Approximate Bayesian computation in population gen...

❌ INVALID DOIs

- 10.48550/arVix.2305.14984 is INVALID
```
