### 2025-03-18 b52b94d
fix #1476: remove duplicate bib entry
