+++ paper/paper.bib
@@ -475,14 +475,6 @@ doi = {10.48550/arXiv.2203.04176}
 	Title = {Adaptive approximate Bayesian computation},
 	Year = {2009}}
 
-@article{BeaumontZhang_02,
-	Author = {Beaumont, M and Zhang, W and Balding, D J},
-	Journal = {Genetics},
-	Number = {4},
-	Title = {Approximate Bayesian computation in population genetics},
-	Volume = {162},
-	Year = {2002}}
-
 @article{BlumFrancois_10,
 	Author = {Blum, Michael GB and Fran{\c{c}}ois, Olivier},
 	Journal = {Statistics and Computing},
