### 2025-03-18 6f891da
Prevent notebook execution upon doc build
