### 2025-03-18 c514c8b
Fix gloeckler2023adversarial bib entry DOI (#1474)
