### 2025-03-18 #7754 @janfb
Thanks @boisgera !

Both have been merged into `joss-submission-2024`
