+++ sbi/neural_nets/embedding_nets/lru.py
@@ -5,6 +5,8 @@ import torch
 from torch import Tensor, nn
 import numpy as np
 
+from torch.utils._pytree import tree_flatten, tree_unflatten
+
 
 class LRUEmbedding(nn.Module):
     """Embedding network backed by a Linear Recurrent Unit (LRU).
@@ -18,13 +20,14 @@ class LRUEmbedding(nn.Module):
         output_dim: int = 10,
         state_dim: int = 20,
         hidden_dim: int = 20,
-        mlp_dim: int = 80,
         num_layers: int = 2,
         r_min=0.0,
         r_max=1.0,
         max_phase=2 * np.pi,
         bidirectional: bool = False,
         dropout: float = 0.0,
+        norm: bool = False,
+        #aggretate_func : Callable,
     ):
         """Fully-connected multi-layer neural network to be used as embedding network.
 
@@ -41,10 +44,8 @@ class LRUEmbedding(nn.Module):
 
         layers = [
             LRUBlock(
-                input_dim=hidden_dim,
-                output_dim=hidden_dim,
+                hidden_dim=hidden_dim,
                 state_dim=state_dim,
-                mlp_dim=mlp_dim,
                 r_min=r_min,
                 r_max=r_max,
                 max_phase=max_phase,
@@ -73,42 +74,54 @@ class LRUEmbedding(nn.Module):
         x = self.embedding(x)
         x_scan = self.net(x)  # (batch_size, len_sequence, output_dim)
 
-        # Aggregate the features of the time dimension.
-        x_embed = self.output(x_scan)
 
+
+        # Pooling
+        # TODO: implement pooling
+
+        # output embedding
+        x_embed = self.output(x_scan)
         return x_embed
 
 
 class LRUBlock(nn.Module):
     def __init__(
         self,
-        input_dim: int,
-        output_dim: int = 10,
+        hidden_dim: int,
         state_dim: int = 20,
-        mlp_dim: int = 20,
         r_min=0.0,
         r_max=1.0,
         max_phase=2 * np.pi,
         bidirectional: bool = False,
         dropout: float = 0.0,
+        norm: Bool=False,
     ):
         super().__init__()
-        self.lru = LRU(input_dim, state_dim, r_min, r_max, max_phase)
-        self.hidden_projection = nn.Linear(state_dim, mlp_dim)
-        self.hidden_nonlinearity = nn.GELU()
-        self.output_projection = nn.Linear(mlp_dim, output_dim)
+        #https://github.com/NicolasZucchet/minimal-LRU/blob/main/lru/model.py
+        self.lru = LRU(hidden_dim, state_dim, r_min, r_max, max_phase)
+        self.out1 = nn.Linear(hidden_dim,hidden_dim)
+        self.out2 = nn.Linear(hidden_dim, hidden_dim)
+        self.GELU = nn.GELU()
+        self.sigmoid = nn.Sigmoid()
         self.dropout = nn.Dropout(dropout) if dropout > 0 else nn.Identity()
+        self.norm = nn.LayerNorm(hidden_dim) if norm else nn.Identity()
 
     def forward(self, x: Tensor) -> Tensor:
-        # TO DO: GLU mixing Layers?
+        # Batch, len_sequence, hidden_dim
+        # normalization
+        y = self.norm(x)
 
         # run LRU
-        x = self.lru(x)
+        y=self.lru(y)
 
         # state mixing
-        y = self.hidden_nonlinearity(self.hidden_projection(x))
+        y = self.GELU(y)
         y = self.dropout(y)
-        y = self.output_projection(y) + x
+        y = self.out1(y)*self.sigmoid(self.out2(y))
+        y = self.dropout(y) 
+
+        # skip connection
+        y = y + x
[…]
