### 2025-03-18 133c9f5
attention and rotary encoding
