+++ sbi/neural_nets/embedding_nets/transformer.py
@@ -1,0 +1,181 @@
+import math
+from typing import Optional, Tuple
+
+import torch
+from torch import nn
+
+
+class PositionalEncoder(nn.Module):
+    def __init__(self, embedding_dim: int, power: Optional[float] = 10e4):
+        super().__init__()
+        self.power = power
+        self.embedding_dim = embedding_dim
+        div_term = self.power ** (torch.arange(0, embedding_dim, 2) / embedding_dim)
+        self.register_buffer("div_term", tensor=div_term, persistent=False)
+
+    def forward(self, x, position_ids: Optional[torch.tensor] = None):
+        seq_length = x.shape[-2]
+        if position_ids is None:
+            position_ids = torch.arange(0, seq_length, 1).to(x)
+
+        div_term = position_ids.view(-1, 1) / self.div_term.to(x)
+
+        pe = torch.zeros_like(x)
+        pe[..., 0::2] += torch.cos(div_term)
+        pe[..., 1::2] += torch.sin(div_term)
+
+        return x + pe
+
+
+class RotaryEncoder(nn.Module):
+    def __init__(self, embedding_dim: int, power: Optional[float] = 10e4):
+        super().__init__()
+        self.power = power
+        self.embedding_dim = embedding_dim
+        div_term = self.power ** (torch.arange(0, embedding_dim, 2) / embedding_dim)
+        self.register_buffer("div_term", tensor=div_term, persistent=False)
+
+    def rotate_half(self, x):
+        """Rotates half the hidden dims of the input."""
+        x1 = x[..., : x.shape[-1] // 2]
+        x2 = x[..., x.shape[-1] // 2 :]
+        return torch.cat((-x2, x1), dim=-1)
+
+    def forward(self, x, position_ids: Optional[torch.tensor] = None) -> torch.Tensor:
+        """Applies Rotary Position Embedding to the query and key tensors.
+
+        Args:
+            x (`torch.Tensor`): query/key tensor
+            position_ids (`torch.Tensor`, *optional*):
+                specify the position ids
+        Returns:
+            `(torch.Tensor)` comprising the query/key tensors rotated using the
+            Rotary Position Embedding.
+        """
+        seq_length = x.shape[-2]
+        rotary_dim = self.embedding_dim // 2
+
+        if position_ids is None:
+            position_ids = torch.arange(0, seq_length, 1).to(x)
+
+        freqs = position_ids.view(-1, 1) / self.div_term.to(x)
+
+        x_rot, x_pass = x[..., :rotary_dim], x[..., rotary_dim:]
+
+        x_embed = torch.cat(
+            [(x_rot * freqs.cos()) + (self.rotate_half(x_rot) * freqs.sin()), x_pass],
+            dim=-1,
+        )
+
+        return x_embed
+
+
+POSITION_EMB = {"positional": PositionalEncoder, "rotary": RotaryEncoder}
+
+
+class FullAttention(nn.Module):
+    # Adapted from https://github.com/huggingface/transformers/main/src/transformers/models/phi3/modeling_phi3.py
+    """Multi-headed attention from 'Attention Is All You Need' paper"""
+
+    def __init__(self, config, layer_idx: Optional[int] = None):
+        super().__init__()
+        self.config = config  # TODO implement config?
+        self.layer_idx = layer_idx
+        self.hidden_size = config["hidden_size"]
+        self.head_dim = config.get(
+            "head_dim", config["hidden_size"] // config["num_attention_heads"]
+        )
+        self.num_heads = config["num_attention_heads"]
+        self.num_key_value_groups = (
+            config["num_attention_heads"] // config["num_key_value_heads"]
+        )
+        self.num_key_value_heads = config["num_key_value_heads"]
+        self.scaling = self.head_dim**-0.5
[…]
