### 2025-03-18 dae3e68
standard rope
