### 2025-03-18 52825d3
basic api on potentials
