### 2025-03-18 abc6742
refactoring noise_schedule and time schedule into base class
- created noise_schedule method to be overwritten by derivatives
- created times_schedules method to be overwritten by derivatives
- created test on times_schedules
- improvded docstrings
