+++ sbi/neural_nets/estimators/score_estimator.py
@@ -6,9 +6,8 @@ from math import pi
 from typing import Callable, Optional, Union
 
 import torch
-from torch import Tensor, nn
-
 from sbi.neural_nets.estimators.base import ConditionalVectorFieldEstimator
+from torch import Tensor, nn
 
 
 class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
@@ -42,10 +41,12 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         input_shape: torch.Size,
         condition_shape: torch.Size,
         weight_fn: Union[str, Callable] = "max_likelihood",
+        beta_min: float = 0.01,
+        beta_max: float = 10.0,
         mean_0: Union[Tensor, float] = 0.0,
         std_0: Union[Tensor, float] = 1.0,
         t_min: float = 1e-3,
-        t_max: float = 1.0,
+        t_max: float = 1.0
     ) -> None:
         r"""Score estimator class that estimates the conditional score function, i.e.,
         gradient of the density p(xt|x0).
@@ -59,6 +60,12 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
                 - "identity": constant weights (1.),
                 - "max_likelihood": weights proportional to the diffusion function, or
                 - a custom function that returns a Callable.
+            beta_min: starting/minimal value for beta, the variance of the noise
+            beta_max: ending/maximal value for beta, the variance of the noise
+            mean_0: expected value of 0th step noise
+            scale_0: sigma of 0th step noise
+            t_min: smallest time step
+            t_max: largest time step
 
         """
         super().__init__(net, input_shape, condition_shape)
@@ -66,7 +73,14 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         # Set lambdas (variance weights) function.
         self._set_weight_fn(weight_fn)
 
-        # Min time for diffusion (0 can be numerically unstable).
+        # Store device of this module
+        self.device = net.device if hasattr(net, "device") else torch.device("cpu")
+
+        # Min/max values for noise variance beta
+        self.beta_min = beta_min
+        self.beta_max = beta_max
+
+        # Min/max values for limits to time
         self.t_min = t_min
         self.t_max = t_max
 
@@ -152,6 +166,7 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
             input: Input variable i.e. theta.
             condition: Conditioning variable.
             times: SDE time variable in [t_min, t_max]. Uniformly sampled if None.
+                if None, will be filled by calling the time_schedule method
             control_variate: Whether to use a control variate to reduce the variance of
                 the stochastic loss estimator.
             control_variate_threshold: Threshold for the control variate. If the std
@@ -161,13 +176,12 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
             MSE between target score and network output, scaled by the weight function.
 
         """
+        # update device if required
+        self.device = input.device if self.device != input.device else self.device
+
         # Sample diffusion times.
         if times is None:
-            times = (
-                torch.rand(input.shape[0], device=input.device)
-                * (self.t_max - self.t_min)
-                + self.t_min
-            )
+            times = self.times_schedule(input.shape[0])
 
         # Sample noise.
         eps = torch.randn_like(input)
@@ -312,6 +326,47 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         """
         raise NotImplementedError
 
+    def noise_schedule(self, times: Tensor) -> Tensor:
[…]
