### 2025-03-18 b7d3be3
added noise schedule test
