### 2025-03-18 1469881
working minimal example
