+++ sbi/inference/potentials/score_fn_util.py
@@ -37,6 +37,7 @@ def get_iid_method(name: str) -> Type["IIDScoreFunction"]:
         )
     return IID_METHODS[name]
 
+
 def get_guidance_method(name: str) -> Type["ScoreAdaptation"]:
     r"""
     Retrieves the guidance method by name.
@@ -48,10 +49,9 @@ def get_guidance_method(name: str) -> Type["ScoreAdaptation"]:
         The guidance method class.
     """
     if name not in GUIDANCE_METHODS:
-        raise NotImplementedError(
-            f"Method {name} for guidance not implemented."
-        )
+        raise NotImplementedError(f"Method {name} for guidance not implemented.")
     return GUIDANCE_METHODS[name]
+
 
 def register_guidance_method(name: str) -> Callable:
     r"""
@@ -96,7 +96,7 @@ class ScoreAdaptation(ABC):
         prior: Optional[Distribution],
         device: str = "cpu",
     ):
-        """ This class manages manipulating the score estimator to impose additional
+        """This class manages manipulating the score estimator to impose additional
         constraints on the posterior via guidance.
 
         Args:
@@ -112,6 +112,7 @@ class ScoreAdaptation(ABC):
     def __call__(self, theta: Tensor, x_o: Tensor, time: Optional[Tensor] = None):
         pass
 
+
 @register_guidance_method("classifier_free")
 class ClassifierFreeGuidance(ScoreAdaptation):
     def __init__(
@@ -124,7 +125,7 @@ class ClassifierFreeGuidance(ScoreAdaptation):
         likelihood_shift: float | Tensor,
         device: str = "cpu",
     ):
-        """ This class manages manipulating the score estimator to temper of shift the
+        """This class manages manipulating the score estimator to temper of shift the
         prior and likelihood.
 
         This is usually known as classifier-free guidance. And works by decomposing the
@@ -138,8 +139,10 @@ class ClassifierFreeGuidance(ScoreAdaptation):
         """
 
         if prior is None:
-            raise ValueError("Prior is required for classifier-free guidance, please"
-                            " provide as least an improper empirical prior.")
+            raise ValueError(
+                "Prior is required for classifier-free guidance, please"
+                " provide as least an improper empirical prior."
+            )
 
         self.prior_scale = prior_scale
         self.prior_shift = prior_shift
@@ -148,29 +151,28 @@ class ClassifierFreeGuidance(ScoreAdaptation):
 
         super().__init__(score_estimator, prior, device)
 
-    def marginal_prior_score(self, theta: Tensor,  time: Tensor):
-        """ Computes the marginal prior score analyticaly (or approximatly)
-        """
+    def marginal_prior_score(self, theta: Tensor, time: Tensor):
+        """Computes the marginal prior score analyticaly (or approximatly)"""
         m = self.score_estimator.mean_t_fn(time)
         std = self.score_estimator.std_fn(time)
-        marginal_prior = marginalize(self.prior, m, std)
+        marginal_prior = marginalize(self.prior, m, std)  # type: ignore
         marginal_prior_score = compute_score(marginal_prior, theta)
         return marginal_prior_score
 
-
-    def __call__(self, theta: Tensor, x_o: Tensor, time: Optional[Tensor] = None):
+    def __call__(self, input: Tensor, condition: Tensor, time: Optional[Tensor] = None):
         if time is None:
             time = torch.tensor([self.score_estimator.t_min])
 
-        posterior_score = self.score_estimator(input=theta, condition=x_o, time=time)
-        prior_score = self.marginal_prior_score(theta, time)
+        posterior_score = self.score_estimator(
[…]
