### 2025-03-18 0237adf
working minimal example
