+++ sbi/neural_nets/__init__.py
@@ -2,6 +2,7 @@ from sbi.neural_nets.factory import (
     classifier_nn,
     flowmatching_nn,
     likelihood_nn,
+    marginal_nn,
     posterior_nn,
     posterior_score_nn,
 )
@@ -24,4 +25,6 @@ def __getattr__(name):
         return posterior_nn
     elif name == "posterior_score_nn":
         return posterior_score_nn
+    elif name == "marginal_nn":
+        return marginal_nn
     raise AttributeError(f"Module '{__name__}' has no attribute '{name}'")
