### 2025-03-18 3919799
add util to generate mcmc samples directly from potential
