### 2025-03-18 777fb2d
first test for mmd misspecification
