### 2025-03-18 1213f69
misspecification test for embedding, bug fix
