### 2025-03-18 bd2735f
internal dataclasses but external stringly...
