+++ sbi/neural_nets/estimators/score_estimator.py
@@ -2,13 +2,13 @@
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
 import math
-from math import pi
 from typing import Callable, Optional, Union
 
 import torch
 from torch import Tensor, nn
 
 from sbi.neural_nets.estimators.base import ConditionalVectorFieldEstimator
+from sbi.utils.vector_field_utils import VectorFieldNet
 
 
 class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
@@ -61,9 +61,10 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
 
     def __init__(
         self,
-        net: nn.Module,
+        net: VectorFieldNet,
         input_shape: torch.Size,
         condition_shape: torch.Size,
+        embedding_net: nn.Module = nn.Identity(),
         weight_fn: Union[str, Callable] = "max_likelihood",
         mean_0: Union[Tensor, float] = 0.0,
         std_0: Union[Tensor, float] = 1.0,
@@ -78,6 +79,8 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
             net: Score estimator neural network with call signature:
                 input, condition, and time (in [0,1]).
             condition_shape: Shape of the conditioning variable.
+            embedding_net: Network to embed the conditioning variable before passing it
+                to the score network.
             weight_fn: Function to compute the weights over time. Can be one of the
                 following:
                 - "identity": constant weights (1.),
@@ -90,6 +93,9 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
 
         """
         super().__init__(net, input_shape, condition_shape, t_min, t_max)
+
+        # store embedding network
+        self._embedding_net = embedding_net
 
         # Set lambdas (variance weights) function.
         self._set_weight_fn(weight_fn)
@@ -109,6 +115,11 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         std_base = self.approx_marginal_std(torch.tensor([t_max])).flatten()[0]
         self._mean_base.fill_(mean_base)
         self._std_base.fill_(std_base)
+
+    @property
+    def embedding_net(self):
+        """Return the embedding network."""
+        return self._embedding_net
 
     def forward(self, input: Tensor, condition: Tensor, time: Tensor) -> Tensor:
         r"""Forward pass of the score estimator
@@ -132,6 +143,9 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         condition = torch.broadcast_to(condition, batch_shape + self.condition_shape)
         time = torch.broadcast_to(time, batch_shape)
 
+        # embed the conditioning variable
+        condition_emb = self._embedding_net(condition)
+
         # Time dependent mean and std of the target distribution to z-score the input
         # and to approximate the score at the end of the diffusion.
         mean = self.approx_marginal_mean(time)
@@ -148,7 +162,7 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
         score_gaussian = (input - mean) / std**2
 
         # Score prediction by the network
-        score_pred = self.net(input_enc, condition, time_enc)
+        score_pred = self.net(input_enc, condition_emb, time_enc)
 
         # Output pre-conditioned score
         # The learnable part will be largly scaled at the beginning of the diffusion
@@ -207,6 +221,9 @@ class ConditionalScoreEstimator(ConditionalVectorFieldEstimator):
                 + self.t_min
             )
 
+        # Embed conditioning variable
+        condition_emb = self._embedding_net(condition)
+
[…]
