### 2025-03-18 9db0eb8
refactor: update imports and enhance ConditionalScoreEstimator
- changed net type in ConditionalScoreEstimator and related classes to VectorFieldNet
- added embedding_net to ConditionalScoreEstimator for condition embedding
