### 2025-03-18 f1b9ba3
Update z-score parameters in flowmatching and posterior score neural networks, keeping others as independent
