+++ sbi/neural_nets/net_builders/flowmatching_nets.py
@@ -1,278 +1,0 @@
-# This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
-# under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
-
-# build function for flow matching neural networks
-# like in classifier.py, we need to build a network that can z-score the inputs
-
-# import ABC for abstract class
-from typing import Optional, Sequence, Union
-
-import torch
-import torch.nn as nn
-import zuko
-from torch import Tensor
-from torch.nn import functional as F
-from zuko.nn import MLP as ZukoMLP
-
-from sbi.neural_nets.estimators.flowmatching_estimator import FlowMatchingEstimator
-from sbi.utils.nn_utils import get_numel
-from sbi.utils.sbiutils import (
-    standardizing_net,
-    z_score_parser,
-    z_standardization,
-)
-from sbi.utils.user_input_checks import check_data_device
-from sbi.utils.vector_field_utils import VectorFieldNet
-
-
-def build_mlp_flowmatcher(
-    batch_x: Tensor,
-    batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
-    hidden_features: Union[Sequence[int], int] = 64,
-    num_layers: int = 5,
-    num_freqs: int = 3,
-    embedding_net: nn.Module = nn.Identity(),
-    **kwargs,
-) -> FlowMatchingEstimator:
-    """Builds a flow matching neural network.
-
-    Args:
-        batch_x: Batch of xs, used to infer dimensionality and (optional) z-scoring.
-        batch_y: Batch of ys, used to infer dimensionality and (optional) z-scoring.
-        z_score_x: Whether to z-score xs passing into the network, can be one of:
-            - `none`, or None: do not z-score.
-            - `independent`: z-score each dimension independently.
-            - `structured`: treat dimensions as related, therefore compute mean and std
-            over the entire batch, instead of per-dimension. Should be used when each
-            sample is, for example, a time series or an image.
-        z_score_y: Whether to z-score ys passing into the network, same options as
-            z_score_x.
-        hidden_features: Number of hidden features.
-        num_layers: Number of num_layers in the vector field regressor.
-        num_freqs: Number of frequencies in the time embeddings.
-        embedding_net: Embedding network for batch_y.
-        kwargs: Additional keyword arguments passed to the FlowMatchingEstimator.
-    """
-    # Infer the output dimensionality of the embedding_net by making a forward pass.
-    check_data_device(batch_x, batch_y)
-    x_numel = get_numel(batch_x)
-    y_numel = get_numel(batch_y, embedding_net=embedding_net)
-
-    # create a list of layers for the regression network; the vector field
-    # regressor is a MLP consisting of num_layers of layers with
-    # hidden_features neurons each
-    if isinstance(hidden_features, int):
-        hidden_features = [hidden_features] * num_layers
-
-    # Wrap zuko MLP as VectorFieldNet to ensure uniform forward signature
-    vectorfield_net = VectorFieldMLP(
-        net=ZukoMLP(
-            in_features=x_numel + y_numel + 2 * num_freqs,
-            out_features=x_numel,
-            hidden_features=hidden_features,
-            activation=nn.ELU,
-        )
-    )
-
-    # input data is only z-scored, not embedded.
-    z_score_x_bool, structured_x = z_score_parser(z_score_x)
-    if z_score_x_bool:
-        t_mean, t_std = z_standardization(batch_x, structured_x)
-        z_score_transform = torch.distributions.AffineTransform(
-            -t_mean / t_std, 1 / t_std
-        )
[…]
