### 2025-03-18 291a213
Update z-score parameters in flowmatching and posterior score neural networks, keeping others as independent
