### 2025-03-18 dab4f07
Update z-score parameters in flowmatching and posterior score neural networks, keeping others as independent
