### 2025-03-18 3b380cf
Update z-score parameters in flowmatching and posterior score neural networks, keeping others as independent
