### 2025-03-18 dca1939
implemented beta schedule for variance-preserving estimators
- added tests too
- inspired by https://arxiv.org/abs/2206.00364
