### 2025-03-18 8be9079
add new ZScoreTypes
