+++ sbi/neural_nets/net_builders/flow.py
@@ -1,6 +1,7 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
+import warnings
 from functools import partial
 from typing import List, Optional, Sequence, Union
 
@@ -17,6 +18,8 @@ from torch import Tensor, nn, relu, tanh, tensor, uint8
 from sbi.neural_nets.estimators import NFlowsFlow, ZukoFlow
 from sbi.utils.nn_utils import MADEMoGWrapper, get_numel
 from sbi.utils.sbiutils import (
+    ZScoreType,
+    logit_transform_zuko,
     standardizing_net,
     standardizing_transform,
     standardizing_transform_zuko,
@@ -31,8 +34,8 @@ nflow_specific_kwargs = ["num_bins", "num_components", "tail_bound"]
 def build_made(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
+    z_score_x: Optional[ZScoreType] = "affine-independent",
+    z_score_y: Optional[ZScoreType] = "affine-independent",
     hidden_features: int = 50,
     num_mixture_components: int = 10,
     embedding_net: nn.Module = nn.Identity(),
@@ -102,8 +105,8 @@ def build_made(
 def build_maf(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
+    z_score_x: Optional[ZScoreType] = "affine-independent",
+    z_score_y: Optional[ZScoreType] = "affine-independent",
     hidden_features: int = 50,
     num_transforms: int = 5,
     embedding_net: nn.Module = nn.Identity(),
@@ -190,8 +193,8 @@ def build_maf(
 def build_maf_rqs(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
+    z_score_x: Optional[ZScoreType] = "affine-independent",
+    z_score_y: Optional[ZScoreType] = "affine-independent",
     hidden_features: int = 50,
     num_transforms: int = 5,
     embedding_net: nn.Module = nn.Identity(),
@@ -302,8 +305,8 @@ def build_maf_rqs(
 def build_nsf(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
+    z_score_x: Optional[ZScoreType] = "affine-independent",
+    z_score_y: Optional[ZScoreType] = "affine-independent",
     hidden_features: int = 50,
     num_transforms: int = 5,
     num_bins: int = 10,
@@ -424,8 +427,8 @@ def build_nsf(
 def build_zuko_nice(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
+    z_score_x: Optional[ZScoreType] = "affine-independent",
+    z_score_y: Optional[ZScoreType] = "affine-independent",
     hidden_features: Union[Sequence[int], int] = 50,
     num_transforms: int = 5,
     embedding_net: nn.Module = nn.Identity(),
@@ -479,8 +482,8 @@ def build_zuko_nice(
 def build_zuko_maf(
     batch_x: Tensor,
     batch_y: Tensor,
-    z_score_x: Optional[str] = "independent",
-    z_score_y: Optional[str] = "independent",
[…]
