### 2025-03-18 23bd180
add new ZScoreTypes
