+++ tests/embedding_net_test.py
@@ -17,6 +17,7 @@ from sbi.neural_nets.embedding_nets import (
     LRUEmbedding,
     PermutationInvariantEmbedding,
 )
+from sbi.neural_nets.embedding_nets.lru import LRU
 from sbi.simulators.linear_gaussian import (
     linear_gaussian,
     true_posterior_linear_gaussian_mvn_prior,
@@ -253,14 +254,40 @@ def test_npe_with_with_iid_embedding_varying_num_trials(trial_factor=50):
             )
 
 
-def test_lru_embedding_isolated(
-    dim_observation: int = 7, output_dim: int = 5, batch_size: int = 16
+def test_lru_isolated(
+    input_dim: int = 7,
+    state_dim: int = 11,
+    r_min: float = 0.1,
+    r_max: float = 1.0,
+    max_phase: float = 2 * torch.pi,
+    bidirectional: bool = False,
+    batch_size: int = 16,
+):
+    """Run some random data trough an LRU network."""
+    lru = LRU(
+        input_dim=input_dim,
+        state_dim=state_dim,
+        r_min=r_min,
+        r_max=r_max,
+        max_phase=max_phase,
+        bidirectional=bidirectional,
+    )
+
+    x = torch.randn(batch_size, 100, input_dim)
+
+    y = lru(x)
+    assert isinstance(y, Tensor)
+    assert y.shape == (batch_size, 100, input_dim)
+
+
+def test_lru_embedding_net_isolated(
+    observation_dim: int = 7, output_dim: int = 5, batch_size: int = 16
 ):
     """Run some random data trough an LRUEmbedding network."""
     embedding_net = LRUEmbedding(
-        input_dim=dim_observation, output_dim=output_dim, hidden_dim=9, num_layers=3
+        input_dim=observation_dim, output_dim=output_dim, hidden_dim=9, num_layers=3
     )
-    x = torch.randn(batch_size, 100, dim_observation)
+    x = torch.randn(batch_size, 100, observation_dim)
 
     x_embed = embedding_net(x)
     assert isinstance(x_embed, Tensor)
