+++ sbi/neural_nets/embedding_nets/lru.py
@@ -1,15 +1,14 @@
-from typing import Callable
+from math import sqrt
+from typing import Optional
 
-import lru
+import numpy as np
 import torch
 from torch import Tensor, nn
-import numpy as np
-
-from torch.utils._pytree import tree_flatten, tree_unflatten
 
 
 class LRUEmbedding(nn.Module):
-    """Embedding network backed by a Linear Recurrent Unit (LRU).
+    """Embedding network backed by a stack of Linear Recurrent Unit (LRU) layers.
+
     See also:
         https://arxiv.org/pdf/2303.06349
     """
@@ -27,7 +26,7 @@ class LRUEmbedding(nn.Module):
         bidirectional: bool = False,
         dropout: float = 0.0,
         norm: bool = False,
-        #aggretate_func : Callable,
+        # aggretate_func : Callable,
     ):
         """Fully-connected multi-layer neural network to be used as embedding network.
 
@@ -74,8 +73,6 @@ class LRUEmbedding(nn.Module):
         x = self.embedding(x)
         x_scan = self.net(x)  # (batch_size, len_sequence, output_dim)
 
-
-
         # Pooling
         # TODO: implement pooling
 
@@ -94,12 +91,12 @@ class LRUBlock(nn.Module):
         max_phase=2 * np.pi,
         bidirectional: bool = False,
         dropout: float = 0.0,
-        norm: Bool=False,
+        norm: bool = False,
     ):
         super().__init__()
-        #https://github.com/NicolasZucchet/minimal-LRU/blob/main/lru/model.py
+        # https://github.com/NicolasZucchet/minimal-LRU/blob/main/lru/model.py
         self.lru = LRU(hidden_dim, state_dim, r_min, r_max, max_phase)
-        self.out1 = nn.Linear(hidden_dim,hidden_dim)
+        self.out1 = nn.Linear(hidden_dim, hidden_dim)
         self.out2 = nn.Linear(hidden_dim, hidden_dim)
         self.GELU = nn.GELU()
         self.sigmoid = nn.Sigmoid()
@@ -112,13 +109,13 @@ class LRUBlock(nn.Module):
         y = self.norm(x)
 
         # run LRU
-        y=self.lru(y)
+        y = self.lru(y)
 
         # state mixing
         y = self.GELU(y)
         y = self.dropout(y)
-        y = self.out1(y)*self.sigmoid(self.out2(y))
-        y = self.dropout(y) 
+        y = self.out1(y) * self.sigmoid(self.out2(y))
+        y = self.dropout(y)
 
         # skip connection
         y = y + x
@@ -126,81 +123,242 @@ class LRUBlock(nn.Module):
 
 
 class LRU(nn.Module):
+    """A single Linear Recurrent Unit (LRU).
+
+    This implementation, just like all other, makes the simplification
+    that `output_dim = input_dim`. The benefor is that you can stack
+    multiple LRU instances without linear layers inbetween, and the
+    skip connections `D` become element-wise, i.e., `D` is a vector.
+    """
+
     def __init__(
         self,
         input_dim: int,
-        state_dim: int = 20,
-        r_min=0.1,
-        r_max=1.0,
-        max_phase=2 * np.pi,
+        state_dim: int,
+        r_min: float = 0.1,
+        r_max: float = 1.0,
+        max_phase: float = 2 * torch.pi,
         bidirectional: bool = False,
     ):
         super().__init__()
+        self.state_dim = state_dim
 
         # between r_min and r_max, with phase in [0, max_phase].
         if bidirectional:
-            state_dim = state_dim *2
+            state_dim = state_dim * 2
 
         u1 = torch.rand(size=(state_dim,))
[…]
