+++ tests/linearGaussian_npse_test.py
@@ -4,61 +4,81 @@ import pytest
 from torch import eye, ones, zeros
 from torch.distributions import MultivariateNormal
 
-from sbi import analysis as analysis
 from sbi import utils as utils
 from sbi.inference import NPSE
 from sbi.simulators import linear_gaussian
 from sbi.simulators.linear_gaussian import (
-    samples_true_posterior_linear_gaussian_mvn_prior_different_dims,
     samples_true_posterior_linear_gaussian_uniform_prior,
     true_posterior_linear_gaussian_mvn_prior,
 )
-from sbi.utils import BoxUniform
 
 from .test_utils import check_c2st, get_dkl_gaussian_prior
 
-DEFAULT_USE = [
-    pytest.param(("ve", None), id="ve-None"),
-    pytest.param(("vp", "gaussian"), id="vp-gaussian"),
-]
 
-SLOW_COMBINATIONS = [
-    pytest.param(("vp", "uniform"), id="vp-uniform", marks=pytest.mark.slow),
-    pytest.param(("vp", None), id="vp-None", marks=pytest.mark.slow),
-    pytest.param(("ve", "gaussian"), id="ve-gaussian", marks=pytest.mark.slow),
-    pytest.param(("ve", "uniform"), id="ve-uniform", marks=pytest.mark.slow),
-    pytest.param(("subvp", "gaussian"), id="subvp-gaussian", marks=pytest.mark.slow),
-    pytest.param(("subvp", "uniform"), id="subvp-uniform", marks=pytest.mark.slow),
-    pytest.param(("subvp", None), id="subvp-None", marks=pytest.mark.slow),
-]
+def _initialize_experiment(
+    num_dim: int, num_samples: int, num_simulations: int, prior_type: str
+):
+    """Set up priors, generate target samples, and create simulation data."""
+    if prior_type in {"gaussian", None}:
+        prior_mean = zeros(num_dim)
+        prior_cov = eye(num_dim)
+        prior = MultivariateNormal(loc=prior_mean, covariance_matrix=prior_cov)
+    elif prior_type == "uniform":
+        prior = utils.BoxUniform(-2.0 * ones(num_dim), 2.0 * ones(num_dim))
+        prior_mean, prior_cov = None, None
+    else:
+        raise NotImplementedError(f"Unsupported prior type: {prior_type}")
+
+    x_o = zeros((1, num_dim))
+    likelihood_shift = -ones(num_dim)
+    likelihood_cov = 0.3 * eye(num_dim)
+
+    if prior_type in {"gaussian", None}:
+        gt_posterior = true_posterior_linear_gaussian_mvn_prior(
+            x_o, likelihood_shift, likelihood_cov, prior_mean, prior_cov
+        )
+        target_samples = gt_posterior.sample((num_samples,))
+    else:  # prior_type == "uniform"
+        target_samples = samples_true_posterior_linear_gaussian_uniform_prior(
+            x_o, likelihood_shift, likelihood_cov, prior=prior, num_samples=num_samples
+        )
+
+    theta = prior.sample((num_simulations,))
+    x = linear_gaussian(theta, likelihood_shift, likelihood_cov)
+    prior = None if prior_type is None else prior
+
+    return (
+        prior,
+        target_samples,
+        theta,
+        x,
+        x_o,
+        likelihood_shift,
+        likelihood_cov,
+        prior_mean,
+        prior_cov,
+    )
 
 
 @pytest.fixture(scope="module")
 def npse_trained_model(request):
     """Module-scoped fixture that trains a score estimator for NPSE tests."""
+    num_dim, num_simulations, num_samples = 2, 5_000, 1_000
     sde_type, prior_type = request.param
-    num_dim = 2
-    num_simulations = 5000
 
-    # likelihood_mean will be likelihood_shift+theta
[…]
