### 2025-03-18 dc45d65
fix: Warn when using append_simulations with exclude_invalid_x=True.
Update SNRE to NRE and SNLE to NLE in append_simulations docstring.

Fixes: #1427
