### 2025-03-19 c827be5
direct posterior: warning for slow batched sampling with large batch size
