### 2025-03-19 4f9524f
review changes: comments, missing import, static type check
