+++ sbi/neural_nets/estimators/mixed_density_estimator.py
@@ -13,10 +13,11 @@ from sbi.neural_nets.estimators.categorical_net import CategoricalMassEstimator
 class MixedDensityEstimator(ConditionalDensityEstimator):
     """Class performing Mixed Neural Density Estimation.
 
-    This estimator combines a Categorical net and a neural density estimator to model
-    data with mixed types (discrete and continuous), e.g., as they occur in
-    decision-making models. It can be used for both likelihood and posterior estimation
-    of mixed data.
+    This estimator combines a categorical mass estimator and a density estimator to
+    model variables with mixed types (discrete and continuous). It can be used for both
+    likelihood estimation (e.g., for discrete decisions and continuous reaction times
+    in decision-making models) or posterior estimation (e.g., for models that have both
+    discrete and continuous parameters).
     """
 
     def __init__(
@@ -55,7 +56,7 @@ class MixedDensityEstimator(ConditionalDensityEstimator):
     def forward(self, input: Tensor):
         raise NotImplementedError(
             """The forward method is not implemented for mixed neural density
-            estimation,use '.sample(...)' to generate samples though a forward
+            estimation, use '.sample(...)' to generate samples though a forward
             pass."""
         )
 
