### 2025-03-19 677f29f
move vector field estimator and extend
