### 2025-03-19 770e29a
common methods for flow and sde estimators
