+++ sbi/neural_nets/estimators/flowmatching_estimator.py
@@ -4,14 +4,14 @@ from typing import Optional, Tuple
 
 import torch
 import torch.nn as nn
-import zuko
+
 from torch import Tensor
 from torch.distributions import Transform
 from zuko.distributions import DiagNormal, NormalizingFlow
 from zuko.transforms import FreeFormJacobianTransform
 from zuko.utils import broadcast
 
-from sbi.neural_nets.estimators.base import ConditionalDensityEstimator
+from sbi.neural_nets.estimators.vector_field_estimator import ConditionalVectorFieldEstimator
 
 
 # abstract class to ensure forward signature for flow matching networks
@@ -20,14 +20,13 @@ class VectorFieldNet(nn.Module, ABC):
     def forward(self, theta: Tensor, x: Tensor, t: Tensor) -> Tensor: ...
 
 
-class FlowMatchingEstimator(ConditionalDensityEstimator):
+class FlowMatchingEstimator(ConditionalVectorFieldEstimator):
     def __init__(
         self,
         net: VectorFieldNet,
         input_shape: torch.Size,
         condition_shape: torch.Size,
         embedding_net: nn.Module,
-        zscore_transform_input: Optional[Transform] = None,
         num_freqs: int = 3,
         noise_scale: float = 1e-3,
         **kwargs,
@@ -40,7 +39,6 @@ class FlowMatchingEstimator(ConditionalDensityEstimator):
             condition_shape: Shape of the condition, e.g., the data.
             noise_scale: Scale of the noise added to the vector field.
             embedding_net: Embedding network for the condition.
-            zscore_transform_input: Transform to z-score the input.
             num_freqs: Number of frequencies to use for the positional time encoding.
             noise_scale: Scale of the noise added to the vector field.
         """
@@ -51,9 +49,6 @@ class FlowMatchingEstimator(ConditionalDensityEstimator):
 
         self.noise_scale = noise_scale
         # Identity transform for z-scoring the input
-        if zscore_transform_input is None:
-            zscore_transform_input = zuko.transforms.IdentityTransform()
-        self.zscore_transform_input: Transform = zscore_transform_input
         self._embedding_net = embedding_net
 
         self.register_buffer("freqs", torch.arange(1, num_freqs + 1) * math.pi)
@@ -71,18 +66,17 @@ class FlowMatchingEstimator(ConditionalDensityEstimator):
 
         # embed the input and condition
         embedded_condition = self._embedding_net(condition)
-        zscored_input = self.zscore_transform_input(input)
 
         # broadcast to match shapes of theta, x, and t
-        zscored_input, embedded_condition, t = broadcast(
-            zscored_input,  # type: ignore
+        input, embedded_condition, t = broadcast(
+            input,  # type: ignore
             embedded_condition,
             t,
             ignore=1,
         )
 
         # return the estimated vector field
-        return self.net(theta=zscored_input, x=embedded_condition, t=t)
+        return self.net(theta=input, x=embedded_condition, t=t)
 
     def loss(self, input: Tensor, condition: Tensor, **kwargs) -> Tensor:
         """Return the loss for training the density estimator. More precisely,
@@ -111,56 +105,30 @@ class FlowMatchingEstimator(ConditionalDensityEstimator):
             (self.forward(theta_prime, condition, t) - vector_field) ** 2, dim=-1
         )
 
-    def log_prob(self, input: Tensor, condition: Tensor, **kwargs) -> Tensor:
-        input_batch_dim = input.shape[1]
-        condition_batch_dim = condition.shape[0]
-
-        assert condition_batch_dim == input_batch_dim, (
-            f"Batch shape of condition {condition_batch_dim} and input "
-            f"{input_batch_dim} do not match."
-        )
-
-        # the flow will apply and take into account input zscoring.
[…]
