+++ sbi/inference/trainers/marginal/marginal_base.py
@@ -2,9 +2,10 @@
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
 import time
-from abc import ABC
 from copy import deepcopy
-from typing import Callable, Optional, Tuple, Union
+from datetime import datetime
+from pathlib import Path
+from typing import Callable, Dict, Optional, Tuple, Union
 
 import torch
 from torch import Tensor
@@ -14,33 +15,47 @@ from torch.utils import data
 from torch.utils.data.sampler import SubsetRandomSampler
 from torch.utils.tensorboard.writer import SummaryWriter
 
-from sbi.inference.trainers.base import NeuralInference
 from sbi.neural_nets.estimators import UnconditionalDensityEstimator
 from sbi.neural_nets.estimators.shape_handling import (
     reshape_to_batch_event,
 )
 from sbi.neural_nets.factory import marginal_nn
-from sbi.utils import check_estimator_arg
-from sbi.utils.torchutils import assert_all_finite
+from sbi.utils import check_estimator_arg, get_log_root
+from sbi.utils.torchutils import assert_all_finite, process_device
 
 
-class MarginalTrainer(NeuralInference, ABC):
+class MarginalTrainer:
     def __init__(
         self,
         density_estimator: Union[str, Callable] = "MAF",
         device: str = "cpu",
-        logging_level: Union[int, str] = "WARNING",
         summary_writer: Optional[SummaryWriter] = None,
         show_progress_bars: bool = True,
     ):
         """Base class for Marginal estimation method."""
 
-        super().__init__(
-            prior=None,
-            device=device,
-            logging_level=logging_level,
-            summary_writer=summary_writer,
-            show_progress_bars=show_progress_bars,
+        self._device = process_device(device)
+        self._neural_net = None
+
+        self._show_progress_bars = show_progress_bars
+
+        # Initialize list that indicates the round from which simulations were drawn.
+        self._data_round_index = []
+
+        self._round = 0
+        self._val_loss = float("Inf")
+
+        self._summary_writer = (
+            self._default_summary_writer() if summary_writer is None else summary_writer
+        )
+
+        # Logging during training (by SummaryWriter).
+        self._summary = dict(
+            epochs_trained=[],
+            best_validation_loss=[],
+            validation_loss=[],
+            training_loss=[],
+            epoch_durations_sec=[],
         )
 
         check_estimator_arg(density_estimator)
@@ -106,7 +121,12 @@ class MarginalTrainer(NeuralInference, ABC):
         Returns:
             Negative log prob.
         """
-        x = reshape_to_batch_event(x, event_shape=self._neural_net.input_shape)
+        if self._neural_net is None:
+            raise ValueError(
+                "Neural network has not been initialized. Please call `train` first."
+            )
+        else:
+            x = reshape_to_batch_event(x, event_shape=self._neural_net.input_shape)
         loss = self._neural_net.loss(x)
         assert_all_finite(loss, "loss")
         return loss
@@ -235,3 +255,138 @@ class MarginalTrainer(NeuralInference, ABC):
         self._neural_net.zero_grad(set_to_none=True)
 
         return deepcopy(self._neural_net)
+
+    def _default_summary_writer(self) -> SummaryWriter:
+        """Return summary writer logging to method- and simulator-specific directory."""
+
+        method = self.__class__.__name__
+        logdir = Path(
+            get_log_root(), method, datetime.now().isoformat().replace(":", "_")
+        )
+        return SummaryWriter(logdir)
+
+    def _converged(self, epoch: int, stop_after_epochs: int) -> bool:
+        """Return whether the training converged yet and save best model state so far.
+
+        Checks for improvement in validation performance over previous epochs.
+
[…]
