### 2025-03-19 7b140b4
simplify mixed nets (default logtransform set for mnle/mnpe), merge conflict with tutorial
