### 2025-03-19 0b7c31f
Changes to compute log_prob of NPSE under iid observations
