### 2025-03-19 23f5085
add option 'importance' to sample_with argument
