### 2025-03-19 35d1404
bugfix: error message for batched sampling not implemented
