### 2025-03-19 b291aac
adapt threshold for warning
