+++ tests/posterior_sampler_test.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 
 import sys
 
+import pymc
 import pytest
 from pyro.infer.mcmc import MCMC
 from torch import Tensor, eye, zeros
@@ -29,10 +30,9 @@ from sbi.simulators.linear_gaussian import diagonal_linear_gaussian
         "hmc_pyro",
         pytest.param(
             "nuts_pymc",
-            marks=pytest.mark.xfail(
-                condition=sys.version_info >= (3, 10),
-                reason="Fails with pymc>=5.20.1 and python>=3.10",
-                raises=TypeError,
+            marks=pytest.mark.skipif(
+                condition=sys.version_info >= (3, 10) and pymc.__version__ >= "5.20.1",
+                reason="Inconsistent behaviour with pymc>=5.20.1 and python>=3.10",
             ),
         ),
         "hmc_pymc",
