### 2025-03-19 #7754 @janfb
Hi @boisgera

thanks for generating the proofs!

Some minor formatting issues:
- on page 1 lower left corner there is a bit of overlap between License disclaimer and the main text / line numbers.
- the formatting of the citations is a bit inconsistent. Sometimes we have first names popping up, e.g., for "Francois Rozet", or "Stefan T. Radev" (independent of whether there are more authors or not).
- Then we sometimes have the first two authors showing up, e.g., "Deistler, Goncalves et al." even though there are more authors and the first two are not equally contributing.
- in contrast, the footnote we show "Boelts et al.", but I guess there is no way to show "Boelts, Deistler et al"?
