### 2025-03-19 3e963a9
cloned VPScoreEstimator to yield improved version
in addition:
- added improved version to benchmarks (for later comparison)
- created new class ImprovedVPScoreEstimator
