### 2025-03-19 c363126
chore: add syrupy dependency for snapshot testing, add snapshot test and snapshot file
