### 2025-03-19 db7dc3b
skipif for all tests using nuts_pymc sampler
