### 2025-03-19 762ed43
Other more general guidance methods
