### 2025-03-19 51cde80
Other more general guidance methods
