### 2025-03-19 0ade94a
Add Pyro distribution wrapper and basic properties tests
