+++ sbi/utils/pyroutils.py
@@ -1,10 +1,16 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
-from typing import Any, Callable
+from abc import abstractmethod
+from typing import Any, Callable, Tuple
 
+import pyro.distributions
+import torch
+import torch.distributions.constraints as constraints
 from pyro import poutine as poutine
 from torch.distributions import biject_to
+
+from sbi.neural_nets.estimators.base import ConditionalEstimator
 
 
 def get_transforms(model: Callable, *model_args: Any, **model_kwargs: Any):
@@ -33,3 +39,125 @@ def get_transforms(model: Callable, *model_args: Any, **model_kwargs: Any):
             transforms[name] = biject_to(fn.support).inv
 
     return transforms
+
+
+class EstimatorDistribution(pyro.distributions.TorchDistribution):
+    """
+    Base class for a conditioned `sbi` estimator wrapped as a Pyro distribution.
+    """
+
+    @property
+    def support(self):
+        # TODO: check if we can infer a more specific support from the estimator
+        return constraints.real  # Assume continuous values
+
+    def __init__(self, estimator: Any, condition: torch.Tensor):
+        self._estimator = estimator
+        self._condition_shape, event_shape = self.get_condition_and_event_shapes(
+            estimator
+        )
+        self._check_condition_shape(condition)
+        self._condition = condition
+        self._condition_reshaped = condition.reshape(-1, *self._condition_shape)
+        super().__init__(
+            batch_shape=condition.shape[: -len(self._condition_shape)],
+            event_shape=event_shape,
+        )
+
+    @abstractmethod
+    def get_condition_and_event_shapes(estimator: Any) -> Tuple[torch.Size, torch.Size]:
+        pass
+
+    @abstractmethod
+    def _estimator_log_prob(
+        self, x: torch.Tensor, condition: torch.Tensor
+    ) -> torch.Tensor:
+        """The log probability method of the estimator."""
+        pass
+
+    @property
+    def estimator(self):
+        """The conditioned estimator."""
+        return self._estimator
+
+    @property
+    def condition(self):
+        """The conditioning parameter."""
+        return self._condition
+
+    def _check_condition_shape(self, condition: torch.Tensor):
+        """Check that the shape of `condition` is compatible with the estimator."""
+        if len(condition.shape) < len(self._condition_shape):
+            raise ValueError(
+                "Dimensionality of condition is too small and does not match the "
+                f"expected dimensionality {len(self._condition_shape)}. "
+                "It should be compatible with condition_shape "
+                f"{self._condition_shape}."
+            )
+        else:
+            condition_shape = condition.shape[-len(self._condition_shape) :]
+            if condition_shape != self._condition_shape:
+                raise ValueError(
+                    f"Condition shape {condition_shape} is not compatible with "
+                    f"estimator condition shape {self._condition_shape}"
+                )
+
+    def sample(self, sample_shape: torch.Size = torch.Size()) -> torch.Tensor:
+        """Generate samples from the conditioned estimator."""
+        if not hasattr(self.estimator, "sample"):
+            raise NotImplementedError("Sampling is not implemented for this estimator.")
+        draws = self.estimator.sample(sample_shape, condition=self._condition_reshaped)
+        # reshape (batch_dim,) -> batch_shape
+        return draws.reshape(*sample_shape, *self.batch_shape, *self.event_shape)
+
+    def log_prob(self, x: torch.Tensor) -> torch.Tensor:
+        """Compute log probability of x given theta."""
[…]
