### 2025-03-19 7137b39
Update documentation and add 1D Resnet
