+++ sbi/neural_nets/embedding_nets/resnet.py
@@ -1,13 +1,14 @@
 from functools import partial
-from typing import Callable, Type
+from typing import Callable, Dict, List, Type, Union
 
 import torch
 import torch.nn as nn
+from torch import Tensor
 
 
-def zero_pad(x: torch.tensor, c_out: int) -> torch.tensor:
+def zero_pad(x: Tensor, c_out: int) -> Tensor:
     """
-    Add additional channels to the input tensor by filling it with zeros.
+    Add additional channels to the input tensor by padding it with zeros.
 
     Args:
         x:      Input tensor
@@ -17,11 +18,13 @@ def zero_pad(x: torch.tensor, c_out: int) -> torch.tensor:
         x_aug:      Input tensor with additional channels filled with zeros
     """
 
-    # Check if it is a image like iput
-    assert len(x.shape) == 4
+    # Check if it is an image like-input
+    if not (len(x.shape) == 4):
+        raise ValueError("Only 4D input tensors are supported.")
 
     # Check if the number of channels is increased
-    assert c_out >= x.shape[1]
+    if c_out <= x.shape[1]:
+        raise ValueError("c_out must be larger than c_in to apply zero padding.")
 
     c_in = x.shape[1]
     a = torch.zeros(x.shape[0], c_out - c_in, x.shape[2], x.shape[3], device=x.device)
@@ -34,9 +37,9 @@ def zero_pad(x: torch.tensor, c_out: int) -> torch.tensor:
 def construct_simple_conv_net(
     c_in: int,
     c_out: int,
-    c_hidden: int = 32,
-    activation: Type[nn.Module] = nn.ReLU,
-    downsample: int = False,
+    c_hidden: int,
+    activation: Type[nn.Module],
+    downsample: bool,
 ) -> nn.Module:
     """
     Default implementation for a simple convolutional network with 3 layers.
@@ -45,7 +48,7 @@ def construct_simple_conv_net(
         c_in:       Number of input channels
         c_out:      Number of output channels
         c_hidden:   Number of hidden channels
-        activation: Constructor for a activation function
+        activation: Constructor for an activation function
         downsample: Apply a strided convolution to halve
                     the height and width of the input
 
@@ -66,131 +69,72 @@ def construct_simple_conv_net(
     return layers
 
 
-class ResidualBLockConv(nn.Module):
-    def __init__(
-        self,
-        c_in: int,
-        c_out: int | None = None,
-        change_c_mode: str = "conv",
-        downsample: bool = False,
-        construct_mapping: Callable = construct_simple_conv_net,
-        construct_mapping_kwargs: dict | None = None,
-        activation: Type[nn.Module] = nn.ReLU,
-    ) -> None:
-        """
-        Single residual block for image like data.
+def construct_simple_fc_net(
+    c_in: int, c_out: int, c_hidden: int, activation: Type[nn.Module]
+) -> nn.Module:
+    """
+    Default implementation for a fully connected neural network.
 
-        Args:
-            c_in: Number of input channels.
-            c_out: Number of output channels. If None, use c_in.
-            change_c_mode: Mode to change the number of channels.
-                           Options are "conv" and "zeros". If "conv" is selected,
-                           1x1 convolutions are applied to change the number
-                           of channels. If "zeros" is selected, the required additional
-                           channels are filled with zeros.
-            downsample: Apply a strided convolution to halve the height and
-                        width of the input.
-            construct_mapping: Constructor for the mapping function. The mapping
-                               function is applied to the input tensor before the
-                               residual connection is added.
-            construct_mapping_kwargs: Additional keyword arguments for the mapping
-                                      function.
-            activation: Constructor for the activation function.
-        """
+    Args:
+        c_in:       Number of input channels
[…]
