+++ sbi/neural_nets/embedding_nets/__init__.py
@@ -4,5 +4,6 @@ from sbi.neural_nets.embedding_nets.permutation_invariant import (
     PermutationInvariantEmbedding,
 )
 from sbi.neural_nets.embedding_nets.resnet import (
-    ResNetEmbedding,
+    ResNetEmbedding1D,
+    ResNetEmbedding2D,
 )
