### 2025-03-19 eb3e9cf
Remove lingering print statements
