### 2025-03-19 2d299bf
Test that SBI and Pyro models agree
