+++ tests/pyroutils_test.py
@@ -1,15 +1,55 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
 
+import numpy as np
 import pyro
+import pyro.distributions as dist
 import pytest
 import torch
+from pyro.infer import MCMC, NUTS
 
 from sbi.inference import NLE, NPE
 from sbi.utils.pyroutils import (
     ConditionedEstimatorDistribution,
     get_transforms,
 )
+from tests.test_utils import check_c2st
+
+
+def single_level_prior_theta(num_dim=10):
+    mu_0 = torch.zeros(num_dim)
+    sigma_0 = torch.ones(num_dim) * 1.0
+    return dist.MultivariateNormal(mu_0, torch.diag(sigma_0))
+
+
+def pyro_gaussian_model(x_o=None, num_dim=None, num_trials=1, sigma_x=0.5):
+    """
+    Pyro model for Bayesian inference with Gaussian likelihoods.
+
+    Args:
+        x_o (torch.Tensor, optional): Observed data (N, D), if available.
+
+    Returns:
+        Pyro model for inference.
+    """
+    if num_dim is None:
+        if x_o is not None:
+            num_dim = x_o.shape[-1]
+        else:
+            raise ValueError("num_dim must be provided if x_o is not provided")
+
+    theta = pyro.sample("theta", single_level_prior_theta(num_dim))  # Shape: (D,)
+
+    # Sample observations (batch shape: N, event shape: D)
+    with pyro.plate("trials", num_trials):
+        x = pyro.sample(
+            "x", dist.MultivariateNormal(theta, sigma_x * torch.eye(num_dim)), obs=x_o
+        )
+
+    if x_o is None:
+        return theta, x
+    else:
+        return x_o
 
 
 def test_unbounded_transform():
@@ -82,3 +122,59 @@ def test_estimator_distribution_basic_properties(
         theta.expand(torch.Size([2, num_simulations, dim])),
     )
     assert estimator_dist_expanded.estimator == estimator_dist.estimator
+
+
+@pytest.mark.parametrize("num_dim", [3, 5])
+@pytest.mark.parametrize("num_trials", [1, 5])
+def test_pyro_gaussian_model(
+    num_dim, num_trials, num_simulations=500, num_samples=500, warmup_steps=500
+):
+    """Test consistency of MCMC samples between the true and estimated likelihood."""
+    _, x_o = pyro_gaussian_model(num_trials=num_trials, num_dim=num_dim)
+
+    nuts_kernel = NUTS(pyro_gaussian_model)
+    mcmc = MCMC(nuts_kernel, num_samples=num_samples, warmup_steps=warmup_steps)
+    mcmc.run(x_o=x_o)
+
+    pyro_samples = torch.from_numpy(np.array(mcmc.get_samples()["theta"]))
+
+    # Run sbi
+    prior = single_level_prior_theta(num_dim=num_dim)
+    theta = []
+    x = []
+    for _ in range(num_simulations):
+        ti, xi = pyro_gaussian_model(num_trials=1, num_dim=num_dim)
+        theta.append(torch.from_numpy(np.array(ti)))
+        x.append(torch.from_numpy(np.array(xi)))
+    theta = torch.stack(theta, dim=0).float()
+    x = torch.stack(x, dim=0).float().squeeze()
+
+    trainer = NLE(prior=prior).append_simulations(theta=theta, x=x)
+    density_estimator = trainer.train()
+
+    def sbi_pyro_model(likelihood, x_o=None):
[…]
