### 2025-03-19 b475446
remove legacy mnle.py that was not interacted with by users
