### 2025-03-19 #7754 @boisgera
> * on page 1 lower left corner there is a bit of overlap between License disclaimer and the main text / line numbers.

Yes, indeed! However the line numbers are only present in the proof-reading stage so I wouldn't worry about this one.
