### 2025-03-19 #7754 @boisgera
> * the formatting of the citations is a bit inconsistent. Sometimes we have first names popping up, e.g., for "Francois Rozet", or "Stefan T. Radev" (independent of whether there are more authors or not).

Mmm, this one is weird. I managed to reproduce the issue in a very simple case, but I am still confused about it! I will ask my fellow editors if they have already encountered the issue.
