### 2025-03-20 #7754 @janfb
Thanks @boisgera !

I made some final fixes in the bib files to minimize the issues in the citations and pushed it to our branch.
