### 2025-03-21 sbi-dev/sbi#1378 @vagechirkov
Now that we have _working_ [regression tests](https://pytest-regressions.readthedocs.io/en/latest/api.html#pytest_regressions.ndarrays_regression.NDArraysRegressionFixture.check), we can significantly speed up testing without compromising coverage (see #1461). Thanks @schroedk! This could be an easy improvement for tests that are slow but not marked as slow because they are critical. I think it would be great to collect the list of the next candidates for similar improvement. Any suggestions for such tests?
