### 2025-03-25 sbi-dev/sbi#1389 @janfb
fixed with the move the ReadTheDocs https://github.com/sbi-dev/sbi/pull/1463
