### 2025-03-26 #7754 @boisgera
> Hi [@boisgera](https://github.com/boisgera)
>
> we reviewed the article proof and are all happy with it. 👍

Thank you. Looks good to me too, nothing to mention about it! 👍

Could you please perform a release of the updated software (if this is necessary) and archive this release to get a DOI? (see [additional author instructions](https://github.com/openjournals/joss-reviews/issues/7754#issuecomment-2730792331))
