### 2025-03-26 #7754 @janfb
The most recent release (0.24.0) is now archived with updated author information here: [10.5281/zenodo.15034786](https://doi.org/10.5281/zenodo.15034786)
