### 2025-04-07 #7754 @crvernon
Thanks @janfb - there is still one missing correction from the list I provided above.  In the reference:  "Papamakarios, G., & Murray, I. (2016). Fast ..."  the "b" in "bayesian" needs to be capitalized.

Thanks.
