### 2025-04-07 #7754 @janfb
Sorry for missing that one, should be fixed now. Thanks @crvernon
