### 2025-04-08 #7754 @crvernon
@janfb - I just noticed that you have special formatting around the package name in the title of your paper.  Could you remove any formatting in the title?  This will ensure that the paper renders correctly when published.

Thanks.
